+++ foundations/conda.md
@@ -17,7 +17,7 @@ Here we will cover:
 
 | Concepts                                                                                                  | Importance | Notes |
 | --------------------------------------------------------------------------------------------------------- | ---------- | ----- |
-| [Installing and Running Python](https://foundations.projectpythia.org/foundations/how-to-run-python.html) | Helpful    |       |
+| [Installing and Running Python](https://foundations.projectpythia.org/foundations/how-to-run-python) | Helpful    |       |
 
 - **Time to learn**: 20 minutes
 
