+++ _preview/550/build/_shared/dist-XJZH3WSO.js
@@ -1,0 +1,1 @@
+import{Aa as o,Ba as Z,Ka as T,Ma as X,Oa as _,Pa as c,ca as t,ea as Q,fa as y,ka as W,na as Y,ua as z,xa as a}from"/_preview/550/build/_shared/chunk-ESBRTUPV.js";import{b as s}from"/_preview/550/build/_shared/chunk-2NH4LW52.js";function S(O){return O>=48&&O<=57}function i(O){return S(O)||O==95}var p,v,d,f,x,q,U,h,l,G,e,w,m,u,V,R,k,j,E,I,A,J,D,H,g,b=s(()=>{c();y();p=1,v=2,d=3,f=4,x=5,q=98,U=101,h=102,l=114,G=69,e=48,w=46,m=43,u=45,V=35,R=34,k=124,j=60,E=62;I=new X((O,n)=>{if(S(O.next)){let P=!1;do O.advance();while(i(O.next));if(O.next==w){if(P=!0,O.advance(),S(O.next))do O.advance();while(i(O.next));else if(O.next==w||O.next>127||/\w/.test(String.fromCharCode(O.next)))return}if(O.next==U||O.next==G){if(P=!0,O.advance(),(O.next==m||O.next==u)&&O.advance(),!i(O.next))return;do O.advance();while(i(O.next))}if(O.next==h){let $=O.peek(1);if($==e+3&&O.peek(2)==e+2||$==e+6&&O.peek(2)==e+4)O.advance(3),P=!0;else return}P&&O.acceptToken(x)}else if(O.next==q||O.next==l){if(O.next==q&&O.advance(),O.next!=l)return;O.advance();let P=0;for(;O.next==V;)P++,O.advance();if(O.next!=R)return;O.advance();O:for(;;){if(O.next<0)return;let $=O.next==R;if(O.advance(),$){for(let r=0;r<P;r++){if(O.next!=V)continue O;O.advance()}O.acceptToken(f);return}}}}),A=new X(O=>{O.next==k&&O.acceptToken(p,1)}),J=new X(O=>{O.next==j?O.acceptToken(v,1):O.next==E&&O.acceptToken(d,1)}),D=t({"const macro_rules struct union enum type fn impl trait let static":Q.definitionKeyword,"mod use crate":Q.moduleKeyword,"pub unsafe async mut extern default move":Q.modifier,"fo
[…]
