+++ _preview/550/build/_shared/csp-PX7FT6ZW.js
@@ -1,0 +1,1 @@
+import{c as s}from"/_preview/550/build/_shared/chunk-2NH4LW52.js";var c=s((t,e)=>{function r(n){return{name:"CSP",case_insensitive:!1,keywords:{$pattern:"[a-zA-Z][a-zA-Z0-9_-]*",keyword:"base-uri child-src connect-src default-src font-src form-action frame-ancestors frame-src img-src media-src object-src plugin-types report-uri sandbox script-src style-src"},contains:[{className:"string",begin:"'",end:"'"},{className:"attribute",begin:"^Content",end:":",excludeEnd:!0}]}}e.exports=r});export default c();

[…]
