+++ _preview/550/build/datetime-ebd7991ddfe43b2af731b1b83aa99dad.md
@@ -1,14 +1,0 @@
-# Datetime
-
-```{note}
-This content is under construction!
-```
-
-This section contains tutorials on dealing with times and calendars in scientific Python. The first and most basic of these tutorials covers the standard Python library known as [datetime](https://docs.python.org/3/library/datetime.html).
-
-When this chapter is fully built out, it will include a comprehensive guide to different time libraries, where to use them, and when they might be useful. This set of time libraries includes these libraries, among others:
-
-- [Numpy `datetime64`](https://numpy.org/doc/stable/reference/arrays.datetime.html) (for efficient vectorized date and time operations)
-- [cftime library](https://unidata.github.io/cftime/) (for dealing with dates and times in non-standard calendars)
-
-These tutorials will be cross-referenced with other tutorials on time-related topics, such as dealing with timeseries data in [Pandas](pandas) and [Xarray](xarray).
