### 2025-07-23 89d5624
Fix foundations links by removing '.html' (#550)
* fix internal links

* Update preamble/how-to-use.md

Co-authored-by: Katelyn FitzGerald <7872563+kafitzgerald@users.noreply.github.com>

---------

Co-authored-by: Katelyn FitzGerald <7872563+kafitzgerald@users.noreply.github.com>
