+++ foundations/terminal.md
@@ -13,7 +13,7 @@ You'd like to learn to run Python in the terminal. Here we will cover:
 
 | Concepts                                                                                                  | Importance | Notes |
 | --------------------------------------------------------------------------------------------------------- | ---------- | ----- |
-| [Installing and Running Python](https://foundations.projectpythia.org/foundations/how-to-run-python.html) | Helpful    |       |
+| [Installing and Running Python](https://foundations.projectpythia.org/foundations/how-to-run-python) | Helpful    |       |
 
 - **Time to learn**: 20 minutes
 
