+++ CONTRIBUTING.md
@@ -1,2 +1,2 @@
 Information on how to contribute to this repository can be found
-in the [Pythia Foundations Contributor's Guide](https://foundations.projectpythia.org/appendix/how-to-contribute.html).
+in the [Pythia Foundations Contributor's Guide](https://foundations.projectpythia.org/appendix/how-to-contribute).
