+++ _preview/550/build/entry.client-UNPC4GT3.js
@@ -1,1 +1,0 @@
-import{o as r}from"/_preview/550/build/_shared/chunk-OCTKKCIL.js";import{a,c as l,d as i}from"/_preview/550/build/_shared/chunk-UAI5KRM7.js";import{e as o}from"/_preview/550/build/_shared/chunk-2NH4LW52.js";var t=o(a()),m=o(l()),e=o(i());function d(){(0,t.startTransition)(()=>{(0,m.hydrateRoot)(document,(0,e.jsx)(t.StrictMode,{children:(0,e.jsx)(r,{})}))})}window.requestIdleCallback?window.requestIdleCallback(d):window.setTimeout(d,1);
