+++ build/github-cloning-forki-3a2e652882d19baa5c4b0bd3ada500f4.md
@@ -1,236 +1,0 @@
-# Cloning and Forking a Repository
-
-```{image} ../../images/GitHub-logo.png
-:alt: GitHub Logo
-:width: 400px
-```
-
-## Overview:
-
-1. Cloning and forking a git repository
-1. Cloning a repository
-1. Forking a repository
-
-## Prerequisites
-
-| Concepts                                | Importance  | Notes                        |
-| --------------------------------------- | ----------- | ---------------------------- |
-| [What is GitHub?](what-is-github)       | Necessary   | GitHub user account required |
-| [GitHub Repositories](github-repos)     | Necessary   |                              |
-| [Issues and Discussions](github-issues) | Recommended |                              |
-| Command-line shell                      | Helpful     |                              |
-
-- **Time to learn**: 30 minutes
-
----
-
-## Cloning and forking
-
-_Cloning_ and _forking_ are two related terms in the GitHub vernacular
-that, unfortunately, are not always used consistently throughout
-the _web-o-sphere_. In Project Pythia we use the term _clone_ to refer to
-making a **local** copy of a **remote** repository; the source for
-the copy is a remote repo, and the destination for the copy is your
-local laptop/desktop. When working with GitHub, a _fork_, on the
-other hand, creates a copy of a GitHub repository on GitHub. In other
-words, both the source and the destination of the _fork_ operations are
-hosted in the cloud on GitHub. Forking is performed via your GitHub
-account. While the forked repository may be owned by anyone, the
-newly created repository will be owned by you. Cloning, on the
-other hand, is performed using a Git command. Naturally, since the
-destination of the clone operation is your local computer, you will
-own the cloned contents. In either case, whether you clone or fork,
-any changes you make to the newly created repository will not impact
-the original without taking explicit action (e.g. performing a
-_push_ or submitting a _Pull Request_, the topics of later sections
-in this guide).
-
-Cloning and forking are often used together (more on this later).
-The illustration below demonstrates the operation of a Fork of a
-remote repository (UPSTREAM), followed by a clone of the newly
-created ORIGIN.
-
-![clone-and-fork](../../images/github-clone-fork.png)
-
-## Cloning a repository
-
-Cloning is ideal for the following scenarios:
-
-1. You wish to download, build, and install the latest version of a software package.
-1. You would like to experiment with a repository on your local computer, but do not desire to maintain a separate copy of it (termed a _fork_, to be covered later in this lesson) on your GitHub account.
-1. You have previously _forked_ a repository to your own GitHub account, and now wish to make changes to it for possible incorporation into the original repo, via a _Pull Request_.
-
-Let's consider the 2nd scenario. Say you wish to copy a GitHub repository to a computer you have access to (which could be your own computer, or one you have access to at work or school).
-
-We'll use a very basic repo that is part of the [Project Pythia organization](https://github.com/ProjectPythia) as our example.
-
-First, point your browser to <https://github.com/ProjectPythia/github-sandbox>:
-
-<img src="../../images/GitHub_SandboxRepo.png" alt="SandboxRepo">
-
----
-
-We see that in the repository, there exists five files. Above the list of files is this row:
-
-<img src="../../images/GitHub_RepoTools.png" alt="RepoTools">
-
----
-
-Click on the green **Code** button to the right:
-
-<img src="../../images/GitHub_CodeClone.png" alt="CodeClone">
-
----
-
-Select the **HTTPS** option, and click on the copy-to-clipboard icon:
-
-<img src="../../images/GitHub_CodeCloneHTTPS.png" alt="CodeCloneHTTPS">
-
----
-
-```{tip}
-This link points to where the repository "lives" on GitHub. We will use the term **origin** to refer to this location.
-```
-
[…]
