+++ _preview/551/build/_shared/clojure-repl-ADYZKNJR.js
@@ -1,0 +1,1 @@
+import{c as n}from"/_preview/551/build/_shared/chunk-2NH4LW52.js";var a=n((t,e)=>{function s(r){return{name:"Clojure REPL",contains:[{className:"meta",begin:/^([\w.-]+|\s*#_)?=>/,starts:{end:/$/,subLanguage:"clojure"}}]}}e.exports=s});export default a();
