+++ _preview/551/build/_shared/dist-NN6GTV4I.js
@@ -1,0 +1,6 @@
+import{a as Fe,e as $e,g as $t}from"/_preview/551/build/_shared/chunk-QA24G7CQ.js";import"/_preview/551/build/_shared/chunk-RBSODAQ6.js";import"/_preview/551/build/_shared/chunk-OWXZLTDO.js";import{a as Xe,g as Ft}from"/_preview/551/build/_shared/chunk-X4YK4Z35.js";import{$ as He,Aa as W,F as Be,Ka as Xt,T as Tt,U as I,V as D,W as V,Y as E,_ as Me,aa as ve,ba as Pe,ca as K,ea as c,fa as Dt,ga as Ne,ha as ye,i as M,ja as Oe,l as Ie,la as X,ma as Re,na as J,oa as Y,pa as ze,r as Ee,ua as Te,w as y,x as zt,za as De}from"/_preview/551/build/_shared/chunk-ESBRTUPV.js";import{b as xe}from"/_preview/551/build/_shared/chunk-2NH4LW52.js";function je(t,e,r){if(r.pos==r.text.length||t!=e.block&&r.indent>=e.stack[r.depth+1].value+r.baseIndent)return!0;if(r.indent>=r.baseIndent+4)return!1;let n=(t.type==u.OrderedList?ke:ge)(r,e,!1);return n>0&&(t.type!=u.BulletList||me(r,e,!1)<0)&&r.text.charCodeAt(r.pos+n-1)==t.value}function A(t){return t==32||t==9||t==10||t==13}function $(t,e=0){for(;e<t.length&&A(t.charCodeAt(e));)e++;return e}function _e(t,e,r){for(;e>r&&A(t.charCodeAt(e-1));)e--;return e}function et(t){if(t.next!=96&&t.next!=126)return-1;let e=t.pos+1;for(;e<t.text.length&&t.text.charCodeAt(e)==t.next;)e++;if(e<t.pos+3)return-1;if(t.next==96){for(let r=e;r<t.text.length;r++)if(t.text.charCodeAt(r)==96)return-1}return e}function tt(t){return t.next!=62?-1:t.text.charCodeAt(t.pos+1)==32?2:1}function me(t,e,r){if(t.next!=42&&t.next!=45&&t.next!=95)return-1;let n=1;for(let s=t.pos+1;s<t.text.length;s++){let i=t.text.charCodeAt(s);if(i==t.next)n++;else if(!A(i))return-1}return r&&t.next==45&&st(t)>-1&&t.depth==e.stack.length&&e.parser.leafBlockParsers.indexOf(lt.SetextHeading)>-1||n<3?-1:1}function rt(t,e){for(let r=t.stack.length-1;r>=0;r--)if(t.stack[r].type==e)return!0;return!1}function ge(t,e,r){return(t.next==45||t.next==43||t.next==42)&&(t.pos==t.text.length-1||A(t.text.charCodeAt(t.pos+1)))&&(!r||rt(e,u.BulletList)||t.skipSpace(t.pos+2)<t.text.length)?1:-1}function ke(t,e,r){let n=t.pos,s=t.next;for(;s>=48&&s<=57;){n++;if(n==t.text.length)return-1;s=t.text.charCodeAt(n)}return n==t.pos||n>t.pos+9||s!=46&&s!=41||n<t.text.length-1&&!A(t.text.charCodeAt(n+1))||r&&!rt(e,u.OrderedList)&&(t.skipSpace(n+1)==t.text.length||n>t.pos+1||t.next!=49)?-1:n+1-t.pos}function nt(t){if(t.next!=35)return-1;let e=t.pos+1;for(;e<t.text.
[…]
