+++ landing-page.md
@@ -3,6 +3,7 @@
 ## A community learning resource for Python-based computing in the geosciences
 
 [![DOI](https://zenodo.org/badge/338145160.svg)](https://zenodo.org/badge/latestdoi/338145160)
+[![status](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15/status.svg)](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15)
 
 :::{image} images/pretty-earth.png
 :alt: Pretty Earth
