+++ README.md
@@ -2,6 +2,7 @@
 
 [![nightly-build](https://github.com/ProjectPythia/pythia-foundations/actions/workflows/nightly-build.yaml/badge.svg)](https://github.com/ProjectPythia/pythia-foundations/actions/workflows/nightly-build.yaml)
 [![DOI](https://zenodo.org/badge/338145160.svg)](https://zenodo.org/badge/latestdoi/338145160)
+[![status](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15/status.svg)](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15)
 
 This is the source repository for the Pythia Foundations content collection.
 
