+++ _preview/551/build/_shared/shell-JJYZC475.js
@@ -1,1 +1,0 @@
-import{c as e}from"/_preview/551/build/_shared/chunk-2NH4LW52.js";var n=e((t,s)=>{function a(l){return{name:"Shell Session",aliases:["console"],contains:[{className:"meta",begin:/^\s{0,3}[/~\w\d[\]()@-]*[>%$#]/,starts:{end:/[^\\](?=\s*$)/,subLanguage:"bash"}}]}}s.exports=a});export default n();
