+++ _preview/554/build/_shared/clean-BB4ZAQ7O.js
@@ -1,0 +1,1 @@
+import{c as n}from"/_preview/554/build/_shared/chunk-2NH4LW52.js";var l=n((t,i)=>{function a(e){return{name:"Clean",aliases:["icl","dcl"],keywords:{keyword:"if let in with where case of class instance otherwise implementation definition system module from import qualified as special code inline foreign export ccall stdcall generic derive infix infixl infixr",built_in:"Int Real Char Bool",literal:"True False"},contains:[e.C_LINE_COMMENT_MODE,e.C_BLOCK_COMMENT_MODE,e.APOS_STRING_MODE,e.QUOTE_STRING_MODE,e.C_NUMBER_MODE,{begin:"->|<-[|:]?|#!?|>>=|\\{\\||\\|\\}|:==|=:|<>"}]}}i.exports=a});export default l();
