+++ _preview/554/build/_shared/gdscript-X7CHA2ZK.js
@@ -1,0 +1,1 @@
+import{a}from"/_preview/554/build/_shared/chunk-ZUDNVCVE.js";import"/_preview/554/build/_shared/chunk-2NH4LW52.js";export default a();
