+++ _preview/554/build/_shared/dist-6TTSG3GK.js
@@ -1,41 +1,0 @@
-import{B as b,b as y,c as B,j as E,m as T,n as S,s as N,t as $}from"/_preview/554/build/_shared/chunk-ND43KHSX.js";import{a as F}from"/_preview/554/build/_shared/chunk-GUCIBHGO.js";import"/_preview/554/build/_shared/chunk-2NH4LW52.js";var P=.7,q=800;function Z(r,e=" "){return r.replace(/\u00A0/g,e).replace(/[\u200B-\u200D\uFEFF]/g,"")}var U="\u{1F4A5}\u{1F3AF}BACKSLASHSPACE\u{1F3AF}\u{1F4A5}",X="\u{1F4A5}\u{1F3AF}BACKSLASH\u{1F3AF}\u{1F4A5}",K="\u{1F4A5}\u{1F3AF}TILDE\u{1F3AF}\u{1F4A5}",O={"&":"\\&","%":"\\%",$:"\\$","#":"\\#",_:"\\_","{":"\\{","}":"\\}","^":"\\^"},de={...O,"\u2019":"'","\u2018":"`","\u201D":"''","\u201C":"``","\xBB":">>","\xAB":"<<","\u2026":"\\dots","\u2013":"--","\u2014":"---","\xA9":"\\textcopyright ","\xAE":"\\textregistered ","\u2122":"\\texttrademark ","<":"\\textless ",">":"\\textgreater ","\xA0":"~","\u202F":"~","\u2009":"\\,"},J={"\u2194":"\\leftrightarrow","\u21D4":"\\Leftrightarrow","\u2192":"\\rightarrow","\u21D2":"\\Rightarrow","\u2190":"\\leftarrow","\u21D0":"\\Leftarrow"},Q={"\u2212":"-","-":"-","\uFE63":"-","\uFF0D":"-","\uFF0B":"+"},fe={"\u2080":"\\textsubscript{0}","\u2081":"\\textsubscript{1}","\u2082":"\\textsubscript{2}","\u2083":"\\textsubscript{3}","\u2084":"\\textsubscript{4}","\u2085":"\\textsubscript{5}","\u2086":"\\textsubscript{6}","\u2087":"\\textsubscript{7}","\u2088":"\\textsubscript{8}","\u2089":"\\textsubscript{9}","\u208A":"\\textsubscript{+}","\u208B":"\\textsubscript{-}","\u208C":"\\textsubscript{=}","\u208D":"\\textsubscript{(}","\u208E":"\\textsubscript{)}","\u2099":"\\textsubscript{n}","\u2070":"\\textsuperscript{0}","\xB9":"\\textsuperscript{1}","\xB2":"\\textsuperscript{2}","\xB3":"\\textsuperscript{3}","\u2074":"\\textsuperscript{4}","\u2075":"\\textsuperscript{5}","\u2076":"\\textsuperscript{6}","\u2077":"\\textsuperscript{7}","\u2078":"\\textsuperscript{8}","\u2079":"\\textsuperscript{9}","\u22C5":"\\textsuperscript{.}","\u207A":"\\textsuperscript{.}","\u207B":"\\textsuperscript{-}","\u207C":"\\textsuperscript{=}","\u207D":"\\textsuperscript{(}","\u207E":"\\textsuperscript{)}",\u207F:"\\textsuperscript{n}","\u2071":"\\textsuperscript{i}"},V={...de,...J,...Q,...fe},A={...J,...Q,"\xBD":"\\frac{1}{2}","\u2153":"\\frac{1}{3}","\u2154":"\\frac{2}{3}","\xBC":"\\frac{1}{4}","\u2155":"\\frac{1}{
[…]
