### 2025-08-08 #295 @labarba
Thank you @acocac !!
Feel free to run `@editorialbot assign me as editor` next time. It just works!
