+++ _preview/555/build/_shared/chunk-PDXVSPKU.js
@@ -1,1 +1,0 @@
-import{c as a}from"/_preview/555/build/_shared/chunk-2NH4LW52.js";var n=a((o,e)=>{e.exports=t;t.displayName="bbcode";t.aliases=["shortcode"];function t(s){s.languages.bbcode={tag:{pattern:/\[\/?[^\s=\]]+(?:\s*=\s*(?:"[^"]*"|'[^']*'|[^\s'"\]=]+))?(?:\s+[^\s=\]]+\s*=\s*(?:"[^"]*"|'[^']*'|[^\s'"\]=]+))*\s*\]/,inside:{tag:{pattern:/^\[\/?[^\s=\]]+/,inside:{punctuation:/^\[\/?/}},"attr-value":{pattern:/=\s*(?:"[^"]*"|'[^']*'|[^\s'"\]=]+)/,inside:{punctuation:[/^=/,{pattern:/^(\s*)["']|["']$/,lookbehind:!0}]}},punctuation:/\]/,"attr-name":/[^\s=\]]+/}}},s.languages.shortcode=s.languages.bbcode}});export{n as a};
