### 2025-08-12 #295 @acocac
Hi @brian-rose, thanks for submitting to JOSE. I confirm the article proof and statement of need section read well. There aren't invalid or missing DOIs.

I do not have any particular suggestions at this stage, so I suggest moving forward to find potential reviewers. Feel free to mention candidates according to the list of JOSE reviewers. You could find them by running `@editorialbot list reviewers`.
