+++ build/overview-8c69f7ff27a77f59e6ba27e9d1734bc6.md
@@ -1,124 +1,0 @@
-# Overview
-
-![Foundations Infographic](../images/ProjectPythia_Foundations_Infographic_v4.jpg)
-
-You've made it to the **Core Packages** section of the book!
-
-As you might know by now that Python is a programming language. To make your job easier, developers of this programming language provide users like you with libraries (or packages). Core libraries will help you with fundamental numerical functions, and high-level libraries will help you efficiently analyze and visualize your data. Some of these libraries are used all across the Python community, while others are domain-specific. Read below to learn more about core and high-level libraries, and domain-specific libraries of the geoscience community. We suggest that new users start with the [Foundational Skills](../foundations/overview) section in order to get the most out of the tutorials below.
-
-## Core libraries
-
-Most geoscience data analysis involves working with numerical arrays.
-The default library for dealing with numerical arrays in Python is [NumPy](http://www.numpy.org/).
-It has some built in functions for calculating very simple statistics
-(e.g. maximum, mean, standard deviation),
-but for more complex analysis
-(e.g. interpolation, integration, linear algebra)
-the [SciPy](https://scipy.org) library is the default.
-If you’re dealing with particularly large arrays,
-[Dask](https://dask.org/) works with the existing Python ecosystem
-(including NumPy) to scale your analysis
-to multi-core machines and/or distributed clusters (i.e. parallel processing).
-
-Another common feature of geo-data science is time series analysis.
-The Python standard library comes with a [datetime](https://docs.python.org/3/library/datetime.html)
-package for manipulating dates and times.
-NumPy also includes a [datetime64](https://numpy.org/doc/stable/reference/arrays.datetime.html)
-module for efficient vectorized datetime operations
-and the [cftime](https://unidata.github.io/cftime/) library
-is useful for dealing with non-standard calendars.
-
-When it comes to data visualization,
-the default library is [Matplotlib](https://matplotlib.org/).
-As you can see at the [Matplotlib gallery](https://matplotlib.org/stable/gallery/index.html),
-this library is great for any simple (e.g. bar charts, contour plots, line graphs),
-static (e.g. .png, .eps, .pdf) plots.
-The [Cartopy](https://scitools.org.uk/cartopy/docs/latest/) library
-provides additional plotting functionality for common geographic map projections.
-
-## High-level libraries
-
-While pretty much all data analysis and visualization tasks
-could be achieved with a combination of the core libraries,
-their flexible, all-purpose nature means relatively common/simple tasks
-can often require quite a bit of work (i.e. many lines of code).
-To make things more efficient for data scientists,
-the scientific Python community has therefore built a number of libraries on top of the core stack.
-These high-levels libraries aren’t as flexible
-– they can’t do _everything_ like the core stack can –
-but they can do common tasks with far less effort.
-
-The most popular high-level data science library is undoubtedly [Pandas](http://pandas.pydata.org/).
-The key advance offered by Pandas is the concept of labeled arrays.
-Rather than referring to the individual elements of a data array using a numeric index
-(as is required with NumPy),
-the actual row and column headings can be used.
-That means information from the cardiac ward on 3 July 2005
-could be obtained from a medical dataset by asking for `data['cardiac'].loc['2005-07-03']`,
-rather than having to remember the numeric index corresponding to that ward and date.
-This labeled array feature,
-combined with a bunch of other features that streamline common statistical and plotting tasks
-traditionally performed with SciPy, datetime and Matplotlib,
-greatly simplifies the code development process (read: less lines of code).
-
-One of the limitations of Pandas
[…]
