### 2025-08-28 #295 @brian-rose
Hi @acocac, here are a few suggestions of folks who would be excellent reviewers for this work, if they are willing. I'll avoid @ tagging people for now:
- [Tom Vo](https://github.com/tomvothecoder), the lead author of the xCDAT climate analysis package and [accompanying JOSS paper](https://doi.org/10.21105/joss.06426)
- [Emma Marshall](https://github.com/e-marshall), PhD student at the University of Utah and lead author of [Cloud-native geospatial data cube workflows with open-source tools](https://doi.org/10.21105/jose.00267) published in JOSE
- [Julia Stewart Lowndes](https://jules32.github.io), founding directory of [Openscapes](https://openscapes.org), or an earlier career Openscapes colleague that Julia might recommend
- [Leah Wasser](https://github.com/lwasser), founder and executive director of [PyOpenSci](https://www.pyopensci.org), or an earlier career PyOpenSci colleague that Leah might recommend
