### 2025-09-02 #295 @brian-rose
@acocac here's another suggestion which I forgot to include last week:
- [Jenna Pearson](https://jlpearso.github.io), Course Director and co-founder of the [Climatematch Academy](https://neuromatch.io/climate-science/). Climatematch has used or repurposed material from Pythia Foundations for their course prep materials.
