+++ build/github-pull-request-7b2f066897050b00fa547c2d9d38410f.md
@@ -1,169 +1,0 @@
-# Opening a Pull Request on GitHub
-
-```{image} ../../images/GitHub-logo.png
-:alt: GitHub Logo
-:width: 400px
-```
-
-A Pull Request, aka a "merge request," is an event that occurs when a project contributor begins the process of merging new code changes from a feature branch with the main project repository.
-
-## Overview:
-
-1. What is a Pull Request?
-1. Opening a Pull Request
-1. Pull Request Features
-1. GitHub Workflows
-
-## Prerequisites
-
-| Concepts                                      | Importance  | Notes |
-| --------------------------------------------- | ----------- | ----- |
-| [What is GitHub](what-is-github)              | Necessary   |       |
-| [GitHub Repositories](github-repos)           | Necessary   |       |
-| [Cloning and Forking](github-cloning-forking) | Necessary   |       |
-| [Basic Version Control with _git_](basic-git) | Necessary   |       |
-| [Issues and Discussions](github-issues)       | Recommended |       |
-| [Branches](git-branches)                      | Necessary   |       |
-
-- **Time to learn**: 60 minutes
-
----
-
-## What is a Pull Request?
-
-A Pull Request (PR) is a formal mechanism for requesting that changes
-that you have made to one repository are integrated (merged) into
-another repository. Typically, the changes are reviewed by the
-maintainers of the destination repository, potentially triggering
-a cycle of revisions, before the PR is “merged”, and your changes
-become part of the destination repo.
-
-Just like Issues, PRs have
-their own discussion forum for communicating about the proposed
-changes. In fact, not only can maintainers or collaborators communicate
-about your PR via GitHub, they can also suggest changes and may
-even be able to make changes of their own by pushing follow-up
-commits. All of the activity, from start to finish, is tracked
-inside of the PR and can be reviewed at any time.
-
-When a contributor to a project creates a PR they are requesting
-that the owners of another destination repository pull a git
-branch from the contributor’s repository and merge the contents of
-the branch into a branch of the destination repository. This means
-that the contributor must provide four pieces of information: the
-contributor’s repository, the contributor’s branch, the destination
-repository, and finally, the destination branch.
-
-A typical sequence of steps consists of the following:
-
-1. A contributor clones a personal remote repository, creating a local copy
-1. The contributor creates a new branch in their local repository
-1. The contributor makes changes to the branch and commits them to
-   their local repository
-1. The contributor _pushes_ the branch to a remote repository
-1. The contributor submits a PR via GitHub
-
-After the maintainers or collaborators of the destination review
-the changes, and any suggested revisions are made, the project
-maintainer merges the feature into the destination repository and
-closes the PR.
-
-## Opening a Pull Request
-
-The demonstration is a continuation from the [GitHub Branches chapter](github-branches). Here, we will move from your local terminal to GitHub.
-
-### Navigate to Your Fork
-
-Go to your fork of the [GitHub Sandbox Repository](https://github.com/ProjectPythia/github-sandbox). One fast way to get to your fork, is to click the "fork" button and then follow the link underneath the message, "You've already forked github-sandbox."
-
-When you've navigated to your fork, you should see a message box alerting you that your branch `branchA` had recent changes with the option to generate an open Pull Request. This Pull Request would take the changes from your `branchA` branch and suggest them for the original upstream ProjectPythia github-sandbox repository. You'll also notice that you are on branch `main`, but that there are now 2 branches.
-
-![GitHub](../../images/8-github.png)
-
-### Switch Branches
-
-If you click on the branch `main` you'll see the list of these branches.
-
-![GitHub Branches](../../images/9-github-seebranches.png)
-
-There you can click on the branch `branchA` to switch branches.
-
[…]
