### 2025-09-22 #308 @editorialbot
[REVIEW]: Pythia Foundations: A community learning resource for Python-based computing in the geosciences
**Submitting author:** @brian-rose (<a href="http://orcid.org/0000-0002-9961-3821">Brian E. J. Rose</a>)
**Repository:** https://github.com/ProjectPythia/pythia-foundations
**Branch with paper.md** (empty if default branch):
**Version:** v2025.06.25
