+++ _preview/556/build/_shared/textile-JCTYTRNK.js
@@ -1,0 +1,1 @@
+import{b as h}from"/_preview/556/build/_shared/chunk-2NH4LW52.js";function g(i,e){e.mode=r.newLayout,e.tableHeading=!1,e.layoutType==="definitionList"&&e.spanningLayout&&i.match(l("definitionListEnd"),!1)&&(e.spanningLayout=!1)}function s(i,e,n){if(n==="_")return i.eat("_")?a(i,e,"italic",/__/,2):a(i,e,"em",/_/,1);if(n==="*")return i.eat("*")?a(i,e,"bold",/\*\*/,2):a(i,e,"strong",/\*/,1);if(n==="[")return i.match(/\d+\]/)&&(e.footCite=!0),u(e);if(n==="("){var o=i.match(/^(r|tm|c)\)/);if(o)return f.specialChar}if(n==="<"&&i.match(/(\w+)[^>]+>[^<]+<\/\1>/))return f.html;if(n==="?"&&i.eat("?"))return a(i,e,"cite",/\?\?/,2);if(n==="="&&i.eat("="))return a(i,e,"notextile",/==/,2);if(n==="-"&&!i.eat("-"))return a(i,e,"deletion",/-/,1);if(n==="+")return a(i,e,"addition",/\+/,1);if(n==="~")return a(i,e,"sub",/~/,1);if(n==="^")return a(i,e,"sup",/\^/,1);if(n==="%")return a(i,e,"span",/%/,1);if(n==="@")return a(i,e,"code",/@/,1);if(n==="!"){var c=a(i,e,"image",/(?:\([^\)]+\))?!/,1);return i.match(/^:\S+/),c}return u(e)}function a(i,e,n,o,c){var d=i.pos>c?i.string.charAt(i.pos-c-1):null,p=i.peek();if(e[n]){if((!p||/\W/.test(p))&&d&&/\S/.test(d)){var y=u(e);return e[n]=!1,y}}else(!d||/\W/.test(d))&&p&&/\S/.test(p)&&i.match(new RegExp("^.*\\S"+o.source+"(?:\\W|$)"),!1)&&(e[n]=!0,e.mode=r.attributes);return u(e)}function u(i){var e=b(i);if(e)return e;var n=[];return i.layoutType&&n.push(f[i.layoutType]),n=n.concat(m(i,"addition","bold","cite","code","deletion","em","footCite","image","italic","link","span","strong","sub","sup","table","tableHeading")),i.layoutType==="header"&&n.push(f.header+"-"+i.header),n.length?n.join(" "):null}function b(i){var e=i.layoutType;switch(e){case"notextile":case"code":case"pre":return f[e];default:return i.notextile?f.notextile+(e?" "+f[e]:""):null}}function m(i){for(var e=[],n=1;n<arguments.length;++n)i[arguments[n]]&&e.push(f[arguments[n]]);return e}function k(i){var e=i.spanningLayout,n=i.layoutType;for(var o in i)i.hasOwnProperty(o)&&delete i[o];i.mode=r.newLayout,e&&(i.layoutType=n,i.spanningLayout=!0)}function l(i){return t.cache[i]||(t.cache[i]=t.createRe(i))}var f,t,r,v,T=h(()=>{f={addition:"inserted",attributes:"propertyName",bold:"strong",cite:"keyword",code:"monospace",definitionList:"list",deletion:"deleted",div:"punctuation",em:"emphasis",footnote:"variable",footCite:"qualifier",header:"heading",html:"comment",image:"atom",italic:"emphasis",link:"link",linkDefinition:"link",list1:"list",list2:"list.special",list3:"list",notextile:"string.special",pre:"operator",p:"content",quote:"bracket",span:"quote",specialChar:"character",strong:"strong",sub:"content.special",sup
[…]
