+++ _preview/556/build/_shared/dockerfile-7J5VVXYT.js
@@ -1,1 +1,0 @@
-import{c as r}from"/_preview/556/build/_shared/chunk-2NH4LW52.js";var i=r((o,n)=>{function a(e){return{name:"Dockerfile",aliases:["docker"],case_insensitive:!0,keywords:"from maintainer expose env arg user onbuild stopsignal",contains:[e.HASH_COMMENT_MODE,e.APOS_STRING_MODE,e.QUOTE_STRING_MODE,e.NUMBER_MODE,{beginKeywords:"run cmd entrypoint volume add copy workdir label healthcheck shell",starts:{end:/[^\\]$/,subLanguage:"bash"}}],illegal:"</"}}n.exports=a});export default i();
