+++ _preview/557/build/_shared/chunk-4PHDSBSU.js
@@ -1,0 +1,1 @@
+import{c as n}from"/_preview/557/build/_shared/chunk-2NH4LW52.js";var r=n((o,a)=>{a.exports=e;e.displayName="yang";e.aliases=[];function e(t){t.languages.yang={comment:/\/\*[\s\S]*?\*\/|\/\/.*/,string:{pattern:/"(?:[^\\"]|\\.)*"|'[^']*'/,greedy:!0},keyword:{pattern:/(^|[{};\r\n][ \t]*)[a-z_][\w.-]*/i,lookbehind:!0},namespace:{pattern:/(\s)[a-z_][\w.-]*(?=:)/i,lookbehind:!0},boolean:/\b(?:false|true)\b/,operator:/\+/,punctuation:/[{};:]/}}});export{r as a};
