+++ _preview/557/build/_shared/powershell-XFWVAV7I.js
@@ -1,0 +1,1 @@
+import{b as P}from"/_preview/557/build/_shared/chunk-2NH4LW52.js";function a(e,n){n=n||{};for(var r=n.prefix!==void 0?n.prefix:"^",t=n.suffix!==void 0?n.suffix:"\\b",i=0;i<e.length;i++)e[i]instanceof RegExp?e[i]=e[i].source:e[i]=e[i].replace(/[-\/\\^$*+?.()|[\]{}]/g,"\\$&");return new RegExp(r+"("+e.join("|")+")"+t,"i")}function o(e,n){var r=n.returnStack[n.returnStack.length-1];if(r&&r.shouldReturnFrom(n))return n.tokenize=r.tokenize,n.returnStack.pop(),n.tokenize(e,n);if(e.eatSpace())return null;if(e.eat("("))return n.bracketNesting+=1,"punctuation";if(e.eat(")"))return n.bracketNesting-=1,"punctuation";for(var t in p)if(e.match(p[t]))return t;var i=e.next();if(i==="'")return I(e,n);if(i==="$")return s(e,n);if(i==='"')return v(e,n);if(i==="<"&&e.eat("#"))return n.tokenize=S,S(e,n);if(i==="#")return e.skipToEnd(),"comment";if(i==="@"){var l=e.eat(/["']/);if(l&&e.eol())return n.tokenize=u,n.startQuote=l[0],u(e,n);if(e.eol())return"error";if(e.peek().match(/[({]/))return"punctuation";if(e.peek().match(c))return s(e,n)}return"error"}function I(e,n){for(var r;(r=e.peek())!=null;)if(e.next(),r==="'"&&!e.eat("'"))return n.tokenize=o,"string";return"error"}function v(e,n){for(var r;(r=e.peek())!=null;){if(r==="$")return n.tokenize=z,"string";if(e.next(),r==="`"){e.next();continue}if(r==='"'&&!e.eat('"'))return n.tokenize=o,"string"}return"error"}function z(e,n){return g(e,n,v)}function T(e,n){return n.tokenize=u,n.startQuote='"',u(e,n)}function D(e,n){return g(e,n,T)}function g(e,n,r){if(e.match("$(")){var t=n.bracketNesting;return n.returnStack.push({shouldReturnFrom:function(i){return i.bracketNesting===t},tokenize:r}),n.tokenize=o,n.bracketNesting+=1,"punctuation"}else return e.next(),n.returnStack.push({shouldReturnFrom:function(){return!0},tokenize:r}),n.tokenize=s,n.tokenize(e,n)}function S(e,n){for(var r=!1,t;(t=e.next())!=null;){if(r&&t==">"){n.tokenize=o;break}r=t==="#"}return"comment"}function s(e,n){var r=e.peek();return e.eat("{")?(n.tokenize=m,m(e,n)):r!=null&&r.match(c)?(e.eatWhile(c),n.tokenize=o,"variable"):(n.tokenize=o,"error")}function m(e,n){for(var r;(r=e.next())!=null;)if(r==="}"){n.tokenize=o;break}return"variable"}function u(e,n){var r=n.startQuote;if(e.sol()&&e.match(new RegExp(r+"@")))n.tokenize=o;else if(r==='"')for(;!e.eol();){var t=e.peek();if(t==="$")return n.tokenize=D,"string";e.next(),t==="`"&&e.next()}else e.skipToEnd();return"string"}var f,c,d,b,C,k,h,x,E,w,y,M,R,p,O,j=P(()=>{
[…]
