### 2025-09-29 86c3742
[pre-commit.ci] pre-commit autoupdate (#557)
updates:
- [github.com/keewis/blackdoc: v0.4.1 → v0.4.2](https://github.com/keewis/blackdoc/compare/v0.4.1...v0.4.2)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
