+++ build/_shared/chunk-NFM2H3KQ.js
@@ -1,0 +1,7 @@
+import{$ as c2,A as D9,B as K,Ba as O2,C as T9,D as g1,E as W1,Ea as I2,F as U9,Fa as B2,G as G1,Ga as j2,H as z9,Ha as P2,I as W9,Ia as k2,J as G9,Ja as N2,K as V1,Ka as E2,L as X9,La as S2,M as K9,N as Y9,O as q9,P as J9,Q as Q9,R as e2,S as o2,T as r2,U as l2,V as b1,W as t2,X as n2,Y as i2,Z as C2,_ as a1,a as x0,aa as d2,b as b9,ba as a2,c as $9,da as s2,e as R1,ea as v2,fa as L1,g as y9,ga as u2,h as _9,ha as h2,i as O9,ia as w2,j as w1,ja as g2,ka as L2,l as U1,la as x2,m as I9,ma as m2,n as B9,na as M2,oa as H2,p as j9,q as P9,r as k9,ra as Z2,s as N9,sa as R2,t as E9,ta as V2,u as S9,ua as b2,v as X,va as $2,w as A9,x as z1,y as F9,ya as y2,z as Q,za as _2}from"/build/_shared/chunk-XWET6RJ7.js";import{m as r1,n as f2,p as p2,r as X1}from"/build/_shared/chunk-HBJK6BW3.js";import{a as c1}from"/build/_shared/chunk-OCWQY3HK.js";import{a as d1}from"/build/_shared/chunk-3CVK3PYF.js";import{k as V9}from"/build/_shared/chunk-ND43KHSX.js";import{a as H9,e as Z9,l as R9}from"/build/_shared/chunk-OCTKKCIL.js";import{a as G,b as L0,d as S}from"/build/_shared/chunk-UAI5KRM7.js";import{e as x}from"/build/_shared/chunk-2NH4LW52.js";var A2={rel:"stylesheet",href:"https://cdn.jsdelivr.net/npm/katex@0.15.2/dist/katex.min.css",integrity:"sha384-MlJdn/WNKDGXveldHDdyRP1R4CTHr3FeuDNfhsLPYrq2t0UBkUdK2jyTnXPEK1NQ",crossOrigin:"anonymous"};var Z=x(S(),1),f1=x(d1(),1);var D=x(S(),1);var $1=x(d1(),1);function m0({license:e,preamble:o="",className:r}){var l;if(!e.id)return null;let t=/^([CBYSAND0-]+)(?:(?:-)([0-9].[0-9]))?$/.exec(e.id);if(!e.CC||!t)return null;let n=`${o}${(l=e.name)!==null&&l!==void 0?l:e.title} (${e.id})`,i=t[1].toUpperCase();return(0,D.jsxs)("a",{href:e.url,target:"_blank",rel:"noopener noreferrer",className:(0,$1.default)("opacity-50 hover:opacity-100 text-inherit hover:text-inherit",r),"aria-label":n,children:[(0,D.jsx)(o2,{width:"1.25rem",height:"1.25rem",className:"inline-block mx-1",title:n}),(i.startsWith("CC0")||i.startsWith("CC-0")||i.includes("ZERO"))&&(0,D.jsx)(e2,{width:"1.25rem",height:"1.25rem",className:"inline-block mr-1",title:"CC0: Work is in the worldwide public domain"}),i.includes("BY")&&(0,D.jsx)(Y9,{width:"1.25rem",height:"1.25rem",className:"inline-block mr-1",title:"Credit must be given to the creator"}),i.includes("NC")&&(0,D.jsx)(q9,{width:"1.25rem",height:"1.25rem",className:"inline-block mr-1",title:"Only noncommercial uses of the work are permitted"}),i.includes("SA")&&(0,D
[…]
