+++ build/routes/_index-SM2DH6C7.js
@@ -1,0 +1,1 @@
+import{a as m,d as u}from"/build/_shared/chunk-NFM2H3KQ.js";import{Ba as f,a as w}from"/build/_shared/chunk-XWET6RJ7.js";import"/build/_shared/chunk-HBJK6BW3.js";import"/build/_shared/chunk-HYMQ7M2K.js";import"/build/_shared/chunk-OHOXABTA.js";import"/build/_shared/chunk-OCWQY3HK.js";import"/build/_shared/chunk-C7FW3E47.js";import"/build/_shared/chunk-3CVK3PYF.js";import"/build/_shared/chunk-J6FHCSRC.js";import"/build/_shared/chunk-ND43KHSX.js";import"/build/_shared/chunk-GUCIBHGO.js";import"/build/_shared/chunk-OCTKKCIL.js";import"/build/_shared/chunk-UAI5KRM7.js";import{e as g}from"/build/_shared/chunk-2NH4LW52.js";var d=g(w());var j=({data:o,location:l})=>{var r,i,n,s,a,p,c;if(!o)return[];let t=o.config,e=o.project;return f({origin:"",url:l.pathname,title:(r=t==null?void 0:t.title)!=null?r:e.title,description:(n=(i=t.description)!=null?i:e.description)!=null?n:void 0,image:(s=e.thumbnailOptimized||e.thumbnail)!=null?s:void 0,keywords:(p=(a=t.keywords)!=null?a:e.keywords)!=null?p:[],twitter:(c=t==null?void 0:t.options)==null?void 0:c.twitter})},x=()=>[m];var M=u;export{M as default,x as links,j as meta};
