### 2025-10-07 ebb0d85
[pre-commit.ci] pre-commit autoupdate (#558)
updates:
- https://github.com/psf/black → https://github.com/psf/black-pre-commit-mirror
- [github.com/keewis/blackdoc: v0.4.2 → v0.4.3](https://github.com/keewis/blackdoc/compare/v0.4.2...v0.4.3)
- [github.com/PyCQA/isort: 6.0.1 → 6.1.0](https://github.com/PyCQA/isort/compare/6.0.1...6.1.0)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
