+++ build/git-branches-a0df3bc6c869aecdef6681384bd9929a.md
@@ -1,0 +1,310 @@
+# Git Branches
+
+```{image} ../../images/Git-Logo-2Color.png
+:alt: Git Logo
+:width: 400px
+```
+
+Git "branches" are an important component of many Git and GitHub workflows. If you plan to use GitHub to manage your own resources, or contribute to a GitHub hosted project, it is essential to have a basic understanding of what branches are and how to use them. For example, the best practices for a simple workflow for suggesting changes to a GitHub repository are: create your own fork of the repository, make a branch from your fork where your changes are made, and then suggest these changes move to the upstream repository with a Pull Request. This section of the GitHub chapter assumes you have read the prior GitHub sections, are at least somewhat familiar with git commands and the vocabulary ("cloning," "forking," "merging," "Pull Request" etc), and that you have already created your own fork of the [GitHub Sandbox Repository](https://github.com/ProjectPythia/github-sandbox) hosted by Project Pythia.
+
+## Overview:
+
+1. What are Git Branches
+1. Creating a New Branch
+1. Switching Branches
+1. Setting up a Remote Branch
+1. Merging Branches
+1. Deleting Branches
+1. Updating Your Branches
+1. Complete Workflow
+
+## Prerequisites
+
+| Concepts                                                   | Importance  | Notes                        |
+| ---------------------------------------------------------- | ----------- | ---------------------------- |
+| [What is GitHub?](what-is-github)                          | Necessary   | GitHub user account required |
+| [GitHub Repositories](github-repos)                        | Necessary   |                              |
+| [Issues and Discussions](github-issues)                    | Recommended |                              |
+| [Cloning and Forking a Repository](github-cloning-forking) | Necessary   |                              |
+| [Configuring your GitHub Account](github-setup-advanced)   | Recommended |                              |
+| [Basic Version Control with _git_](basic-git)              | Necessary   |                              |
+
+- **Time to learn**: 30 minutes
+
+---
+
+## What are Git branches?
+
+Git branches allow for non-linear or differing revision histories of a repository. At a point in time, you can split your repository into multiple development paths (branches) where you can make different commits in each, typically with the ultimate intention of merging these branches and development changes together at a later time.
+
+Branching is one of git's methods for helping with collaborative document editing, much like "change tracking" in Google Docs or Microsoft Word. It enables multiple people to edit copies of the same document content, while reducing or managing edit collisions, and with the ultimate aim of merging everyone's changes together later. It also allows the same person to edit multiple copies of the same document, but with different intentions. Some reasons for wanting to split your repository into multiple paths (i.e. branches) is to experiment with different methods of solving a problem (before deciding which method will ultimately be merged) and to work on different problems within the same codebase (without confusing which code changes are relevant to which problem).
+
+These branches can live on your computer (local) or on GitHub (remote). They are brought together through Git _pushes_, _pulls_, _merges_, and _Pull Requests_. _Pushing_ is how you transfer changes from your local repository to a remote repository. _Pulling_ is how you fetch upstream changes into your branch. _Merging_ is how you piece the forked history back together again (i.e. join two branches). And _Pull Requests_ are how you suggest the changes you've made on your branch to the upstream codebase.
+
+```{admonition} Pull Requests
+:class: info
+We will cover [Pull Requests]((github-pull-request)) more in-depthly in the next section.
+```
+
[…]
