+++ _preview/559/build/_shared/timeline-definition-8e5a9bc6-CLYOC45O.js
@@ -1,62 +1,0 @@
-import{a as Mt,b as $t,c as G,d as Y,g as Lt,h as st,i as rt,j as at,l as At,m as Pt,n as Ht,o as E,u as D,v as ot,w as ct,x as lt,y as ht}from"/_preview/559/build/_shared/chunk-EMBXDKWE.js";import"/_preview/559/build/_shared/chunk-OHOXABTA.js";import{e as F}from"/_preview/559/build/_shared/chunk-2NH4LW52.js";var oe=F(Mt()),ce=F($t()),le=F(Lt());var xe=F(At()),_e=F(Pt()),be=F(Ht());var tt=function(){var i=function(k,r,a,h){for(a=a||{},h=k.length;h--;a[k[h]]=r);return a},t=[1,2],e=[1,5],s=[6,9,11,17,18,20,22,23,26,27,28],n=[1,15],c=[1,16],l=[1,17],p=[1,18],g=[1,19],d=[1,23],f=[1,24],v=[1,27],_=[4,6,9,11,17,18,20,22,23,26,27,28],b={trace:function(){},yy:{},symbols_:{error:2,start:3,timeline:4,document:5,EOF:6,directive:7,line:8,SPACE:9,statement:10,NEWLINE:11,openDirective:12,typeDirective:13,closeDirective:14,":":15,argDirective:16,title:17,acc_title:18,acc_title_value:19,acc_descr:20,acc_descr_value:21,acc_descr_multiline_value:22,section:23,period_statement:24,event_statement:25,period:26,event:27,open_directive:28,type_directive:29,arg_directive:30,close_directive:31,$accept:0,$end:1},terminals_:{2:"error",4:"timeline",6:"EOF",9:"SPACE",11:"NEWLINE",15:":",17:"title",18:"acc_title",19:"acc_title_value",20:"acc_descr",21:"acc_descr_value",22:"acc_descr_multiline_value",23:"section",26:"period",27:"event",28:"open_directive",29:"type_directive",30:"arg_directive",31:"close_directive"},productions_:[0,[3,3],[3,2],[5,0],[5,2],[8,2],[8,1],[8,1],[8,1],[7,4],[7,6],[10,1],[10,2],[10,2],[10,1],[10,1],[10,1],[10,1],[10,1],[24,1],[25,1],[12,1],[13,1],[16,1],[14,1]],performAction:function(r,a,h,u,y,o,$){var x=o.length-1;switch(y){case 1:return o[x-1];case 3:this.$=[];break;case 4:o[x-1].push(o[x]),this.$=o[x-1];break;case 5:case 6:this.$=o[x];break;case 7:case 8:this.$=[];break;case 11:u.getCommonDb().setDiagramTitle(o[x].substr(6)),this.$=o[x].substr(6);break;case 12:this.$=o[x].trim(),u.getCommonDb().setAccTitle(this.$);break;case 13:case 14:this.$=o[x].trim(),u.getCommonDb().setAccDescription(this.$);break;case 15:u.addSection(o[x].substr(8)),this.$=o[x].substr(8);break;case 19:u.addTask(o[x],0,""),this.$=o[x];break;case 20:u.addEvent(o[x].substr(2)),this.$=o[x];break;case 21:u.parseDirective("%%{","open_directive");break;case 22:u.parseDirective(o[x],"type_directive");break;case 23:o[x]=o[x].trim().replace(/'/g,'"'),u.parseDirective(o[x],"arg_directive");break;case 24:u.parseDirective("}%%","close_directive","timeline");break}},table:[{3:1,4:t,7:3,12:4,28:
[…]
