+++ _preview/560/build/_shared/dist-IQDPXIQV.js
@@ -1,0 +1,1 @@
+import{a as d,b as u}from"/_preview/560/build/_shared/chunk-6IQOW7V3.js";import"/_preview/560/build/_shared/chunk-3CVK3PYF.js";import{d as p,e as c}from"/_preview/560/build/_shared/chunk-ND43KHSX.js";import"/_preview/560/build/_shared/chunk-GUCIBHGO.js";import"/_preview/560/build/_shared/chunk-2NH4LW52.js";var m={name:"exercise",alias:["exercise-start"],arg:{type:"myst"},options:{...d("exercise"),nonumber:{type:Boolean,doc:"Legacy flag to disable numbering of exercises; equivalent to `enumerated: false`"},hidden:{type:Boolean}},body:{type:"myst"},run(e){var n,o,t,r,s;let l=[];e.arg&&l.push({type:"admonitionTitle",children:e.arg}),e.body&&l.push(...e.body);let i={type:"exercise",hidden:(n=e.options)===null||n===void 0?void 0:n.hidden,children:l};u(e,i),((o=e.options)===null||o===void 0?void 0:o.nonumber)!==void 0?i.enumerated=!e.options.nonumber:i.enumerated=(r=(t=e.options)===null||t===void 0?void 0:t.enumerated)!==null&&r!==void 0?r:!0;let v=i.enumerated?`exercise-${p()}`:void 0,y=((s=e.options)===null||s===void 0?void 0:s.label)||v,{label:x,identifier:b}=c(y)||{};return i.label=x,i.identifier=b,e.name.endsWith("-start")&&(i.gate="start"),[i]}},a={name:"solution",alias:["solution-start"],arg:{type:String,required:!0},options:{...d("solution"),hidden:{type:Boolean}},body:{type:"myst"},run(e){var n;let o=[];if(e.arg){let{label:r,identifier:s}=c(e.arg)||{};o.push({type:"admonitionTitle",children:[{type:"text",value:"Solution to "},{type:"crossReference",label:r,identifier:s}]})}e.body&&o.push(...e.body);let t={type:"solution",hidden:(n=e.options)===null||n===void 0?void 0:n.hidden,children:o};return u(e,t),e.name.endsWith("-start")&&(t.gate="start"),[t]}},f={name:"solution-end",run:()=>[{type:"solution",gate:"end"}]},h={name:"exercise-end",run:()=>[{type:"exercise",gate:"end"}]},g=[m,h,a,f];export{m as exerciseDirective,g as exerciseDirectives,a as solutionDirective};
