+++ _preview/560/build/_shared/factor-I3GXM5GL.js
@@ -1,0 +1,1 @@
+import{a}from"/_preview/560/build/_shared/chunk-5VPNU2DV.js";import"/_preview/560/build/_shared/chunk-2NH4LW52.js";export default a();
