### 2025-10-21 a58aa30
[pre-commit.ci] pre-commit autoupdate (#560)
updates:
- [github.com/keewis/blackdoc: v0.4.3 → v0.4.4](https://github.com/keewis/blackdoc/compare/v0.4.3...v0.4.4)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
