### 2025-10-27 e56a384
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/keewis/blackdoc: v0.4.4 → v0.4.5](https://github.com/keewis/blackdoc/compare/v0.4.4...v0.4.5)
