+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
       - id: black
 
   - repo: https://github.com/keewis/blackdoc
-    rev: v0.4.4
+    rev: v0.4.5
     hooks:
       - id: blackdoc
 
