+++ _preview/561/build/_shared/dist-GZM3ZWIY.js
@@ -1,0 +1,1 @@
+import{a,b,c,d,e,f,g,h,i,j,k,l,m}from"/_preview/561/build/_shared/chunk-OWXZLTDO.js";import"/_preview/561/build/_shared/chunk-X4YK4Z35.js";import"/_preview/561/build/_shared/chunk-ESBRTUPV.js";import"/_preview/561/build/_shared/chunk-2NH4LW52.js";m();export{k as autoCloseTags,d as completionPath,l as esLint,j as javascript,f as javascriptLanguage,h as jsxLanguage,c as localCompletionSource,e as scopeCompletionSource,a as snippets,i as tsxLanguage,g as typescriptLanguage,b as typescriptSnippets};
