+++ build/xarray-4e7483826d9a835f415dc6e924780f95.md
@@ -1,19 +1,0 @@
-# Xarray
-
-![xarray Logo](https://docs.xarray.dev/en/stable/_static/Xarray_Logo_RGB_Final.svg "xarray Logo")
-
-This section contains tutorials on using [Xarray][xarray home]. Xarray {cite:p}`hoyer2017xarray` is used widely in the geosciences and beyond for analysis of gridded N-dimensional datasets.
-
----
-
-From the [Xarray website][xarray home]:
-
-> Xarray (formerly Xray) is an open source project and Python package that makes working with labelled multi-dimensional arrays simple, efficient, and fun!
->
-> Xarray introduces labels in the form of dimensions, coordinates and attributes on top of raw NumPy-like arrays, which allows for a more intuitive, more concise, and less error-prone developer experience. The package includes a large and growing library of domain-agnostic functions for advanced analytics and visualization with these data structures.
->
-> Xarray is inspired by and borrows heavily from pandas, the popular data analysis package focused on labelled tabular data. It is particularly tailored to working with netCDF files, which were the source of xarray’s data model, and integrates tightly with dask for parallel computing.
-
-You should have a basic familiarity with [Numpy arrays](numpy) prior to working through the Xarray notebooks presented here.
-
-[xarray home]: http://xarray.pydata.org/en/stable/
