+++ build/jupyter-34068945c77d447f75fef8ea9445bada.md
@@ -1,123 +1,0 @@
-# Python in Jupyter
-
----
-
-## Overview
-
-You'd like to learn to run Python in a Jupyter session. Here we will cover:
-
-1.  Installing Python in Jupyter
-2.  Running Python code in Jupyter
-3.  Saving your notebook and exiting
-
-## Prerequisites
-
-| Concepts                                                                                                  | Importance | Notes |
-| --------------------------------------------------------------------------------------------------------- | ---------- | ----- |
-| [Installing and Running Python](https://foundations.projectpythia.org/foundations/how-to-run-python) | Helpful    |       |
-
-- **Time to learn**: 20 minutes
-
----
-
-(installing-python-in-jupyter)=
-
-## Installing Python in Jupyter
-
-To run a Jupyter session, you will need to install some necessary packages into your {term}`Conda` environment.
-
-Install `miniforge` by following the [instructions for your machine](https://github.com/conda-forge/miniforge).
-
-[Learn more about Conda here](conda.md)
-
-Next, create a {term}`Conda` environment with {term}`Jupyter Lab` installed. In the terminal, type:
-
-```
-$ conda create --name pythia_foundations_env jupyterlab
-```
-
-Test that you have installed everything correctly by first activating your environment and then launching a {term}`Jupyter Lab` session:
-
-```
-$ conda activate pythia_foundations_env
-$ jupyter lab
-```
-
-A new window should open automatically in your default browser. You can change the browser when launching from the terminal with (for example):
-
-```
-jupyter lab —browser=chrome
-```
-
-## Running Python in Jupyter
-
-1. With your Conda environment activated and Jupyter session launched (see above), create a directory to store our work. Let's call it `pythia-foundations`.
-
-   ![Jupyter GUI](../images/jupyter_gui.png)
-
-   You can do this in the GUI left sidebar by clicking the new-folder icon. If you prefer to use the command line, you can access a terminal by clicking the icon under the "Other" heading in the Launcher.
-
-2. Create a new `mysci.ipynb` file within the `pythia-foundations` folder:
-
-   Do this in the GUI on the left sidebar by clicking the "+" icon.
-
-   This will open a new launcher window where you can select a Python kernel under the "Notebooks" heading for your project. _You should see "Python 3" as in the screenshot above._ Depending on the details of your system, you might see some additional buttons with different kernels.
-
-   Selecting a kernel will open a Jupyter notebook instance and add an untitled file to the left sidebar navigator, which you can then rename to `mysci.ipynb`.
-
-   Select "Python 3" to use the Python version you just installed in the `pythia_foundations_env` conda environment.
-
-3. Change the first notebook cell to include the classic first command: printing, "Hello, world!".
-
-   ```python
-   print("Hello, world!")
-   ```
-
-4. Run your cell with {kbd}`Shift`\+{kbd}`Enter` and see that the results are printed below the cell.
-
-   ![Jupyter - Hello World](../images/mysci.png)
-
-**Congratulations!** You have just set up your first Python environment and run your first Python code in a Jupyter notebook.
-
-<div class="admonition alert alert-info">
-    <p class="admonition-title" style="font-weight:bold">Info</p>
-    There are a few code-syntax differences between running Python in a Jupyter notebook and a script:
-    <ul>
-        <li>In a Python script, to print a variable to the terminal, you need to call <code>print([THING YOU WANT TO PRINT])</code>, whereas in Jupyter, you can simply have the last line of a code cell be <code>[THING YOU WANT TO PRINT]</code>.</li>
[…]
