+++ build/numpy-0ddf840df3ee66a1ae8767429c93641f.md
@@ -1,0 +1,16 @@
+# NumPy
+
+:::{image} https://github.com/numpy/numpy/raw/main/branding/logo/primary/numpylogo.svg
+:alt: NumPy Logo
+:width: 250
+:::
+
+This section contains tutorials on array computing with [NumPy](https://numpy.org) {cite:p}`harris2020array`.
+
+---
+
+From the [NumPy documentation](https://numpy.org/doc/stable/user/whatisnumpy.html):
+
+> NumPy is the fundamental package for scientific computing in Python. It is a Python library that provides a multidimensional array object, various derived objects (such as masked arrays and matrices), and an assortment of routines for fast operations on arrays, including mathematical, logical, shape manipulation, sorting, selecting, I/O, discrete Fourier transforms, basic linear algebra, basic statistical operations, random simulation, and much more.
+
+NumPy's position at the center of the scientific Python ecosystem means that all users should start here in their learning journey through the core scientific packages.
