+++ build/landing-page-4f747ec55f6df1caed753fda8f3f3425.md
@@ -1,39 +1,0 @@
-# Pythia Foundations
-
-## A community learning resource for Python-based computing in the geosciences
-
-[![DOI](https://zenodo.org/badge/338145160.svg)](https://zenodo.org/badge/latestdoi/338145160)
-[![status](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15/status.svg)](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15)
-
-:::{image} images/pretty-earth.png
-:alt: Pretty Earth
-:width: 600px
-:align: center
-:::
-
-This collection covers the foundational skills everyone needs to get started with scientific computing in the open-source Python ecosystem.
-
-:::::{grid} 2 2 2 2
-::::{card}
-:link: https://projectpythia.org
-:::{image} images/logos/ProjectPythia_Logo_Final-01-Blue.svg
-:alt: Project Pythia logo
-:width: 80px
-:align: center
-:::
-::::
-
-::::{card}
-:link: https://pangeo.io
-:::{image} https://raw.githubusercontent.com/pangeo-data/pangeo/master/docs/_static/small_e_logo_cropped.png
-:alt: Pangeo logo
-:width: 80px
-:align: center
-:::
-::::
-:::::
-
-:::{div}
-:class: col-page-middle
-_Brought to you by [Project Pythia](https://projectpythia.org), the education working group for [Pangeo](https://pangeo.io)_
-:::
