### 2025-11-07 #308 @jatkinson1000
@brian-rose @acocac My busyness persisted through October but I am making an effort to chip away at this.
I will place some interim comments and feedback below. I have currently been addressing technical aspects and will then move on to the educational content itself.

My general feeling is that this is a very useful resource, and indeed I have found myself using it to brush up on my Cartopy as part of the [30DayMapChallenge](https://30daymapchallenge.com/).
I think it is well-suited to the JOSE philosophy and ecosystem and imagine I will recommend eventual publication following full review. Whilst introducing basic concepts in Python programming it builds on this to prepare researchers in the geoscience space with the tools they need to perform quality effective research. I support this as one of the most common issues I see is researchers simply not knowing about the tools they need in the open-source ecosystem.

- **Authorship**: Has the submitting author made visible contributions to the module? Does the full list of authors seem appropriate and complete?
    - The authorship list [remains your choice](https://openjournals.readthedocs.io/en/jose/review_criteria.html#authorship), and I am very much a supporter of the [all contributors](https://allcontributors.org/) philosophy recognising that not all contributions can be measured in lines of code. However, based on the available evidence I would raise:
        - jnmorley is not included, but has made larger and more relevant GitHub contributions that some named authors.
        - kafitzgerald made typo and link changes, were there other contributions?
        - bl-freeman seems to have made similar (repository) contributions to kafitzgerald but is not listed
        - Members of CISL and NVIDIA have no repo contributions, but I presume provided invaluable technical support.
           You may want to consider using all-contributors, or listing people on the community pages, but this is simply for you consideration and not required for publication.
    - If you can confirm you are happy I'll check off the authorship.
- **Installation instructions**: Is there a clearly stated list of dependencies?
    - Repo Installation
        - This is an online resource so I'm happy to waive missing dependencies list on the repo. There are some conda environment installation details as part of the contributor's guide for those wishing to install and develop locally.
        - However, I cannot find a clear list of developer dependencies in the docs. I was only able to see this by looking at the conda environment file. It is, of course, possible I missed this as the docs are extensive in which case my apologies.
    - Using/Examples:
        - Setup instructions suggest using anaconda or miniconda.
            - Given the changes to the anaconda licensing last year (see e.g. [anaconda academia press release](https://www.anaconda.com/blog/update-on-anacondas-terms-of-service-for-academia-and-research), [datacamp](https://www.datacamp.com/blog/navigating-anaconda-licensing) or [plain Python](https://python.plainenglish.io/anacondas-2024-licensing-changes-and-alternatives-e63b25741000)) I am glad that you promote miniconda and conda-forge here. (I also see the environment file is set up to use conda-forge.). However, the pros-cons do not discuss this point, instead focussing on installation size and GUI/IDE provision.
            - Whist this is an education resource so is covered by the license changes, it is intended to educate researchers and research activities are no longer covered. I feel this should be made clear.
    - Is there a reason there is no discussion of virtual environments and pip installation? This is, arguably, an equally if not more popular approach to installing python dependencies, and I'd expect researchers to encounter this in the wild.
        -  I accept that it does not make sense to package the foundations project this way as installation would be much more challenging given external dependencies like NetCDF, ffmpeg, especially for newcomers to software and computing.
        - Often a smaller research package will not be available on conda-forge and instead needs to be installed from a local clone using pip, even if this is from within conda.
    - The expectation seems to be that users install the entire conda environment at the start and then work from this for all exercises.
        - In other contexts users should be creating their own environments/packaging for their projects and not rely on using the one from this resource [which is discussed](https://foundations.projectpythia.org/foundations/conda/#conda-commands).
        - I would suggest you consider, for clarity, listing specific dependencies for each notebook/module at the start of that section to hammer home the concepts of dependencies and environments. This would also help those who want to work through a specific chapter or section rather than following from the very beginning.
- **References**
    - There is a citation for the original FAIR principles in data. However, perhaps consider also citing the FAIR-4-Research-Software as well, as this is applicable in my opinion: https://doi.org/10.1038/s41597-022-01710-x
