+++ _preview/562/build/_shared/vbscript-html-YYZDKXRC.js
@@ -1,0 +1,1 @@
+import{c as t}from"/_preview/562/build/_shared/chunk-2NH4LW52.js";var i=t((a,n)=>{function e(s){return{name:"VBScript in HTML",subLanguage:"xml",contains:[{begin:"<%",end:"%>",subLanguage:"vbscript"}]}}n.exports=e});export default i();
