### 2025-11-10 324032d
[pre-commit.ci] pre-commit autoupdate (#562)
