+++ build/root-ZJZ4LRPY.js
@@ -1,1 +1,0 @@
-import{K as u,Q as we,a as Ke,ba as ue,ca as fe,d as Y,f as ge,ia as xe,ja as ye,k as me,ma as ve,o as be,oa as je,pa as ke,ya as ze}from"/build/_shared/chunk-N2N6M3QH.js";import"/build/_shared/chunk-HBJK6BW3.js";import"/build/_shared/chunk-HYMQ7M2K.js";import"/build/_shared/chunk-OHOXABTA.js";import"/build/_shared/chunk-OCWQY3HK.js";import"/build/_shared/chunk-CPTH56EW.js";import{a as he}from"/build/_shared/chunk-3CVK3PYF.js";import"/build/_shared/chunk-J6FHCSRC.js";import{r as A,s as k,t as N}from"/build/_shared/chunk-S4SWV34C.js";import{a as ce}from"/build/_shared/chunk-GUCIBHGO.js";import{f as le,l as pe}from"/build/_shared/chunk-OCTKKCIL.js";import{a as P,d as I}from"/build/_shared/chunk-UAI5KRM7.js";import{e as f}from"/build/_shared/chunk-2NH4LW52.js";var Te="/build/_assets/app-PKKJHBHV.css";var _e="/build/_assets/thebe-core-VKVHG5VY.css";var yt=f(Ke());function Z(o,e,t,r){function i(a){return a instanceof t?a:new t(function(d){d(a)})}return new(t||(t=Promise))(function(a,d){function n(p){try{l(r.next(p))}catch(g){d(g)}}function s(p){try{l(r.throw(p))}catch(g){d(g)}}function l(p){p.done?a(p.value):i(p.value).then(n,s)}l((r=r.apply(o,e||[])).next())})}var rt="ENTRIES",Me="KEYS",Le="VALUES",x="",q=class{constructor(e,t){let r=e._tree,i=Array.from(r.keys());this.set=e,this._type=t,this._path=i.length>0?[{node:r,keys:i}]:[]}next(){let e=this.dive();return this.backtrack(),e}dive(){if(this._path.length===0)return{done:!0,value:void 0};let{node:e,keys:t}=H(this._path);if(H(t)===x)return{done:!1,value:this.result()};let r=e.get(H(t));return this._path.push({node:r,keys:Array.from(r.keys())}),this.dive()}backtrack(){if(this._path.length===0)return;let e=H(this._path).keys;e.pop(),!(e.length>0)&&(this._path.pop(),this.backtrack())}key(){return this.set._prefix+this._path.map(({keys:e})=>H(e)).filter(e=>e!==x).join("")}value(){return H(this._path).node.get(x)}result(){switch(this._type){case Le:return this.value();case Me:return this.key();default:return[this.key(),this.value()]}}[Symbol.iterator](){return this}},H=o=>o[o.length-1],ot=(o,e,t)=>{let r=new Map;if(e===void 0)return r;let i=e.length+1,a=i+t,d=new Uint8Array(a*i).fill(t+1);for(let n=0;n<i;++n)d[n]=n;for(let n=1;n<a;++n)d[n*i]=n;return Ie(o,e,t,r,d,1,i,""),r},Ie=(o,e,t,r,i,a,d,n)=>{let s=a*d;e:for(let l of o.keys())if(l===x){let p=i[s-1];p<=t&&r.set(n,[o.get(l),p])}else{let p=a;for(let g=0;g<l.length;++g,++p){let c=l[g],m=d*p,h=m-d,b=i[m],w=Math.max(0,p-t-1),U=Math.min(d-1,p+t);for(let v=w;v<
[…]
