### 2025-11-17 6796563
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/keewis/blackdoc: v0.4.5 → v0.4.6](https://github.com/keewis/blackdoc/compare/v0.4.5...v0.4.6)
