+++ _preview/563/build/_shared/python-repl-7IXW45LL.js
@@ -1,0 +1,1 @@
+import{c as s}from"/_preview/563/build/_shared/chunk-2NH4LW52.js";var t=s((i,n)=>{function a(e){return{aliases:["pycon"],contains:[{className:"meta",starts:{end:/ |$/,starts:{end:"$",subLanguage:"python"}},variants:[{begin:/^>>>(?=[ ]|$)/},{begin:/^\.\.\.(?=[ ]|$)/}]}]}}n.exports=a});export default t();
