+++ build/basic-git-cb899bd295eb56da4ce01afbcc259dc4.md
@@ -1,471 +1,0 @@
-# Basic Version Control with _git_
-
-```{image} ../../images/Git-Logo-2Color.png
-:alt: Git Logo
-:width: 400px
-```
-
-## Overview:
-
-1. The need for version control
-1. Basic git usage
-1. Making your first git commit
-1. Viewing and comparing across the commit history
-
-## Prerequisites
-
-| Concepts                                                   | Importance  | Notes                        |
-| ---------------------------------------------------------- | ----------- | ---------------------------- |
-| [What is GitHub?](what-is-github)                          | Necessary   | GitHub user account required |
-| [GitHub Repositories](github-repos)                        | Necessary   |                              |
-| [Issues and Discussions](github-issues)                    | Recommended |                              |
-| [Cloning and Forking a Repository](github-cloning-forking) | Recommended |                              |
-| [Configuring your GitHub Account](github-setup-advanced)   | Recommended |                              |
-
-- **Time to learn**: 45 minutes
-
----
-
-## About version control and git
-
-### What is version control (and why should we care)?
-
-[Version Control](https://en.wikipedia.org/wiki/Version_control) refers generally to systems for managing changes to documents or files. Version control systems let us keep track of what changes were made to a file, when they were made, and by whom. If you've ever used "Tracked changes" on a Word document with multiple authors, then you've seen a form of version control in action (though NOT one that is well suited to working with computer code!).
-
-The need for version control is particularly acute when _working with computer code_, where small changes to the text can have huge impacts on the results of running the code.
-
-Do you have a directory somewhere on your machine right now with five different versions of a Python script like this?
-
-```
-analysis_script_OLD.py
-analysis_script.py
-analysis_script_09122021.py
-analysis_script_09122021_edit.py
-analysis_script_NEW.py
-```
-
-A Version Control System (VCS) like git will replace this mess with a _well-ordered and labelled history_ of edits that you can freely browse through, and will greatly simplify collaborating with other people on writing new code.
-
-### What is git?
-
-#### Git is not GitHub
-
-That's the first thing to understand. GitHub is a web-based platform for hosting code and collaborating with other people. On the other hand, **git is a command-line Version Control System (VCS)** that you can download and install. It runs on your local computer as well as under the hood on GitHub. You need to understand something about version control with git in order to use many of GitHub's collaboration features.
-
-#### A little history and nomenclature
-
-Git has been around [since the mid-2000s](https://en.wikipedia.org/wiki/Git). It was originally written by Linus Torvalds specifically for use in development of the Linux kernel. Git is [FOSS](https://foundations.projectpythia.org/foundations/github/what-is-github#free-and-open-source-software-foss) and comes pre-installed on many Linux and Mac OS systems.
-
-There are many other VCSs out there. A few that you might encounter in scientific codebases include [Subversion](https://subversion.apache.org), [Mercurial](https://www.mercurial-scm.org), and [CVS](https://en.wikipedia.org/wiki/Concurrent_Versions_System). However, git is overwhelmingly the VCS of choice for open-source projects in the Scientific Python ecosystem these days (as well as among software developers more generally).
-
-There is no universally agreed-upon meaning of the name "git". From the [git project's own README file](https://github.com/git/git/blob/master/README.md):
-
-> The name "git" was given by Linus Torvalds when he wrote the very first version. He described the tool as "the stupid content tracker" and the name as (depending on your mood):
->
[…]
