+++ build/_shared/diagram-S2PKOQOG-T6AMRORV.js
@@ -1,0 +1,24 @@
+import{a as P}from"/build/_shared/chunk-UJXF6HD2.js";import{a as z}from"/build/_shared/chunk-XTAZ4CNP.js";import{a as E}from"/build/_shared/chunk-EJAB6C5B.js";import"/build/_shared/chunk-7ZGJO7S4.js";import"/build/_shared/chunk-TQOV233I.js";import"/build/_shared/chunk-WRUPT7W2.js";import"/build/_shared/chunk-FYSFXZKR.js";import"/build/_shared/chunk-WJCEDSXQ.js";import"/build/_shared/chunk-ZJA2U5VY.js";import"/build/_shared/chunk-AMWGDE74.js";import"/build/_shared/chunk-K5DBCL5N.js";import"/build/_shared/chunk-W25R6CBQ.js";import"/build/_shared/chunk-GLKM22BQ.js";import{l as u}from"/build/_shared/chunk-G2WMYO6W.js";import"/build/_shared/chunk-ILX2UBHN.js";import{L as x,P as y,Q as $,R as B,S as C,T as S,U as D,V as T,q as w,y as v}from"/build/_shared/chunk-TRJHAMCF.js";import"/build/_shared/chunk-UVP6FRQW.js";import{b as h,d as m}from"/build/_shared/chunk-ZHBLEF4E.js";import"/build/_shared/chunk-MBMQ6BQC.js";import"/build/_shared/chunk-OZE3FFNP.js";var N=w.packet,F=class{constructor(){this.packet=[],this.setAccTitle=$,this.getAccTitle=B,this.setDiagramTitle=D,this.getDiagramTitle=T,this.getAccDescription=S,this.setAccDescription=C}static{h(this,"PacketDB")}getConfig(){let t=u({...N,...v().packet});return t.showBits&&(t.paddingY+=10),t}getPacket(){return this.packet}pushWord(t){t.length>0&&this.packet.push(t)}clear(){y(),this.packet=[]}},L=1e4,M=h((t,e)=>{z(t,e);let a=-1,o=[],n=1,{bitsPerRow:l}=e.getConfig();for(let{start:r,end:s,bits:d,label:c}of t.blocks){if(r!==void 0&&s!==void 0&&s<r)throw new Error(`Packet block ${r} - ${s} is invalid. End must be greater than start.`);if(r??=a+1,r!==a+1)throw new Error(`Packet block ${r} - ${s??r} is not contiguous. It should start from ${a+1}.`);if(d===0)throw new Error(`Packet block ${r} is invalid. Cannot have a zero bit field.`);for(s??=r+(d??1)-1,d??=s-r+1,a=s,m.debug(`Packet block ${r} - ${a} with label ${c}`);o.length<=l+1&&e.getPacket().length<L;){let[p,i]=Y({start:r,end:s,bits:d,label:c},n,l);if(o.push(p),p.end+1===n*l&&(e.pushWord(o),o=[],n++),!i)break;({start:r,end:s,bits:d,label:c}=i)}}e.pushWord(o)},"populate"),Y=h((t,e,a)=>{if(t.start===void 0)throw new Error("start should have been set during first phase");if(t.end===void 0)throw new Error("end should have been set during first phase");if(t.start>t.end)throw new Error(`Block start ${t.start} is greater than block end ${t.end}.`);if(t.end+1<=e*a)return[t,void 0];let o=e*a-1,n=e*a;return[{start:t.start,end:o,label:t.label,bits:o-t.start},{start:n,end:t.end,label:t.label,bits:t.end-n}]},"getNextFittingBlock"),A={parser:{yy:void 0},parse:h(async t=>{let e=await E("packet",t),a=A.parser?.yy;if(!(
[…]
