### 2025-11-26 af457e1
add configuration option to minimize warnings about matplotlib output
