### 2025-11-26 6dc7028
fix and/or update a bunch of references
