+++ foundations/github/git-branches.md
@@ -22,7 +22,7 @@ Git "branches" are an important component of many Git and GitHub workflows. If y
 
 | Concepts                                                   | Importance  | Notes                        |
 | ---------------------------------------------------------- | ----------- | ---------------------------- |
-| [What is GitHub?](what-is-github)                          | Necessary   | GitHub user account required |
+| [What is GitHub?](#what-is-github)                         | Necessary   | GitHub user account required |
 | [GitHub Repositories](github-repos)                        | Necessary   |                              |
 | [Issues and Discussions](github-issues)                    | Recommended |                              |
 | [Cloning and Forking a Repository](github-cloning-forking) | Necessary   |                              |
