+++ foundations/github/github-cloning-forking.md
@@ -15,7 +15,7 @@
 
 | Concepts                                | Importance  | Notes                        |
 | --------------------------------------- | ----------- | ---------------------------- |
-| [What is GitHub?](what-is-github)       | Necessary   | GitHub user account required |
+| [What is GitHub?](#what-is-github)      | Necessary   | GitHub user account required |
 | [GitHub Repositories](github-repos)     | Necessary   |                              |
 | [Issues and Discussions](github-issues) | Recommended |                              |
 | Command-line shell                      | Helpful     |                              |
