+++ foundations/github/github-pull-request.md
@@ -18,7 +18,7 @@ A Pull Request, aka a "merge request," is an event that occurs when a project co
 
 | Concepts                                      | Importance  | Notes |
 | --------------------------------------------- | ----------- | ----- |
-| [What is GitHub](what-is-github)              | Necessary   |       |
+| [What is GitHub](#what-is-github)             | Necessary   |       |
 | [GitHub Repositories](github-repos)           | Necessary   |       |
 | [Cloning and Forking](github-cloning-forking) | Necessary   |       |
 | [Basic Version Control with _git_](basic-git) | Necessary   |       |
@@ -148,7 +148,7 @@ The steps leading up to creating your PR depend on your workflow. The main diffe
 the contributor now, for the Feature Workflow, navigates to the upstream, remote
 repository, not a personal remote fork, and initiates the PR there.
 
-We will cover [GitHub Workflows](github-workflows) in greater detail in the next chapter.
+We will cover [GitHub Workflows](#gh-workflows) in greater detail in the next chapter.
 
 ---
 
