+++ foundations/github/github-issues.md
@@ -15,7 +15,7 @@
 
 | Concepts                            | Importance | Notes |
 | ----------------------------------- | ---------- | ----- |
-| [What is GitHub?](what-is-github)   | Necessary  |       |
+| [What is GitHub?](#what-is-github)  | Necessary  |       |
 | [GitHub Repositories](github-repos) | Necessary  |       |
 
 - **Time to learn**: 5 minutes
