+++ foundations/github/review-pr.md
@@ -21,7 +21,7 @@ This tutorial covers the following topics:
 
 | Concepts                                      | Importance  | Notes |
 | --------------------------------------------- | ----------- | ----- |
-| [What is GitHub](what-is-github)              | Necessary   |       |
+| [What is GitHub](#what-is-github)             | Necessary   |       |
 | [GitHub Repositories](github-repos)           | Necessary   |       |
 | [Cloning and Forking](github-cloning-forking) | Necessary   |       |
 | [Basic Version Control with _git_](basic-git) | Necessary   |       |
@@ -118,7 +118,7 @@ There are specific elements of PRs that are more commonly prioritized during a r
 
 ### What's Next?
 
-The next tutorial will cover standards and other details about [GitHub Workflows](github-workflows).
+The next tutorial will cover standards and other details about [GitHub Workflows](#gh-workflows).
 
 ## Resources and References
 
