+++ foundations/github/what-is-github.md
@@ -1,3 +1,4 @@
+(what-is-github)=
 # What is GitHub?
 
 ```{image} ../../images/GitHub-logo.png
