+++ foundations/github/basic-git.md
@@ -14,13 +14,13 @@
 
 ## Prerequisites
 
-| Concepts                                                   | Importance  | Notes                        |
-| ---------------------------------------------------------- | ----------- | ---------------------------- |
-| [What is GitHub?](#what-is-github)                         | Necessary   | GitHub user account required |
-| [GitHub Repositories](github-repos)                        | Necessary   |                              |
-| [Issues and Discussions](github-issues)                    | Recommended |                              |
-| [Cloning and Forking a Repository](github-cloning-forking) | Recommended |                              |
-| [Configuring your GitHub Account](github-setup-advanced)   | Recommended |                              |
+| Concepts                                                      | Importance  | Notes                        |
+| ------------------------------------------------------------- | ----------- | ---------------------------- |
+| [What is GitHub?](what-is-github.md)                          | Necessary   | GitHub user account required |
+| [GitHub Repositories](github-repos.md)                        | Necessary   |                              |
+| [Issues and Discussions](github-issues.md)                    | Recommended |                              |
+| [Cloning and Forking a Repository](github-cloning-forking.md) | Recommended |                              |
+| [Configuring your GitHub Account](github-setup-advanced.md)   | Recommended |                              |
 
 - **Time to learn**: 45 minutes
 
