+++ foundations/github/github-workflows.md
@@ -1,4 +1,3 @@
-(gh-workflows)=
 # GitHub Workflows
 
 ```{image} ../../images/Git-Logo-2Color.png
@@ -16,16 +15,16 @@ A workflow is a series of activities or tasks that must be completed sequentiall
 
 ## Prerequisites
 
-| Concepts                                      | Importance  | Notes |
-| --------------------------------------------- | ----------- | ----- |
-| [What is GitHub](#what-is-github)             | Necessary   |       |
-| [GitHub Repositories](github-repos)           | Necessary   |       |
-| [Cloning and Forking](github-cloning-forking) | Necessary   |       |
-| [Basic Version Control with _git_](basic-git) | Necessary   |       |
-| [Issues and Discussions](github-issues)       | Recommended |       |
-| [Branches](git-branches)                      | Necessary   |       |
-| [Pull Requests](github-pull-request)          | Necessary   |       |
-| [Reviewing Pull Requests](review-pr)          | Recommended |       |
+| Concepts                                         | Importance  | Notes |
+| ------------------------------------------------ | ----------- | ----- |
+| [What is GitHub](what-is-github.md)              | Necessary   |       |
+| [GitHub Repositories](github-repos.md)           | Necessary   |       |
+| [Cloning and Forking](github-cloning-forking.md) | Necessary   |       |
+| [Basic Version Control with _git_](basic-git.md) | Necessary   |       |
+| [Issues and Discussions](github-issues.md)       | Recommended |       |
+| [Branches](git-branches.md)                      | Necessary   |       |
+| [Pull Requests](github-pull-request.md)          | Necessary   |       |
+| [Reviewing Pull Requests](review-pr.md)          | Recommended |       |
 
 - **Time to learn**: 60 minutes
 
