### 2025-11-26 31c5980
add additional citations
