+++ _preview/565/build/_shared/chunk-6YGIEX7D.js
@@ -1,0 +1,1 @@
+import{a as u}from"/_preview/565/build/_shared/chunk-R3ABELDT.js";import{c as a}from"/_preview/565/build/_shared/chunk-OZE3FFNP.js";var d=a((g,s)=>{var o=u();s.exports=e;e.displayName="json5";e.aliases=[];function e(n){n.register(o),function(r){var t=/("|')(?:\\(?:\r\n?|\n|.)|(?!\1)[^\\\r\n])*\1/;r.languages.json5=r.languages.extend("json",{property:[{pattern:RegExp(t.source+"(?=\\s*:)"),greedy:!0},{pattern:/(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*:)/,alias:"unquoted"}],string:{pattern:t,greedy:!0},number:/[+-]?\b(?:NaN|Infinity|0x[a-fA-F\d]+)\b|[+-]?(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:[eE][+-]?\d+\b)?/})}(n)}});export{d as a};
