+++ _preview/565/build/_shared/wren-FGHU2HOT.js
@@ -1,0 +1,1 @@
+import{a}from"/_preview/565/build/_shared/chunk-MOOSZ6D3.js";import"/_preview/565/build/_shared/chunk-OZE3FFNP.js";export default a();
