+++ _preview/565/build/_shared/chunk-UFEOMFEA.js
@@ -1,0 +1,1 @@
+import{L as w}from"/_preview/565/build/_shared/chunk-TRJHAMCF.js";import{b as r,d as c}from"/_preview/565/build/_shared/chunk-ZHBLEF4E.js";var g=r((t,e,i,h)=>{t.attr("class",i);let{width:o,height:n,x,y:u}=s(t,e);w(t,n,o,h);let a=m(x,u,o,n,e);t.attr("viewBox",a),c.debug(`viewBox configured: ${a} with padding: ${e}`)},"setupViewPortForSVG"),s=r((t,e)=>{let i=t.node()?.getBBox()||{width:0,height:0,x:0,y:0};return{width:i.width+e*2,height:i.height+e*2,x:i.x,y:i.y}},"calculateDimensionsWithPadding"),m=r((t,e,i,h,o)=>`${t-o} ${e-o} ${i} ${h}`,"createViewBox");export{g as a};

[…]
