### 2025-12-01 ProjectPythia/pythia-foundations#568 @e-marshall
EM JOSE Review - Preamble
- If you were to include a requirements list, I think it would fit nicely in the [Running Pythia Foundations examples](https://foundations.projectpythia.org/preamble/how-to-use/#running-pythia-foundations-examples) section
- Would it be possible to add links to the foundational skills and core scientific python packages sections when they're referenced in text in the [Organization of Pythia Foundations section](https://foundations.projectpythia.org/preamble/how-to-use/#organization-of-pythia-foundations)?
