### 2025-12-01 ProjectPythia/pythia-foundations#572 @e-marshall
EM JOSE Review - Foundations: General comments
- Could you provide a bit of a roadmap/overview of where we're headed in this section in the [overview page](https://github.com/ProjectPythia/pythia-foundations/blob/main/foundations/overview.md) ? Something letting the reader know that the section starts by explaining why we're using python, before moving into a quickstart example of using python and walking through how to use python on different platforms, etc? It would be helpful to give more detail on what the user can expect to learn by the end of this section
