### 2025-12-01 ProjectPythia/pythia-foundations#573 @e-marshall
EM JOSE Review - Foundations: Python Quickstart
- in the [quickstart notebook](https://foundations.projectpythia.org/foundations/quickstart/)
- Is this how the binder icon is meant to render?

- Would be helpful to link to terms like C, Fortran, etc.
-  Could you flesh out the [Resources and references](https://foundations.projectpythia.org/foundations/quickstart/#resources-and-references) section on this page? I think at least the matplotlib and numpy documentation/tutorial pages would be relevant here, and potentially more, such as other tutorials you find particularly helpful or interesting. Also, jupyter notebook and binder?
- Confused on the jupyter notebook ref in the [Run this code from your browser section](https://foundations.projectpythia.org/foundations/quickstart/#run-this-code-from-your-browser) but maybe I'm missing something here?
