### 2025-12-01 ProjectPythia/pythia-foundations#574 @e-marshall
EM JOSE Review - Foundations: Installing and running Python
##  [Choosing a Python Platform](https://foundations.projectpythia.org/foundations/how-to-run-python/#choosing-a-python-platform)
- Can the bullet points in this section (1. The terminal; 2. Jupyter notebooks, and 3. IDEs...) link to the corresponding section headers on the page? otherwise, its a bit confusing because we haven't been introduced to jupyter yet
