### 2025-12-01 ProjectPythia/pythia-foundations#575 @e-marshall
EM JOSE Review - Foundations: Getting started with Jupyter
## [Main page](https://foundations.projectpythia.org/foundations/getting-started-jupyter/#prerequisites)
- in the [pre reqs](https://foundations.projectpythia.org/foundations/getting-started-jupyter/#prerequisites), [Installing and running Python: Python in Jupyter](https://foundations.projectpythia.org/foundations/jupyter/) is listed, but this page (getting started with jupyter) has relevant introductory info about Project Jupyter and Jupyter Notebooks that I think would be helpful to introduce, or at least point to, back when Jupyter is first used. Could you either link to this in an admonition box or pull out the overview section and place it earlier?
- I like how the four main components of jupyter are laid out in the introduction, but I find the section structure of this page a bit confusing. Instead of each of the components having their own H2 header and a short description, could you explore maybe having a Jupyter Components H2 section, and the individual components inside of it, and then Executing Jupyter as the next H2 section?
- or consider using MyST [dropdown components](https://mystmd.org/guide/dropdowns-cards-and-tabs#dropdowns) here?
- [jupyter notebook sub-section](https://foundations.projectpythia.org/foundations/getting-started-jupyter/#jupyter-notebooks): this is more filled out than the other component sub-sections. I'd suggest possibly adding additional helpful information about the other components and why someone might want to use them, and/or linking to external references
- [jupyter kernel sub-section](https://foundations.projectpythia.org/foundations/getting-started-jupyter/#jupyter-kernels): I think this is a bit terse, a bit more about what kernels are would be helpful, and potentially addressing questions like what does this mean for me as a user and what do I need to know how to do, would be helpful
- [jupyterhub sub-section](https://foundations.projectpythia.org/foundations/getting-started-jupyter/#jupyter-hub): Potentially add a bit more to help users who may not be familiar with remote systems. Something along the lines of... these are a way of connecting to computational resources outside of your local machine, you may encounter them when using high-performance computing (HPC) systems or cloud-computing platforms....
