### 2025-12-01 ProjectPythia/pythia-foundations#576 @e-marshall
EM JOSE Review - Foundations: Getting Started with GitHub 
## [What is GitHub?](https://foundations.projectpythia.org/foundations/github/what-is-github/)
-  in the 'stay tuned' admonition box [here](https://foundations.projectpythia.org/foundations/github/what-is-github/#foss-and-vcs-on-the-internet), I think it would be helpful to link to where these concepts will come up later in the series so that the user can see what you're referencing (ie. '...later in this series in the XX section').
- the references section of this page is a bit sparse, and there are links earlier in the text that aren't included here, could you included all pages linked in this page in the reference section?
