### 2025-12-01 ProjectPythia/pythia-foundations#577 @e-marshall
EM JOSE Review - Core Scientific Packages - Overview
- this [page](https://foundations.projectpythia.org/core/overview/) provides a lot of great info but it doesn't include much about what to expect in the rest of this section. Can you add a text section that gives a bit of a roadmap and helps the reader set expectations about what they can learn in the subsequent tutorials?
- similarly, can you provide a bit more info about what the user can expect within each tutorial
