### 2025-12-01 ProjectPythia/pythia-foundations#577 @e-marshall
➕ ➕ to this paragraph! Could you add a few examples of the operations that pandas and xarray have implementations for so that a user who is very new to the field can start to build a better idea? just very basic things like rolling averages, computing a min or max along a given dimension
