### 2025-12-01 ProjectPythia/pythia-foundations#578 @e-marshall
EM JOSE Review - Core Scientific Packages - Numpy
All three of these pages have great content but I think could be a bit overwhelming, especially for a reader coming straight from the overview page, which also has a lot of information
- can you add a section on the numpy overview/landing page telling the reader what to expect and where to find certain things
- for instance, in NumPy basics, you will learn how to....
	- for examples of slicing and indexing see intermediate...
	- for a detailed walk-through of broadcasting and the math underlying these operations see the broadcasting tutorial.
