### 2025-12-01 ProjectPythia/pythia-foundations#579 @e-marshall
EM JOSE Review - Core Scientific Packages - Matplotlib
- at the bottom of the landing [page](https://foundations.projectpythia.org/core/matplotlib/#in-this-section) where there are bullet points to types of plots and other content included in the book, can you add links so that the user could easily navigate to those sections?
