### 2025-12-01 #308 @e-marshall
Hi all,
thanks for your patience as I chip away at this. I have worked through the majority of the material and will try to get through the rest this week.

This is a comprehensive and detailed resource that will be very helpful for learners across the geosciences. This could accompany a researcher through many stages of learning/ their career and be very helpful at each. I think that this will be very valuable for the open-source scientific community and the geoscience community. The github chapter of the foundations section I think, will be particularly helpful to many users. I would recommend it for publication after some modifications have been made.

below are a few comments on the general requirements and then on specific sections:

## General checks
- I see that there is an overall license for the repository; I think this submission might need a separate category for code and for the text? You can see this [comment](https://github.com/openjournals/jose-reviews/issues/243#issuecomment-2225660468) for additional explanation but in my experience, you would have a LICENSE file for the docs and a LICENSE-CODE, this could be the MIT license) file for the code

## Documentation
- I echo @jatkinson1000 's comments on the dependencies. a list of dependencies more readily found in the jupyter book would be helpful but not critical (See preamble comment below)

## Pedagogy / Instructional design
- **Learning Objectives:** Only one of the tutorials has explicitly stated learning objectives. While not required, I think that they would be very helpful elsewhere throughout the resource, especially one that covers so many topics.
-  **Content scope and length:** The content is substantial; in some places, I think the length of individual notebooks may be too long, and would suggest breaking up the content into multiple notebooks/pages. These are noted in the comments below.
- **Pedagogy**: There are some places where the modules felt difficult to follow and the cognitive load high. I added some specific comments on these instances below with suggestions to add text formatting/different organization to address this.
- **Content quality**: The content is of very good quality, and it is engaging. There is one section that is marked as under construction. I wasn't sure the JOSE guidelines on completeness and in-progress sections, is this acceptable?
-  **Instructional design:** For the most part, I think the instructional design is deliberate and apparent; however, there are some areas where design could be more consistent across pages and sections, and where additional features and formatting could be used to reduce cognitive load.

**Overall comments**
- As stated above, I think that this resource has a very large amount of educational and informative content and demonstrates the objectives and philosophy of JOSE. However, I think that the volume of content can be a bit overwhelming to the reader. I'd suggest:
- adding additional framing text, such as learning objectives (ideally in every multi-page section and notebook, though not critical), to give the user an idea of what they are going to see
- In the overview sections at the top of each page, add hyperlinks to the bullet point section headers that link to the corresponding section further down on the page. (See https://github.com/ProjectPythia/pythia-foundations/issues/574#issue-3680045838 as an example for a single page)
- I'd also suggest additional links throughout the book, whether to external sites or internal references. I think this would help break up the text visually and aid the user in building connections and understanding within the content.
- There are a number of helpful figures in the text; some have in-text references, and some have admonitions with captions. Is there JOSE guidance on this? Do all images need to be treated as 'figures' with numbers, captions, and in-text references, or are more informal images acceptable?

I added a number of section-specific comments in the issues listed at the bottom of this comment. However, some of them verge into the category of nits and suggestions and do not need to be addressed, while some are more critical. The issues that are most important to address are those listed above and the following:

- Discussion of virtual environments, package management, and conda, as mentioned by @jatkinson1000 : https://github.com/ProjectPythia/pythia-foundations/issues/574#issuecomment-3595606049
- Jupyter is mentioned without an introduction or external link before it is properly introduced in the text: https://github.com/ProjectPythia/pythia-foundations/issues/575#issue-3680061516 and https://github.com/ProjectPythia/pythia-foundations/issues/574#issuecomment-3595605854
- Consider breaking the JupyterLab page of the 'getting started with jupyter' section into two notebooks: https://github.com/ProjectPythia/pythia-foundations/issues/575#issuecomment-3595517924
- I think the wording in the github workflows section could easily be confused with github workflows as they pertain to github actions: https://github.com/ProjectPythia/pythia-foundations/issues/576#issuecomment-3595504938
- Broken link: https://github.com/ProjectPythia/pythia-foundations/issues/576#issuecomment-3595503793
- Broken link on the core packages overview page: https://github.com/ProjectPythia/pythia-foundations/issues/577#issuecomment-3595664979

Section-specific issues:
- Preamble: https://github.com/ProjectPythia/pythia-foundations/issues/568
- Foundations:
  - General comments: https://github.com/ProjectPythia/pythia-foundations/issues/572
  - Python quickstart: https://github.com/ProjectPythia/pythia-foundations/issues/573
  - Installing and Running Python: https://github.com/ProjectPythia/pythia-foundations/issues/574
  - Getting started with Jupyter: https://github.com/ProjectPythia/pythia-foundations/issues/575
  - Getting started with GitHub: https://github.com/ProjectPythia/pythia-foundations/issues/576
 - Core Sci. Packages:
 - Overview: https://github.com/ProjectPythia/pythia-foundations/issues/577
 - Matplotlib: https://github.com/ProjectPythia/pythia-foundations/issues/579
 - Numpy: https://github.com/ProjectPythia/pythia-foundations/issues/578
