### 2025-12-03 920af4f
update links in matplotlib.md
