### 2025-12-03 ProjectPythia/pythia-foundations#576 @kafitzgerald
Assigned this to myself since I think it should be a relatively easy lift for me and I think some may be covered by #564, but if someone else is excited to take this on let me know.
