+++ _preview/581/build/_shared/mojolicious-27MQ6CSC.js
@@ -1,1 +1,0 @@
-import{c as n}from"/_preview/581/build/_shared/chunk-OZE3FFNP.js";var a=n((g,e)=>{function u(s){return{name:"Mojolicious",subLanguage:"xml",contains:[{className:"meta",begin:"^__(END|DATA)__$"},{begin:"^\\s*%{1,2}={0,2}",end:"$",subLanguage:"perl"},{begin:"<%{1,2}={0,2}",end:"={0,1}%>",subLanguage:"perl",excludeBegin:!0,excludeEnd:!0}]}}e.exports=u});export default a();
