+++ _preview/581/build/entry.client-PCJPW7TK.js
@@ -1,1 +1,0 @@
-import{o as r}from"/_preview/581/build/_shared/chunk-AQ2CODAG.js";import{a,c as l,d as i}from"/_preview/581/build/_shared/chunk-JJXTQVMA.js";import{f as o}from"/_preview/581/build/_shared/chunk-OZE3FFNP.js";var t=o(a()),m=o(l()),e=o(i());function d(){(0,t.startTransition)(()=>{(0,m.hydrateRoot)(document,(0,e.jsx)(t.StrictMode,{children:(0,e.jsx)(r,{})}))})}window.requestIdleCallback?window.requestIdleCallback(d):window.setTimeout(d,1);

[…]
