### 2025-12-04 7bb13ef
add abbreviations and central glossary along with some minor editorial changes to why-python.md
