### 2025-12-04 0941563
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
