### 2025-12-04 ProjectPythia/pythia-foundations#583 @brian-rose
Add Brittany Freeman as co-author of JOSE manuscript
It was pointed out (correctly) in our JOSE manuscript review https://github.com/openjournals/jose-reviews/issues/308 that @bl-freeman (Brittany Freeman) has made substantial contributions to the Foundations repo and should be included as co-author on the manuscript.

This was just an accident of timing; Brittany did a lot of work on Foundations improvements after the original manuscript submission. We are going to include her now.
