### 2025-12-04 #308 @brian-rose
Responding specifically to the authorship points raised by @jatkinson1000 :
> * **Authorship**: Has the submitting author made visible contributions to the module? Does the full list of authors seem appropriate and complete?
>
>   * The authorship list [remains your choice](https://openjournals.readthedocs.io/en/jose/review_criteria.html#authorship), and I am very much a supporter of the [all contributors](https://allcontributors.org/) philosophy recognising that not all contributions can be measured in lines of code. However, based on the available evidence I would raise:
>
>     * jnmorley is not included, but has made larger and more relevant GitHub contributions that some named authors.
>     * kafitzgerald made typo and link changes, were there other contributions?
>     * bl-freeman seems to have made similar (repository) contributions to kafitzgerald but is not listed
>     * Members of CISL and NVIDIA have no repo contributions, but I presume provided invaluable technical support.
>       You may want to consider using all-contributors, or listing people on the community pages, but this is simply for you consideration and not required for publication.
>   * If you can confirm you are happy I'll check off the authorship.

In deciding who to include in the co-author list for this manuscript, we had to look back over four years of work on Project Pythia and Pythia Foundations in particular. It is a large collaboration and people have come and gone from the team over the lifetime of the project. The intent here was to include as co-authors all the individuals who contributed ideas, effort, or code that led to the current form of Pythia Foundations. Some of these contributions occurred through brainstorm sessions on mapping out content. Other contributions occurred in the form of design and development of the shared infrastructure used by Project Pythia to create, maintain, and host our content (including Foundations). Much of this work is not reflected in Pull Requests on the Foundations repo itself.

A couple of comments are in order about specific individuals pointed out by the reviewer:

@bl-freeman should definitely be included. This was an oversight that occurred because Brittany's contribution came after the original manuscript was submitted to JOSE. It is being addressed in https://github.com/ProjectPythia/pythia-foundations/issues/583.

@jnmorley **should** also be included based on past contributions, which occurred several years ago as part of a summer internship. We made a good faith effort to get it touch with James and ask for his consent to be included, but we did not ever manage to get in contact. We'd still be happy to include him if we can reach him.

Great points from the reviewer, thanks @jatkinson1000!
