### 2025-12-04 573f972
Quickstart: add links to C/Fortan and np/plt tutorials
