### 2025-12-04 ProjectPythia/pythia-foundations#573 @r-ford
For the first screenshot, that looks like the [MyST in-page execution](https://mystmd.org/guide/in-page-execution), but I'm not sure if that currently works. The link in the second screenshot should be fixed by #564.
