### 2025-12-04 92aad79
Add Brittany Freeman to author list (#584)
