+++ build/manifest-51D9907C.js
@@ -1,1 +1,0 @@
-window.__remixManifest={"entry":{"module":"/build/entry.client-PCJPW7TK.js","imports":["/build/_shared/chunk-AQ2CODAG.js","/build/_shared/chunk-JJXTQVMA.js","/build/_shared/chunk-OZE3FFNP.js"]},"routes":{"root":{"id":"root","path":"","module":"/build/root-TRTRBUUN.js","imports":["/build/_shared/chunk-UGO2RSJF.js","/build/_shared/chunk-HHKX4TLX.js","/build/_shared/chunk-YATQFWLS.js","/build/_shared/chunk-MBMQ6BQC.js","/build/_shared/chunk-OCWQY3HK.js","/build/_shared/chunk-CVDX4WR7.js","/build/_shared/chunk-3EBOCCHJ.js","/build/_shared/chunk-O4VQNZ62.js","/build/_shared/chunk-NK7RXO5Q.js","/build/_shared/chunk-GUCIBHGO.js"],"hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":true},"routes/$":{"id":"routes/$","parentId":"root","path":"*","module":"/build/routes/$-VLPGWU3Z.js","imports":["/build/_shared/chunk-6NUWI4AQ.js","/build/_shared/chunk-M5KN76GI.js"],"hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":true},"routes/($project)_.($a).($b).($c).($d).$slug[.json]":{"id":"routes/($project)_.($a).($b).($c).($d).$slug[.json]","parentId":"root","path":":project?/:a?/:b?/:c?/:d?/:slug.json","module":"/build/routes/($project)_.($a).($b).($c).($d).$slug[.json]-DIB2VQ4M.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[favicon.ico]":{"id":"routes/[favicon.ico]","parentId":"root","path":"favicon.ico","module":"/build/routes/[favicon.ico]-MH4XVVZI.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[objects.inv]":{"id":"routes/[objects.inv]","parentId":"root","path":"objects.inv","module":"/build/routes/[objects.inv]-RUZS76RX.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[robots.txt]":{"id":"routes/[robots.txt]","parentId":"root","path":"robots.txt","module":"/build/routes/[robots.txt]-JULN6D72.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[sitemap.xml]":{"id":"routes/[sitemap.xml]","parentId":"root","path":"sitemap.xml","module":"/build/routes/[sitemap.xml]-NGYNA2IS.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[sitemap_style.xsl]":{"id":"routes/[sitemap_style.xsl]","parentId":"root","path":"sitemap_style.xsl","module":"/build/routes/[sitemap_style.xsl]-GJUCKAG6.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/_index":{"id":"routes/_index","parentId":"root","index":true,"module":"/build/routes/_index-6NG6KCJM.js","imports":["/build/_shared/chunk-6NUWI4AQ.js","/build/_shared/chunk-M5KN76GI.js"],"hasAction":false,"hasLo
[…]

[…]
