### 2025-12-05 1ad2688
add references to BibTeX file to avoid lookup and fix bibliography config
