+++ references.bib
@@ -75,3 +75,84 @@ url = {https://scitools.org.uk/cartopy}
     doi          = {10.5281/zenodo.3509134},
     url          = {https://doi.org/10.5281/zenodo.3509134}
 }
+
+@inbook{Kluyver_Thomas_2016,
+	author = {Thomas, Kluyver and Benjamin, Ragan-Kelley and Fernando, P\&eacute;rez and Brian, Granger and Matthias, Bussonnier and Jonathan, Frederic and Kyle, Kelley and Jessica, Hamrick and Jason, Grout and Sylvain, Corlay and Paul, Ivanov and Dami\&aacute;n, Avila and Safia, Abdalla and Carol, Willing and Team, Jupyter Development},
+	booktitle = {Positioning and {Power} in {Academic} {Publishing}: Players, {Agents} and {Agendas}},
+	doi = {10.3233/978-1-61499-649-1-87},
+	year = {2016},
+	publisher = {IOS Press},
+	title = {Jupyter {Notebooks} \&amp;ndash; a publishing format for reproducible computational workflows},
+	url = {http://dx.doi.org/10.3233/978-1-61499-649-1-87},
+}
+
+
+@inproceedings{Jupyter_2025,
+	author = {Jupyter, Project and Bolyen, Evan and Caporaso, J Gregory and Cockett, Rowan and Garside, Danny and Holdgraf, Chris and Hollands, Angus and Kent, Julia and Koch, Franklin and Madge, Jim and McKay, Matthew and Morrison, Mike and P{\' e}rez, Fernando and Purves, Steve and Kelley, Min Ragan and Rose, Brian E. J. and Sharan, Malvika and Sip{\H o}cz, Brigitta M. and Walt, St{\' e}fan J. and Whitaker, Kirstie Jane},
+	series = {SciPy},
+	booktitle = {Proceedings of the 24th {Python} in {Science} {Conference}},
+	doi = {10.25080/hwcj9957},
+	issn = {2575-9752},
+	year = {2025},
+	month = {7},
+	pages = {173--193},
+	organization = {SciPy},
+	title = {Jupyter {Book} 2 and the {MyST} {Document} {Stack}: A modular, extensible, web-native stack for authoring and publishing computational narratives},
+	url = {http://dx.doi.org/10.25080/hwcj9957},
+}
+
+
+@article{Granger_2021,
+	author = {Granger, Brian E. and Perez, Fernando},
+	journal = {Computing in Science \&amp; Engineering},
+	doi = {10.1109/mcse.2021.3059263},
+	issn = {1558-366X},
+	number = {2},
+	year = {2021},
+	month = {3},
+	pages = {7--14},
+	publisher = {Institute of Electrical},
+	title = {Jupyter: Thinking and {Storytelling} {With} {Code} and {Data}},
+	url = {http://dx.doi.org/10.1109/MCSE.2021.3059263},
+	volume = {23},
+}
+
+
+@article{Rew_1990,
+	author = {Rew, R. and Davis, G.},
+	journal = {IEEE Computer Graphics and Applications},
+	doi = {10.1109/38.56302},
+	issn = {0272-1716},
+	number = {4},
+	year = {1990},
+	month = {7},
+	pages = {76--82},
+	publisher = {Institute of Electrical},
+	title = {NetCDF: an interface for scientific data access},
+	url = {http://dx.doi.org/10.1109/38.56302},
+	volume = {10},
+}
+
+
+@misc{https://doi.org/10.5281/zenodo.14274886,
[…]
