+++ _preview/565/build/_shared/architecture-U656AL7Q-WLB6T2II.js
@@ -1,1 +1,0 @@
-import{a as e,b as r}from"/_preview/565/build/_shared/chunk-FYSFXZKR.js";import"/_preview/565/build/_shared/chunk-K5DBCL5N.js";import"/_preview/565/build/_shared/chunk-W25R6CBQ.js";import"/_preview/565/build/_shared/chunk-GLKM22BQ.js";import"/_preview/565/build/_shared/chunk-UVP6FRQW.js";import"/_preview/565/build/_shared/chunk-OZE3FFNP.js";export{e as ArchitectureModule,r as createArchitectureServices};
