+++ _preview/565/build/_shared/infoDiagram-F6ZHWCRC-MUQ5PJQF.js
@@ -1,2 +1,0 @@
-import{a as n}from"/_preview/565/build/_shared/chunk-EC4C7QZ2.js";import{a as s}from"/_preview/565/build/_shared/chunk-UJXF6HD2.js";import{a as p}from"/_preview/565/build/_shared/chunk-EJAB6C5B.js";import"/_preview/565/build/_shared/chunk-7ZGJO7S4.js";import"/_preview/565/build/_shared/chunk-TQOV233I.js";import"/_preview/565/build/_shared/chunk-WRUPT7W2.js";import"/_preview/565/build/_shared/chunk-FYSFXZKR.js";import"/_preview/565/build/_shared/chunk-WJCEDSXQ.js";import"/_preview/565/build/_shared/chunk-ZJA2U5VY.js";import"/_preview/565/build/_shared/chunk-AMWGDE74.js";import"/_preview/565/build/_shared/chunk-K5DBCL5N.js";import"/_preview/565/build/_shared/chunk-W25R6CBQ.js";import"/_preview/565/build/_shared/chunk-GLKM22BQ.js";import{L as i}from"/_preview/565/build/_shared/chunk-TRJHAMCF.js";import"/_preview/565/build/_shared/chunk-UVP6FRQW.js";import{b as r,d as a}from"/_preview/565/build/_shared/chunk-ZHBLEF4E.js";import"/_preview/565/build/_shared/chunk-MBMQ6BQC.js";import"/_preview/565/build/_shared/chunk-OZE3FFNP.js";var m={parse:r(async e=>{let t=await p("info",e);a.debug(t)},"parse")},v={version:n.version+""},d=r(()=>v.version,"getVersion"),f={getVersion:d},c=r((e,t,g)=>{a.debug(`rendering info diagram
-`+e);let o=s(t);i(o,100,400,!0),o.append("g").append("text").attr("x",100).attr("y",40).attr("class","version").attr("font-size",32).style("text-anchor","middle").text(`v${g}`)},"draw"),u={draw:c},S={parser:m,db:f,renderer:u};export{S as diagram};

[…]
