### 2025-12-08 819935d
change citation format
