+++ _preview/585/build/_shared/chunk-AHSZ7MKH.js
@@ -1,15 +1,0 @@
-import{b as e}from"/_preview/585/build/_shared/chunk-7DW7L33J.js";var l=e(()=>`
-  /* Font Awesome icon styling - consolidated */
-  .label-icon {
-    display: inline-block;
-    height: 1em;
-    overflow: visible;
-    vertical-align: -0.125em;
-  }
-  
-  .node .label-icon path {
-    fill: currentColor;
-    stroke: revert;
-    stroke-width: revert;
-  }
-`,"getIconStyles");export{l as a};

[…]
