### 2025-12-08 d68ee21
Address build warnings and fix links (#564)
* update link to netcdf logo

* add configuration option to minimize warnings about matplotlib output

* fix bibliography config

* fix and/or update a bunch of references

* fix links

* link to md pages
