+++ _preview/582/build/_shared/chunk-X25FUMBP.js
@@ -1,0 +1,1 @@
+import{b as i}from"/_preview/582/build/_shared/chunk-7DW7L33J.js";var o=i(({flowchart:t})=>{let r=t?.subGraphTitleMargin?.top??0,a=t?.subGraphTitleMargin?.bottom??0,e=r+a;return{subGraphTitleTopMargin:r,subGraphTitleBottomMargin:a,subGraphTitleTotalMargin:e}},"getSubGraphTitleMargins");export{o as a};
