+++ _preview/586/build/_shared/ichigojam-SIH6OLN4.js
@@ -1,0 +1,1 @@
+import{a}from"/_preview/586/build/_shared/chunk-6I7R2J3E.js";import"/_preview/586/build/_shared/chunk-OZE3FFNP.js";export default a();
