### 2025-12-09 ProjectPythia/pythia-foundations#587 @brian-rose
JOSE Review: cite FAIR 4 Software paper
It was mentioned in the review thread for the JOSE paper [here](https://github.com/openjournals/jose-reviews/issues/308#issuecomment-3502535042) that it would be appropriate to cite the "FAIR Principles for Research Software" paper <https://doi.org/10.1038/s41597-022-01710-x> in the paper.
