+++ _preview/588/build/_shared/erb-WVVQ56SN.js
@@ -1,0 +1,1 @@
+import{c as u}from"/_preview/588/build/_shared/chunk-OZE3FFNP.js";var a=u((g,e)=>{function r(n){return{name:"ERB",subLanguage:"xml",contains:[n.COMMENT("<%#","%>"),{begin:"<%[%=-]?",end:"[%-]?%>",subLanguage:"ruby",excludeBegin:!0,excludeEnd:!0}]}}e.exports=r});export default a();
