+++ preamble/how-to-use.md
@@ -46,7 +46,7 @@ of the Scientific Python Ecosystem.
 
 All of the content in Pythia Foundations is authored in Markdown
 and presented in the form of a [Jupyter
-Book](https://jupyterbook.org)[@https://doi.org/10.25080/hwcj9957] with
+Book](https://jupyterbook.org) [@https://doi.org/10.25080/hwcj9957] with
 “chapters” written as MyST Markdown documents or Jupyter Notebooks
 [@https://doi.org/10.3233/978-1-61499-649-1-87]. The power
 of Jupyter Notebooks is that they can contain both static text and
