### 2025-12-10 fbdc69d
fix typo
Co-authored-by: Brian Rose <brose@albany.edu>
