### 2025-12-10 e3d15d3
formatting
Co-authored-by: Brian Rose <brose@albany.edu>
