### 2025-12-10 803e588
update BinderHub definition
Co-authored-by: Brian Rose <brose@albany.edu>
