+++ appendix/glossary.md
@@ -8,7 +8,7 @@ description: Glossary of terms used in Pythia Foundations.
 [Binder](https://jupyter.org/binder)
 : An open-source service that allows users to create sharable, interactive computing environments from {term}`Jupyter Notebooks` and other repositories. Binder can reproduce a computational environment directly from a GitHub repository, providing a seamless way to share and interact with code and data.
 
-: The public service to run Binder is on <https://mybinder.org>, which is running {term}`BinderHub`. ProjectPythia
+: The public service to run Binder is on <https://mybinder.org>, which is running {term}`BinderHub`. The Binder links on most Project Pythia pages point to <https://binder.projectpythia.org>, which is a Pythia-specific implementation of {term}`BinderHub` running on the National Science Foundation-supported [Jetstream2](https://jetstream-cloud.org/) cloud computing service.
 
 BinderHub
 : The underlying technology and infrastructure that powers {term}`Binder`. BinderHub deploys and manages the interactive computing environments for {term}`Jupyter Notebooks`, ensuring that users can access and share reproducible computational work.
