### 2025-12-10 cd43d7e
Update myst.yml
Co-authored-by: Brian Rose <brose@albany.edu>
