+++ _preview/582/build/github-issues-7fe1a28b5ba6a739062998f8fa4f758a.md
@@ -1,0 +1,100 @@
+# Issues and Discussions
+
+```{image} ../../images/GitHub-logo.png
+:alt: GitHub Logo
+:width: 400px
+```
+
+## Overview:
+
+1. What are Issues and Discussions?
+1. Examine an existing Issue
+1. Examine an existing Discussion
+
+## Prerequisites
+
+| Concepts                               | Importance | Notes |
+| -------------------------------------- | ---------- | ----- |
+| [What is GitHub?](what-is-github.md)   | Necessary  |       |
+| [GitHub Repositories](github-repos.md) | Necessary  |       |
+
+- **Time to learn**: 5 minutes
+
+---
+
+## What are Issues and Discussions?
+
+GitHub provides two different, but related mechanisms for communicating
+within a repository about a project: _Issues_ and _Discussions_.
+Issues are more like “to-do” items; they are task-focused. For example, Issues
+are often used to report and track bugs, request new features, or
+perhaps note a performance problem. Ultimately, the maintainers of
+a project may resolve the issue by fixing the bug, adding the
+feature, etc., and then closing the resolved issue, marking the
+task as completed. GitHub _Discussions_, much like the name implies,
+are more open ended, and may not have a resolution. Asking about a
+topic, discussing the merits of a new feature, or even advertising
+an event, such as a tutorial for your project, are all examples
+of _Discussions_.
+
+In the text below we discuss _Issues_ in more detail, followed by
+a discussion on Discussions. Keep in mind that when initiating a
+conversation on GitHub, it is often unclear whether something is
+more suited as an _Issue_ or a _Discussion_. We, the creators of
+Project Pythia, struggle with this ourselves. If you’re not sure, simply pick
+one. Fortunately, the GitHub developers recognized this dilemma, and
+made it easy to convert _Issues_ into _Discussions_ and vice versa.
+
+## Issues
+
+To get started, let's take a look at the [Issues page](https://github.com/ProjectPythia/pythia-foundations/issues) in Project Pythia's `pythia-foundations` repository:
+
+<img src="../../images/GitHubPythiaIssues.png" alt="Pythia Issues">
+
+By default, it shows all open Issues, but we can see all [closed Issues](https://github.com/ProjectPythia/pythia-foundations/issues?q=is%3Aissue+is%3Aclosed) by clicking "Closed".
+
+<img src="../../images/GitHubPythiaIssuesClosed.png" alt="Pythia Closed Issues">
+
+Issues, Discussions, and Pull Requests are all numbered for easy reference. By opening, resolving, and then closing an issue, we are leaving behind a searchable public record of what the issue was, why we thought it was important, and how we resolved it. This is great for project management, since it gets old Issues out of the way without actually deleting them.
+
+Let's now examine [Issue \#144](https://github.com/ProjectPythia/pythia-foundations/issues/144).
+
+<img src="../../images/GitHubPythiaIssue144.png" alt="Pythia Issue 144">
+
+As you can see, some broken links were found in one of the Pythia Foundations tutorials, likely because the site being linked recently had its structure changed. An additional comment was added, as well as a label to help filtering/sorting Issues by topic. We then see that this issue was mentioned (by typing the issue number) elsewhere in the repository. In this case, it was mentioned in [Pull Request \#145](https://github.com/ProjectPythia/pythia-foundations/pull/145), which makes the changes to fix the issue. We can also see that the PR has been merged, which means the changes have been incorporated into the main branch of the code.
+
[…]
