### 2025-12-10 84d141e
capitalize all abbreviations
Co-authored-by: Brian Rose <brose@albany.edu>
