### 2025-12-10 211b25b
Use MyST wiki links
