+++ _preview/589/build/_shared/chunk-3TGUKLBR.js
@@ -1,0 +1,1 @@
+import{a as n}from"/_preview/589/build/_shared/chunk-CBDDV5MN.js";import{c as a}from"/_preview/589/build/_shared/chunk-OZE3FFNP.js";var c=a((i,o)=>{var r=n();o.exports=t;t.displayName="objectivec";t.aliases=["objc"];function t(e){e.register(r),e.languages.objectivec=e.languages.extend("c",{string:{pattern:/@?"(?:\\(?:\r\n|[\s\S])|[^"\\\r\n])*"/,greedy:!0},keyword:/\b(?:asm|auto|break|case|char|const|continue|default|do|double|else|enum|extern|float|for|goto|if|in|inline|int|long|register|return|self|short|signed|sizeof|static|struct|super|switch|typedef|typeof|union|unsigned|void|volatile|while)\b|(?:@interface|@end|@implementation|@protocol|@class|@public|@protected|@private|@property|@try|@catch|@finally|@throw|@synthesize|@dynamic|@selector)\b/,operator:/-[->]?|\+\+?|!=?|<<?=?|>>?=?|==?|&&?|\|\|?|[~^%?*\/@]/}),delete e.languages.objectivec["class-name"],e.languages.objc=e.languages.objectivec}});export{c as a};
