+++ _preview/589/build/_shared/dust-OBHDZTHF.js
@@ -1,0 +1,1 @@
+import{c as s}from"/_preview/589/build/_shared/chunk-OZE3FFNP.js";var l=s((i,e)=>{function n(a){let t="if eq ne lt lte gt gte select default math sep";return{name:"Dust",aliases:["dst"],case_insensitive:!0,subLanguage:"xml",contains:[{className:"template-tag",begin:/\{[#\/]/,end:/\}/,illegal:/;/,contains:[{className:"name",begin:/[a-zA-Z\.-]+/,starts:{endsWithParent:!0,relevance:0,contains:[a.QUOTE_STRING_MODE]}}]},{className:"template-variable",begin:/\{/,end:/\}/,illegal:/;/,keywords:t}]}}e.exports=n});export default l();

[…]
