### 2025-12-10 91b2219
Add additional citations (#565)
* add additional citations

* add references to BibTeX file to avoid lookup and fix bibliography config

* change citation format

* typo

Co-authored-by: Brian Rose <brose@albany.edu>

* update section title

---------

Co-authored-by: Brian Rose <brose@albany.edu>
