### 2025-12-10 a8a5fa0
Add central glossary, additional abbreviations, links, and minor editorial changes (#582)
* add abbreviations and central glossary along with some minor editorial changes to why-python.md

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix typo

Co-authored-by: Brian Rose <brose@albany.edu>

* formatting

Co-authored-by: Brian Rose <brose@albany.edu>

* update BinderHub definition

Co-authored-by: Brian Rose <brose@albany.edu>

* Update myst.yml

Co-authored-by: Brian Rose <brose@albany.edu>

* capitalize all abbreviations

Co-authore
[…]
