+++ myst.yml
@@ -17,17 +17,28 @@ project:
   settings:
     output_matplotlib_strings: remove
   abbreviations:
-    FOSS: Free and open-source software
-    VCS: Version control system
-    FTP: File transfer protocol
+    API: Application Programming Interface
+    DOI: Digital Object Identifier
+    FOSS: Free and Open-Source Software
+    FTP: File Transfer Protocol
+    HTTP: Hypertext Transfer Protocol
+    ID: Identifier
+    IDL: Interactive Data Language
+    NCAR: NSF National Center for Atmospheric Research
+    NCL: NCAR Command Language
+    NSF: U.S. National Science Foundation
+    PR: Pull Request
+    SPE: Scientific Python Ecosystem
+    SSH: Secure Shell Protocol
     TCP: Transmission Control Protocol
+    VCS: Version Control System
   toc:
     - file: landing-page.md
     - title: Preamble
       children:
         - file: preamble/how-to-use.md
         - file: preamble/how-to-cite.md
-    - title: Foundational skills
+    - title: Foundational Skills
       children:
         - file: foundations/overview.md
         - file: foundations/why-python.md
@@ -56,7 +67,7 @@ project:
             - file: foundations/github/review-pr.md
             - file: foundations/github/github-workflows.md
             - file: foundations/github/contribute-to-pythia.md
-    - title: Core Scientific Python packages
+    - title: Core Scientific Python Packages
       children:
         - file: core/overview.md
         - file: core/numpy.md
@@ -89,6 +100,7 @@ project:
             - file: core/xarray/dask-arrays-xarray.ipynb
     - title: Appendix
       children:
+        - file: appendix/glossary.md
         - file: appendix/how-to-contribute.md
           children:
             - file: appendix/template.ipynb
