+++ build/github-setup-advance-1e3673fb36c66d5d2546cc535eeffba7.md
@@ -1,193 +1,0 @@
-# Configuring Your GitHub Account
-
-```{image} ../../images/GitHub-logo.png
-:alt: GitHub Logo
-:width: 400px
-```
-
-## Overview:
-
-1. Configure your GitHub account for secure logins via ssh and/or https
-1. Set up notifications on repositories you own or follow
-
-## Prerequisites
-
-| Concepts                                                      | Importance  | Notes                        |
-| ------------------------------------------------------------- | ----------- | ---------------------------- |
-| [What is GitHub?](what-is-github.md)                          | Necessary   | GitHub user account required |
-| [GitHub Repositories](github-repos.md)                        | Necessary   |                              |
-| [Issues and Discussions](github-issues.md)                    | Recommended |                              |
-| [Cloning and Forking a Repository](github-cloning-forking.md) | Recommended |                              |
-
-- **Time to learn**: 35 minutes
-
----
-
-## GitHub secure key generation
-
-When you signed up for your free account on [GitHub](https://github.com), you established a _user ID_ and its corresponding _password_. Many of the repositories that GitHub serves are readable from anywhere, not even requiring a GitHub account.
-
-However, especially when you use the git command-line interface to access a GitHub-hosted repo, there are cases when you need to provide an additional set of login credentials. Some of these cases are:
-
-1. When you want to clone a _private_, as opposed to _public_ GitHub repository (**read-access**)
-2. When you wish to _push_ to a repo (**write-access**)
-
-For these use-cases, you won't be able to simply type your GitHub user ID and password from the command line. Instead, you need to set up _access tokens_ that live in two places: in your GitHub account, and in your local computer's file system.
-
-GitHub supports two means of key-based access: via _https_, and via _ssh_.
-
-For example, one can clone [Project Pythia's Sandbox repository](https://github.com/ProjectPythia/github-sandbox) using a URL for the https protocol:
-
-```{image} ../../images/GitHub_Setup_Advanced_https_URL.png
-:alt: GitHub Clone https
-:width: 600px
-```
-
----
-
-The URL in this case is **https://github.com/ProjectPythia/github-sandbox.git**
-
-Similarly, if you click on the **SSH** tab:
-
-```{image} ../../images/GitHub_Setup_Advanced_ssh_URL.png
-:alt: GitHub Clone ssh
-:width: 600px
-```
-
----
-
-Here, the URL is **git@github.com:ProjectPythia/github-sandbox.git**
-
-## Generate a secure personal access token for https
-
-First, you will create a secure token in your GitHub account settings, and then use the token on your local computer.
-
-Follow the steps with helpful screenshots at [GitHub's PAT Creation](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token) page.
-
-```{admonition} Tip:
-:class: tip
-If using the *https* protocol to *push* to a remote repo, you must have generated and downloaded a personal access token. You *may* also need it when cloning, *if* the remote repo is *not* open to all.
-```
-
-## Generate an SSH public/private keypair
-
-First, on your local computer, you will create an SSH _public/private keypair_, and then upload the _public key_ to your GitHub account.
-
-Follow the steps with helpful screenshots at [GitHub's Connecting to GitHub with SSH](https://docs.github.com/en/github/authenticating-to-github/connecting-to-github-with-ssh) page.
-
-```{admonition} Tip:
-:class: tip
-If using the *ssh* protocol to clone *or* push, you *must* have generated and created an ssh key-pair.
-```
-
----
-
[…]
