+++ build/how-to-use-78398efd1546d2289b0b4fdd59da1815.md
@@ -1,0 +1,141 @@
+# How to Use This Book
+
+## Overview
+
+Pythia Foundations is a geoscience-flavored introduction to essential tools in the Scientific Python Ecosystem (SPE) and [Pangeo](https://pangeo.io) stack. It covers the foundational knowledge that's needed to get started with Python in the computational geosciences, as well as to become an effective citizen-practitioner in key open geoscience software ecosystems. The intended audience is anyone from undergraduate students through established geoscientists who are relatively new to working in Python. The tutorials in this book also serve as references and prerequisites for the more advanced and domain-specific content in the [Pythia Cookbook Gallery](https://cookbooks.projectpythia.org).
+
+## What's included
+
+The Scientific Python Ecosystem (SPE) is a collection of
+open source Python packages that support analysis, manipulation,
+and visualization of scientific data. While Project Pythia is focused
+on the geoscience communities, the material contained in Pythia
+Foundations applies broadly, and may be useful to scientists, or
+aspiring scientists, working in any number of disciplines. Pythia
+Foundations makes very few assumptions about the experience level
+of the reader other than having a background in math or science,
+and being comfortable using a computer, including the command line
+terminal (i.e., the [Unix shell](https://en.wikipedia.org/wiki/Unix_shell)).
+Prior programming experience is not required.
+
+Lastly, in addition to the Python language and a number of fundamental
+scientific Python packages, Pythia Foundations covers two topics
+that we believe are essential
+to effectively using the SPE: GitHub, for sharing workflows and
+managing code; and Jupyter Notebooks, the technology by which all
+Pythia Foundations material is presented and which is quickly
+becoming a standard format for scientific communication of computational
+results.
+
+## Organization of Pythia Foundations
+
+Pythia Foundations is organized into two main sections as seen in
+the sidebar on the left and the schematic below: Foundational Skills, and Core Scientific
+Python Packages. The first, Foundational Skills, covers essential
+material that all users of Project Pythia are expected to feel
+comfortable with in order to make the most of the rest of the Project
+Pythia content. The second, Core Scientific Python Packages, covers
+what we believe are the most important and fundamental packages
+in the Scientific Python Ecosystem. These packages serve as the
+building blocks for many of the more geoscience focused components
+of the Scientific Python Ecosystem.
+
+![Foundations Infographic](../images/ProjectPythia_Foundations_Infographic_v4.jpg)
+
+## Running Pythia Foundations examples
+
+All of the content in Pythia Foundations is authored in Markdown
+and presented in the form of a [Jupyter
+Book](https://jupyterbook.org) [@https://doi.org/10.25080/hwcj9957] with
+“chapters” written as MyST Markdown documents or Jupyter Notebooks
+[@https://doi.org/10.3233/978-1-61499-649-1-87]. The power
+of Jupyter Notebooks is that they can contain both static text and
+executable code that you can interact with. When you navigate to a
+book chapter such as [Matplotlib
+Basics](../core/matplotlib/matplotlib-basics)
+you will see static text, Python code, and the rendered output of
+that code in the form of the many figures that appear. In the case
+of Matplotlib Basics these figures are produced by Matplotlib itself.
+Viewing content in this manner is not much different than reading
+a hardbound textbook. To get the full benefit of Jupyter Books you
+can run, and even modify, the example code in real time! This
+interaction allows you to experiment with different parameters and
+observe instantly how results change.
+
+There are two ways that you can execute the Pythia Foundations book
+chapters. Both are described below.
+
+![Annotated Pythia Foundations home page](../images/foundations_diagram.png)
+
+### Interacting with Jupyter Notebooks in the cloud via Binder
+
+The simplest way to interact with a Jupyter Notebook is through
[…]
