### 2025-12-10 85cf97f
Add more links (#589)
* Quickstart: add links to C/Fortan and np/plt tutorials

* Use MyST wiki links
