+++ _preview/589/build/_shared/chunk-PDV5BDST.js
@@ -1,1 +1,0 @@
-import{a as o,b as n,c as a,d as s,e as m,f as e,g as u,j as d,o as c,q as l}from"/_preview/589/build/_shared/chunk-ILONG3JM.js";var v=class extends l{static{e(this,"PieTokenBuilder")}constructor(){super(["pie","showData"])}},C=class extends c{static{e(this,"PieValueConverter")}runCustomConverter(t,r,i){if(t.name==="PIE_SECTION_LABEL")return r.replace(/"/g,"").trim()}},P={parser:{TokenBuilder:e(()=>new v,"TokenBuilder"),ValueConverter:e(()=>new C,"ValueConverter")}};function p(t=s){let r=a(n(t),u),i=a(o({shared:r}),d,P);return r.ServiceRegistry.register(i),{shared:r,Pie:i}}e(p,"createPieServices");export{P as a,p as b};
