### 2025-12-12 23f22a6
add additional abbreviations
