### 2025-12-12 89e34c9
add additional abbreviations
