+++ myst.yml
@@ -20,6 +20,7 @@ project:
     output_matplotlib_strings: remove
   abbreviations:
     API: Application Programming Interface
+    CVS: Concurrent Versions System
     DOI: Digital Object Identifier
     FOSS: Free and Open-Source Software
     FTP: File Transfer Protocol
@@ -32,7 +33,9 @@ project:
     PR: Pull Request
     SPE: Scientific Python Ecosystem
     SSH: Secure Shell Protocol
+    SVN: Apache Subversion
     TCP: Transmission Control Protocol
+    URL: Uniform Resource Locator
     VCS: Version Control System
 
   error_rules:
