### 2026-01-13 #308 @acocac
@brian-rose - Happy New Year! I appreciate if you have any updates on addressing relevant changes of the submitted resource as suggested by the reviewers.
