+++ _preview/594/build/_shared/dist-DAMRF7E6.js
@@ -1,0 +1,1 @@
+import{a as o,b as n}from"/_preview/594/build/_shared/chunk-6BVT5IQV.js";import"/_preview/594/build/_shared/chunk-CUKUDK3R.js";import"/_preview/594/build/_shared/chunk-3EBOCCHJ.js";import{d as t}from"/_preview/594/build/_shared/chunk-4OEDG4JQ.js";import"/_preview/594/build/_shared/chunk-GUCIBHGO.js";import"/_preview/594/build/_shared/chunk-OZE3FFNP.js";function m(i){let e=(i??"1 2 2 3").split(/\s/).map(r=>Number(r.trim())).filter(r=>!Number.isNaN(r)).map(r=>Math.min(Math.max(Math.floor(r),1),12));return e.length===0||e.length>4?[1,2,2,3]:e}var d={name:"grid",arg:{type:String},options:{...o("grid")},body:{type:"myst",required:!0},run(i){let e={type:"grid",columns:m(i.arg),children:i.body};return n(i,e),[e]}},p={name:"grid-item",options:{label:{type:String,alias:["name"]},class:{type:String}},body:{type:"myst",required:!0},run(i){var e,r;let{label:s,identifier:l}=t((e=i.options)===null||e===void 0?void 0:e.label)||{};return[{type:"div",children:i.body,class:(r=i.options)===null||r===void 0?void 0:r.class,label:s,identifier:l}]}},g=[d,p];export{d as gridDirective,g as gridDirectives,p as gridItemDirective};
