+++ _preview/594/build/_shared/chunk-CMMKXMGV.js
@@ -1,7 +1,0 @@
-import{a as T,b as V,c as pt}from"/_preview/594/build/_shared/chunk-PLXGGK6F.js";import{a as kt,d as X}from"/_preview/594/build/_shared/chunk-V73H6EQZ.js";import{a as ft}from"/_preview/594/build/_shared/chunk-X25FUMBP.js";import{c as ht,d as mt}from"/_preview/594/build/_shared/chunk-44Y5F2RG.js";import{g as yt}from"/_preview/594/build/_shared/chunk-IFGNTSJO.js";import{m as $}from"/_preview/594/build/_shared/chunk-H5UGLSR3.js";import{F as K,W as v}from"/_preview/594/build/_shared/chunk-JU57NILB.js";import{D as et,G as nt,K as st,L as ot,M as it,N as lt,O as ct,P as dt,b as h,d as m,j as D,u as J,v as tt,x as rt,y as at,z as U}from"/_preview/594/build/_shared/chunk-7DW7L33J.js";var vt=h((r,t,a,s,i,n)=>{t.arrowTypeStart&&ut(r,"start",t.arrowTypeStart,a,s,i,n),t.arrowTypeEnd&&ut(r,"end",t.arrowTypeEnd,a,s,i,n)},"addEdgeMarkers"),$t={arrow_cross:{type:"cross",fill:!1},arrow_point:{type:"point",fill:!0},arrow_barb:{type:"barb",fill:!0},arrow_circle:{type:"circle",fill:!1},aggregation:{type:"aggregation",fill:!1},extension:{type:"extension",fill:!1},composition:{type:"composition",fill:!0},dependency:{type:"dependency",fill:!0},lollipop:{type:"lollipop",fill:!1},only_one:{type:"onlyOne",fill:!1},zero_or_one:{type:"zeroOrOne",fill:!1},one_or_more:{type:"oneOrMore",fill:!1},zero_or_more:{type:"zeroOrMore",fill:!1},requirement_arrow:{type:"requirement_arrow",fill:!1},requirement_contains:{type:"requirement_contains",fill:!1}},ut=h((r,t,a,s,i,n,e)=>{let o=$t[a];if(!o){m.warn(`Unknown arrow type: ${a}`);return}let c=o.type,d=`${i}_${n}-${c}${t==="start"?"Start":"End"}`;if(e&&e.trim()!==""){let f=e.replace(/[^\dA-Za-z]/g,"_"),p=`${d}_${f}`;if(!document.getElementById(p)){let y=document.getElementById(d);if(y){let L=y.cloneNode(!0);L.id=p,L.querySelectorAll("path, circle, line").forEach(k=>{k.setAttribute("stroke",e),o.fill&&k.setAttribute("fill",e)}),y.parentNode?.appendChild(L)}}r.attr(`marker-${t}`,`url(${s}#${p})`)}else r.attr(`marker-${t}`,`url(${s}#${d})`)},"addEdgeMarker"),P=new Map,x=new Map,nr=h(()=>{P.clear(),x.clear()},"clear"),I=h(r=>r?r.reduce((a,s)=>a+";"+s,""):"","getLabelStyles"),sr=h(async(r,t)=>{let a=K(v().flowchart.htmlLabels),{labelStyles:s}=mt(t);t.labelStyle=s;let i=await yt(r,t.label,{style:t.labelStyle,useHtmlLabels:a,addSvgBackground:!0,isNode:!1});m.info("abc82",t,t.labelType);let n=r.insert("g").attr("class","edgeLabel"),e=n.insert("g").attr("class","label").attr("data-id",t.id);e.node().appendChild(i);let o=i.getBBox();if(a){let l=i.
[…]
