### 2026-01-20 ProjectPythia/pythia-foundations#590 @brian-rose
This was discussed at today's Pythia Community meeting. The consensus was that we are well justified in sticking to an opinionated presentation of our recommended best practices for environment management, and we will continue to promote the use of mini forge and the Conda-forge channel.

We will update the text in Foundations to address:
- why we recommend against using pip (prevalence of dependencies on compiled libraries in geoscience workflows)
- existence of newer tools like pixi, with external links but no explicit recommendation
- Focus on use of the free, community-supported Conda-forge stack
- Mention existence of Anaconda as a commercially supported option for those that are interested

We will include instructions on the landing page for each book section for how to set up a minimal environment for that section alone, as part of our overall goal of teaching readers about package and environment management (but not require readers to do this). At the most granular notebook-by-notebook level, we will emphasize that `import` statements at the top of each notebook should give enough information for minimal environment descriptions.
