### 2026-01-20 1bf6b5d
Add min version for mystmd
