+++ environment.yml
@@ -18,5 +18,5 @@ dependencies:
   - xarray
   - python-graphviz
   - graphviz
-  - mystmd
+  - mystmd>=1.6.0
   - jupyterlab-myst
