+++ _preview/596/build/_shared/chunk-2RRX75WA.js
@@ -1,1 +1,0 @@
-import{b as i}from"/_preview/596/build/_shared/chunk-7DW7L33J.js";function t(c,e){c.accDescr&&e.setAccDescription?.(c.accDescr),c.accTitle&&e.setAccTitle?.(c.accTitle),c.title&&e.setDiagramTitle?.(c.title)}i(t,"populateCommonDb");export{t as a};
