+++ _preview/596/build/_shared/inform7-O7PU5KJH.js
@@ -1,1 +1,0 @@
-import{c as o}from"/_preview/596/build/_shared/chunk-OZE3FFNP.js";var r=o((t,s)=>{function a(i){let e="\\[",n="\\]";return{name:"Inform 7",aliases:["i7"],case_insensitive:!0,keywords:{keyword:"thing room person man woman animal container supporter backdrop door scenery open closed locked inside gender is are say understand kind of rule"},contains:[{className:"string",begin:'"',end:'"',relevance:0,contains:[{className:"subst",begin:e,end:n}]},{className:"section",begin:/^(Volume|Book|Part|Chapter|Section|Table)\b/,end:"$"},{begin:/^(Check|Carry out|Report|Instead of|To|Rule|When|Before|After)\b/,end:":",contains:[{begin:"\\(This",end:"\\)"}]},{className:"comment",begin:e,end:n,contains:["self"]}]}}s.exports=a});export default r();
