+++ _preview/596/build/_shared/ttcn-cfg-EUK6SGDO.js
@@ -1,1 +1,0 @@
-import{b as A}from"/_preview/596/build/_shared/chunk-OZE3FFNP.js";function I(e){for(var n={},T=e.split(" "),E=0;E<T.length;++E)n[T[E]]=!0;return n}function L(e,n){var T=e.next();if(T=='"'||T=="'")return n.tokenize=D(T),n.tokenize(e,n);if(/[:=]/.test(T))return t=T,"punctuation";if(T=="#")return e.skipToEnd(),"comment";if(/\d/.test(T))return e.eatWhile(/[\w\.]/),"number";if(O.test(T))return e.eatWhile(O),"operator";if(T=="[")return e.eatWhile(/[\w_\]]/),"number";e.eatWhile(/[\w\$_]/);var E=e.current();return U.propertyIsEnumerable(E)?"keyword":R.propertyIsEnumerable(E)?"atom":u.propertyIsEnumerable(E)?"deleted":"variable"}function D(e){return function(n,T){for(var E=!1,i,l=!1;(i=n.next())!=null;){if(i==e&&!E){var C=n.peek();C&&(C=C.toLowerCase(),(C=="b"||C=="h"||C=="o")&&n.next()),l=!0;break}E=!E&&i=="\\"}return(l||!(E||S))&&(T.tokenize=null),"string"}}function _(e,n,T,E,i){this.indented=e,this.column=n,this.type=T,this.align=E,this.prev=i}function r(e,n,T){var E=e.indented;return e.context&&e.context.type=="statement"&&(E=e.context.indented),e.context=new _(E,n,T,null,e.context)}function o(e){var n=e.context.type;return(n==")"||n=="]"||n=="}")&&(e.indented=e.context.indented),e.context=e.context.prev}var N,U,R,u,S,P,O,t,c,M=A(()=>{N={name:"ttcn-cfg",keywords:I("Yes No LogFile FileMask ConsoleMask AppendFile TimeStampFormat LogEventTypes SourceInfoFormat LogEntityName LogSourceInfo DiskFullAction LogFileNumber LogFileSize MatchingHints Detailed Compact SubCategories Stack Single None Seconds DateTime Time Stop Error Retry Delete TCPPort KillTimer NumHCs UnixSocketsEnabled LocalAddress"),fileNCtrlMaskOptions:I("TTCN_EXECUTOR TTCN_ERROR TTCN_WARNING TTCN_PORTEVENT TTCN_TIMEROP TTCN_VERDICTOP TTCN_DEFAULTOP TTCN_TESTCASE TTCN_ACTION TTCN_USER TTCN_FUNCTION TTCN_STATISTICS TTCN_PARALLEL TTCN_MATCHING TTCN_DEBUG EXECUTOR ERROR WARNING PORTEVENT TIMEROP VERDICTOP DEFAULTOP TESTCASE ACTION USER FUNCTION STATISTICS PARALLEL MATCHING DEBUG LOG_ALL LOG_NOTHING ACTION_UNQUALIFIED DEBUG_ENCDEC DEBUG_TESTPORT DEBUG_UNQUALIFIED DEFAULTOP_ACTIVATE DEFAULTOP_DEACTIVATE DEFAULTOP_EXIT DEFAULTOP_UNQUALIFIED ERROR_UNQUALIFIED EXECUTOR_COMPONENT EXECUTOR_CONFIGDATA EXECUTOR_EXTCOMMAND EXECUTOR_LOGOPTIONS EXECUTOR_RUNTIME EXECUTOR_UNQUALIFIED FUNCTION_RND FUNCTION_UNQUALIFIED MATCHING_DONE MATCHING_MCSUCCES
[…]
