### 2026-01-23 #308 @brian-rose
> [@brian-rose](https://github.com/brian-rose) - Happy New Year! I appreciate if you have any updates on addressing relevant changes of the submitted resource as suggested by the reviewers.

Hi @acocac, we are making progress! The work definitely got delayed over the holidays. We are working our way through all the issues raised by the reviewers and also addressing some necessary maintenance of the Pythia Foundations infrastructure.
