+++ build/getting-started-pyth-f4548dc0c66c6b1cebc06f6ba3c8a6bf.md
@@ -1,8 +1,0 @@
-# Getting Started with Python
-
-**_New Python users, start here!_**
-
-## Topics
-
-- [Quickstart: Zero to Python](quickstart): For the impatient among us: run your first Python code in the cloud!
-- [Installing and Running Python](how-to-run-python): Detailed instructions for choosing a Python platform and getting up and running on a laptop, including using the conda package manager.
