+++ build/routes/_index-ZMAZD7U6.js
@@ -1,1 +1,0 @@
-import{a as m,d as u}from"/build/_shared/chunk-ECLX7DIY.js";import{a as w,ya as f}from"/build/_shared/chunk-7UUHRSK3.js";import"/build/_shared/chunk-C4DFGG5C.js";import"/build/_shared/chunk-J7TUH54J.js";import"/build/_shared/chunk-FZ2S7OYD.js";import"/build/_shared/chunk-JEM6JXYA.js";import"/build/_shared/chunk-34XIY2DH.js";import"/build/_shared/chunk-KQM5FBHR.js";import"/build/_shared/chunk-OCWQY3HK.js";import"/build/_shared/chunk-7HNKBP4B.js";import"/build/_shared/chunk-CUKUDK3R.js";import"/build/_shared/chunk-3EBOCCHJ.js";import"/build/_shared/chunk-O4VQNZ62.js";import"/build/_shared/chunk-4OEDG4JQ.js";import"/build/_shared/chunk-GUCIBHGO.js";import"/build/_shared/chunk-AQ2CODAG.js";import"/build/_shared/chunk-JJXTQVMA.js";import{f as g}from"/build/_shared/chunk-OZE3FFNP.js";var d=g(w());var j=({data:o,location:l})=>{var r,i,n,s,a,p,c;if(!o)return[];let t=o.config,e=o.project;return f({origin:"",url:l.pathname,title:(r=t==null?void 0:t.title)!=null?r:e.title,description:(n=(i=t.description)!=null?i:e.description)!=null?n:void 0,image:(s=e.thumbnailOptimized||e.thumbnail)!=null?s:void 0,keywords:(p=(a=t.keywords)!=null?a:e.keywords)!=null?p:[],twitter:(c=t==null?void 0:t.options)==null?void 0:c.twitter})},x=()=>[m];var M=u;export{M as default,x as links,j as meta};
