+++ build/manifest-3595279D.js
@@ -1,1 +1,0 @@
-window.__remixManifest={"entry":{"module":"/build/entry.client-PCJPW7TK.js","imports":["/build/_shared/chunk-AQ2CODAG.js","/build/_shared/chunk-JJXTQVMA.js","/build/_shared/chunk-OZE3FFNP.js"]},"routes":{"root":{"id":"root","path":"","module":"/build/root-SIO6LUTY.js","imports":["/build/_shared/chunk-CH4FVTDV.js","/build/_shared/chunk-C4DFGG5C.js","/build/_shared/chunk-J7TUH54J.js","/build/_shared/chunk-FZ2S7OYD.js","/build/_shared/chunk-JEM6JXYA.js","/build/_shared/chunk-34XIY2DH.js","/build/_shared/chunk-KQM5FBHR.js","/build/_shared/chunk-OCWQY3HK.js","/build/_shared/chunk-7HNKBP4B.js","/build/_shared/chunk-CUKUDK3R.js","/build/_shared/chunk-3EBOCCHJ.js","/build/_shared/chunk-O4VQNZ62.js","/build/_shared/chunk-4OEDG4JQ.js","/build/_shared/chunk-GUCIBHGO.js"],"hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":true},"routes/$":{"id":"routes/$","parentId":"root","path":"*","module":"/build/routes/$-PRP77N34.js","imports":["/build/_shared/chunk-FAHZZXAC.js"],"hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":true},"routes/($project)_.($a).($b).($c).($d).$slug[.json]":{"id":"routes/($project)_.($a).($b).($c).($d).$slug[.json]","parentId":"root","path":":project?/:a?/:b?/:c?/:d?/:slug.json","module":"/build/routes/($project)_.($a).($b).($c).($d).$slug[.json]-DIB2VQ4M.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[favicon.ico]":{"id":"routes/[favicon.ico]","parentId":"root","path":"favicon.ico","module":"/build/routes/[favicon.ico]-MH4XVVZI.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[objects.inv]":{"id":"routes/[objects.inv]","parentId":"root","path":"objects.inv","module":"/build/routes/[objects.inv]-RUZS76RX.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[robots.txt]":{"id":"routes/[robots.txt]","parentId":"root","path":"robots.txt","module":"/build/routes/[robots.txt]-JULN6D72.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[sitemap.xml]":{"id":"routes/[sitemap.xml]","parentId":"root","path":"sitemap.xml","module":"/build/routes/[sitemap.xml]-NGYNA2IS.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/[sitemap_style.xsl]":{"id":"routes/[sitemap_style.xsl]","parentId":"root","path":"sitemap_style.xsl","module":"/build/routes/[sitemap_style.xsl]-GJUCKAG6.js","hasAction":false,"hasLoader":true,"hasCatchBoundary":false,"hasErrorBoundary":false},"routes/_index":{"id":"routes/_index","parentId":"root","index":true,"module":"/build/routes/_index-QUJ6PDLY.js
[…]
