### 2026-02-16 0748996
Fix Zenodo badge on landing page
