### 2026-02-16 1adc1f0
Add outline to NumPy overview
