### 2026-02-17 ProjectPythia/pythia-foundations#590 @r-ford
Note that a PR closing this issue also closes #568
