### 2026-02-17 21ae38f
Update core/numpy.md
Co-authored-by: Brian Rose <brose@albany.edu>
