### 2026-02-17 1f2f57f
Update core/numpy.md
Co-authored-by: Brian Rose <brose@albany.edu>
