+++ _preview/602/build/_shared/infoDiagram-WHAUD3N6-LGW7VANE.js
@@ -1,2 +1,0 @@
-import{a as n}from"/_preview/602/build/_shared/chunk-6ZWZTF4Z.js";import{a as s}from"/_preview/602/build/_shared/chunk-PSIFSBYU.js";import{a as p}from"/_preview/602/build/_shared/chunk-GS7QPNKU.js";import"/_preview/602/build/_shared/chunk-D4I4FMU3.js";import"/_preview/602/build/_shared/chunk-LHWWMMIN.js";import"/_preview/602/build/_shared/chunk-PDV5BDST.js";import"/_preview/602/build/_shared/chunk-B7PDIXGE.js";import"/_preview/602/build/_shared/chunk-536FPRQA.js";import"/_preview/602/build/_shared/chunk-MB4EJF2W.js";import"/_preview/602/build/_shared/chunk-5GONE7L7.js";import"/_preview/602/build/_shared/chunk-ILONG3JM.js";import"/_preview/602/build/_shared/chunk-JMV5HLZN.js";import{L as i}from"/_preview/602/build/_shared/chunk-JU57NILB.js";import{b as r,d as a}from"/_preview/602/build/_shared/chunk-7DW7L33J.js";import"/_preview/602/build/_shared/chunk-FZ2S7OYD.js";import"/_preview/602/build/_shared/chunk-JEM6JXYA.js";import"/_preview/602/build/_shared/chunk-34XIY2DH.js";import"/_preview/602/build/_shared/chunk-KQM5FBHR.js";import"/_preview/602/build/_shared/chunk-OZE3FFNP.js";var m={parse:r(async e=>{let t=await p("info",e);a.debug(t)},"parse")},v={version:n.version+""},d=r(()=>v.version,"getVersion"),f={getVersion:d},c=r((e,t,g)=>{a.debug(`rendering info diagram
-`+e);let o=s(t);i(o,100,400,!0),o.append("g").append("text").attr("x",100).attr("y",40).attr("class","version").attr("font-size",32).style("text-anchor","middle").text(`v${g}`)},"draw"),u={draw:c},S={parser:m,db:f,renderer:u};export{S as diagram};
