+++ _preview/602/build/_shared/mizar-PMIAEKSD.js
@@ -1,1 +1,0 @@
-import{c as i}from"/_preview/602/build/_shared/chunk-OZE3FFNP.js";var n=i((r,e)=>{function t(s){return{name:"Mizar",keywords:"environ vocabularies notations constructors definitions registrations theorems schemes requirements begin end definition registration cluster existence pred func defpred deffunc theorem proof let take assume then thus hence ex for st holds consider reconsider such that and in provided of as from be being by means equals implies iff redefine define now not or attr is mode suppose per cases set thesis contradiction scheme reserve struct correctness compatibility coherence symmetry assymetry reflexivity irreflexivity connectedness uniqueness commutativity idempotence involutiveness projectivity",contains:[s.COMMENT("::","$")]}}e.exports=t});export default n();
