### 2026-02-17 230670a
Replace old HTML admonitions with proper MyST syntax throughout the book (#601)
* Update call to numpy.trapezoid and fix admonitions

* Replace all remaining HTML admonitions with MyST syntax

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
