+++ build/numpy-f1626ba224ee80e7e8222c64ba1d126b.md
@@ -14,3 +14,9 @@ From the [NumPy documentation](https://numpy.org/doc/stable/user/whatisnumpy.htm
 > NumPy is the fundamental package for scientific computing in Python. It is a Python library that provides a multidimensional array object, various derived objects (such as masked arrays and matrices), and an assortment of routines for fast operations on arrays, including mathematical, logical, shape manipulation, sorting, selecting, I/O, discrete Fourier transforms, basic linear algebra, basic statistical operations, random simulation, and much more.
 
 NumPy's position at the center of the scientific Python ecosystem means that all users should start here in their learning journey through the core scientific packages.
+
+In [NumPy Basics](numpy-basics), you will get an introduction to NumPy arrays, including how to create and perform basic operations on them, as well as the basics of indexing and slicing arrays.
+
+[Intermediate NumPy](intermediate-numpy) covers additional concepts related to indexing and slicing multidimensional arrays,
+and [NumPy Broadcasting](numpy-broadcasting) introduces the way NumPy treats arithmetic on arrays of different
+shapes, which can help make your code more efficient by avoiding loops.
