+++ _preview/603/build/_shared/chunk-JRVCZGCT.js
@@ -1,1 +1,0 @@
-import{a as p}from"/_preview/603/build/_shared/chunk-QQNNMZV3.js";import{a as n}from"/_preview/603/build/_shared/chunk-XK3K4AT7.js";import{c as l}from"/_preview/603/build/_shared/chunk-OZE3FFNP.js";var o=l((x,s)=>{var u=p(),c=n();s.exports=r;r.displayName="tsx";r.aliases=[];function r(a){a.register(u),a.register(c),function(e){var g=e.util.clone(e.languages.typescript);e.languages.tsx=e.languages.extend("jsx",g),delete e.languages.tsx.parameter,delete e.languages.tsx["literal-property"];var t=e.languages.tsx.tag;t.pattern=RegExp(/(^|[^\w$]|(?=<\/))/.source+"(?:"+t.pattern.source+")",t.pattern.flags),t.lookbehind=!0}(a)}});export{o as a};
