+++ _preview/603/build/_shared/ldif-VPUYKXAU.js
@@ -1,1 +1,0 @@
-import{c as t}from"/_preview/603/build/_shared/chunk-OZE3FFNP.js";var r=t((s,e)=>{function a(n){return{name:"LDIF",contains:[{className:"attribute",begin:"^dn",end:": ",excludeEnd:!0,starts:{end:"$",relevance:0},relevance:10},{className:"attribute",begin:"^\\w",end:": ",excludeEnd:!0,starts:{end:"$",relevance:0}},{className:"literal",begin:"^-",end:"$"},n.HASH_COMMENT_MODE]}}e.exports=a});export default r();
