+++ _preview/591/build/_shared/chunk-FQLWQKN7.js
@@ -1,1 +1,0 @@
-import{a as o}from"/_preview/591/build/_shared/chunk-R3ABELDT.js";import{c as a}from"/_preview/591/build/_shared/chunk-OZE3FFNP.js";var u=a((r,s)=>{var t=o();s.exports=e;e.displayName="jsonp";e.aliases=[];function e(n){n.register(t),n.languages.jsonp=n.languages.extend("json",{punctuation:/[{}[\]();,.]/}),n.languages.insertBefore("jsonp","punctuation",{function:/(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*\()/})}});export{u as a};
