+++ _preview/591/build/_shared/chunk-SZAK5OLM.js
@@ -1,1 +1,0 @@
-import{a as g}from"/_preview/591/build/_shared/chunk-SWM73M5L.js";import{a as s}from"/_preview/591/build/_shared/chunk-37MEAX4M.js";import{c as t}from"/_preview/591/build/_shared/chunk-OZE3FFNP.js";var n=t((u,r)=>{var c=g(),l=s();r.exports=a;a.displayName="t4Cs";a.aliases=[];function a(e){e.register(c),e.register(l),e.languages.t4=e.languages["t4-cs"]=e.languages["t4-templating"].createT4("csharp")}});export{n as a};
