### 2026-02-18 cb21998
Use myst class formatting for badges
