### 2026-02-18 747983c
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
