### 2026-02-18 81f32f1
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
