### 2026-02-18 6cad0a7
Use simple grid, not subfigures
