### 2026-02-18 4e5f58a
Remove badges and let myst handle DOI
