+++ landing-page.md
@@ -2,12 +2,6 @@
 
 ## A community learning resource for Python-based computing in the geosciences
 
-::::{grid} 2 2 2 2
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7884571.svg)](https://doi.org/10.5281/zenodo.7884571)
-[![status](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15/status.svg)](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15)
-::::
-
-
 :::{image} images/pretty-earth.png
 :alt: Pretty Earth
 :width: 600px
