### 2026-02-19 fafcf85
Fix Zenodo badge on landing page (#598)
* Fix Zenodo badge on landing page

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Try to fix line formatting

* Use myst class formatting for badges

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* tweak line break

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Use simple grid, not subfigures

* Remove badges and let myst handle DOI

---------

Co-authored-by: pre-commit-ci[bo
[…]
