+++ _preview/598/build/_shared/dist-SXUICQHC.js
@@ -1,1 +1,0 @@
-import{Da as N,G as R,Ka as I,Na as Oe,Pa as C,Qa as Y,Ra as j,Sa as oe,W as ee,fa as z,ha as u,i as E,ia as te,na as A,oa as P,qa as k,x as J,xa as G}from"/_preview/598/build/_shared/chunk-6FEZWRIS.js";import{b as _}from"/_preview/598/build/_shared/chunk-OZE3FFNP.js";function Se(e){return e==45||e==46||e==58||e>=65&&e<=90||e==95||e>=97&&e<=122||e>=161}function ge(e){return e==9||e==10||e==13||e==32}function h(e,t){let o=e.pos+t;if(Z==e&&B==o)return U;for(;ge(e.peek(t));)t++;let O="";for(;;){let r=e.peek(t);if(!Se(r))break;O+=String.fromCharCode(r),t++}return Z=e,B=o,U=O||null}function D(e,t){this.name=e,this.parent=t}function W(e,t){return new C(o=>{let O=0,r=t.charCodeAt(0);e:for(;!(o.next<0);o.advance(),O++)if(o.next==r){for(let a=1;a<t.length;a++)if(o.peek(a)!=t.charCodeAt(a))continue e;break}O&&o.acceptToken(e)})}var b,re,ne,ae,le,se,ie,ce,me,$e,U,Z,B,pe,ue,fe,de,qe,xe,M,F=_(()=>{oe();te();b=1,re=2,ne=3,ae=4,le=5,se=36,ie=37,ce=38,me=11,$e=13;U=null,Z=null,B=0;pe=new Y({start:null,shift(e,t,o,O){return t==b?new D(h(O,1)||"",e):e},reduce(e,t){return t==me&&e?e.parent:e},reuse(e,t,o,O){let r=t.type.id;return r==b||r==$e?new D(h(O,1)||"",e):e},strict:!1}),ue=new C((e,t)=>{if(e.next==60){if(e.advance(),e.next==47){e.advance();let o=h(e,0);if(!o)return e.acceptToken(le);if(t.context&&o==t.context.name)return e.acceptToken(re);for(let O=t.context;O;O=O.parent)if(O.name==o)return e.acceptToken(ne,-2);e.acceptToken(ae)}else if(e.next!=33&&e.next!=63)return e.acceptToken(b)}},{contextual:!0});fe=W(se,"-->"),de=W(ie,"?>"),qe=W(ce,"]]>"),xe=z({Text:u.content,"StartTag StartCloseTag EndTag SelfCloseEndTag":u.angleBracket,TagName:u.tagName,"MismatchedCloseTag/TagName":[u.tagName,u.invalid],AttributeName:u.attributeName,AttributeValue:u.attributeValue,Is:u.definitionOperator,"EntityReference CharacterReference":u.character,Comment:u.blockComment,ProcessingInst:u.processingInstruction,DoctypeDecl:u.documentMeta,Cdata:u.spec
[…]
