### 2026-02-21 3b51c74
updates to Git and GitHub content
