+++ foundations/github/github-workflows.md
@@ -1,20 +1,21 @@
-# GitHub Workflows
+# Git Workflows
 
 ```{image} ../../images/Git-Logo-2Color.png
 :alt: Git Logo
 :width: 400px
 ```
 
-A workflow is a series of activities or tasks that must be completed sequentially or parallel to achieve the desired outcome. Here we outline two different GitHub workflows that take you through the steps leading up to opening a Pull Request.
+A workflow is a series of activities or tasks that must be completed sequentially or parallel to achieve the desired outcome. Here we outline two different Git workflows that take you through the steps leading up to opening a pull request.
 
 ## Overview:
 
-1. GitHub workflows overview
-1. Git Feature Branch Workflow
+1. Overview of Git workflows
+1. Feature branch workflow
 1. Forking workflow
 
 ## Prerequisites
 
+<<<<<<< Updated upstream
 | Concepts                                         | Importance  | Notes |
 | ------------------------------------------------ | ----------- | ----- |
 | [What is GitHub](what-is-github.md)              | Necessary   |       |
@@ -25,14 +26,26 @@ A workflow is a series of activities or tasks that must be completed sequentiall
 | [Branches](git-branches.md)                      | Necessary   |       |
 | [Pull Requests](github-pull-request.md)          | Necessary   |       |
 | [Reviewing Pull Requests](review-pr.md)          | Recommended |       |
+=======
+| Concepts                                      | Importance  | Notes |
+| --------------------------------------------- | ----------- | ----- |
+| [What is GitHub](what-is-github)              | Necessary   |       |
+| [GitHub Repositories](github-repos)           | Necessary   |       |
+| [Cloning and Forking](github-cloning-forking) | Necessary   |       |
+| [Basic Version Control with Git](basic-git)   | Necessary   |       |
+| [Issues and Discussions](github-issues)       | Recommended |       |
+| [Branches](git-branches)                      | Necessary   |       |
+| [Pull Requests](github-pull-request)          | Necessary   |       |
+| [Reviewing Pull Requests](review-pr)          | Recommended |       |
+>>>>>>> Stashed changes
 
 - **Time to learn**: 60 minutes
 
 ---
 
-## GitHub workflows
+## Git workflows
 
-GitHub, together with Git, are powerful tools for managing and
+GitHub, together with Git, is a powerful tool for managing and
 collaborating on all kinds of digital assets, such as software,
 documentation, and even manuscripts for research papers. Like other
 complex software environments, often these tools can be employed
@@ -451,8 +464,8 @@ git merge upstream/main
 ## Summary
 
 - The steps that lead up to
-  the PR depend your GitHub Workflow.
-- Two commonly used GitHub Worflows are **Git Feature Branch Workflow** and
+  the PR depend your chosen workflow.
+- Two commonly used worflows include the **Branch Workflow** and the
   **Forking Workflow**. The former is appropriate for teams of collaborators
   where everyone has write access to the GitHub repository. The latter
   is commonly used when a developer wishes to contribute to a public GitHub
@@ -463,7 +476,7 @@ git merge upstream/main
 In the next lesson we will put the **Forking Workflow** to work and show you
 how to use it to [contribute to Project Pythia](contribute-to-pythia).
 
-## References
+## Additional Resources
 
-1. Atlassian's tutorial on [workflows](https://www.atlassian.com/git/tutorials/comparing-workflows)
-1. GitHub's [Collaborating with Pull Requests](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests)
+- Atlassian's tutorial on [workflows](https://www.atlassian.com/git/tutorials/comparing-workflows)
+- GitHub's [Collaborating with Pull Requests](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests)
