+++ foundations/github/github-pull-request.md
@@ -148,7 +148,11 @@ The steps leading up to creating your PR depend on your workflow. The main diffe
 the contributor now, for the Feature Workflow, navigates to the upstream, remote
 repository, not a personal remote fork, and initiates the PR there.
 
+<<<<<<< Updated upstream
 We will cover [GitHub Workflows](github-workflows.md) in greater detail in the next chapter.
+=======
+We will cover [Git workflows](git-workflows.md) in greater detail in the next chapter.
+>>>>>>> Stashed changes
 
 ---
 
