+++ foundations/github/basic-git.md
@@ -1,15 +1,15 @@
-# Basic Version Control with _git_
+# Basic Version Control with Git
 
 ```{image} ../../images/Git-Logo-2Color.png
 :alt: Git Logo
 :width: 400px
 ```
 
-## Overview:
+## Overview
 
 1. The need for version control
-1. Basic git usage
-1. Making your first git commit
+1. Basic Git usage
+1. Making your first Git commit
 1. Viewing and comparing across the commit history
 
 ## Prerequisites
@@ -26,11 +26,11 @@
 
 ---
 
-## About version control and git
+## About version control and Git
 
 ### What is version control (and why should we care)?
 
-[Version Control](https://en.wikipedia.org/wiki/Version_control) refers generally to systems for managing changes to documents or files. Version control systems let us keep track of what changes were made to a file, when they were made, and by whom. If you've ever used "Tracked changes" on a Word document with multiple authors, then you've seen a form of version control in action (though NOT one that is well suited to working with computer code!).
+[Version control](https://en.wikipedia.org/wiki/Version_control) refers generally to systems for managing changes to documents or files. Version control systems let us keep track of what changes were made to a file, when they were made, and by whom. If you've ever used "Track Changes" in a Word document with multiple authors, then you've seen a form of version control in action (though NOT one that is well suited to working with computer code!).
 
 The need for version control is particularly acute when _working with computer code_, where small changes to the text can have huge impacts on the results of running the code.
 
@@ -44,21 +44,21 @@ analysis_script_09122021_edit.py
 analysis_script_NEW.py
 ```
 
-A Version Control System (VCS) like git will replace this mess with a _well-ordered and labelled history_ of edits that you can freely browse through, and will greatly simplify collaborating with other people on writing new code.
+A version control system (VCS) like Git will replace this mess with a _well-ordered and labelled history_ of edits that you can freely browse through, and will greatly simplify collaborating with other people on writing new code.
 
-### What is git?
+### What is Git?
 
 #### Git is not GitHub
 
-That's the first thing to understand. GitHub is a web-based platform for hosting code and collaborating with other people. On the other hand, **git is a command-line Version Control System (VCS)** that you can download and install. It runs on your local computer as well as under the hood on GitHub. You need to understand something about version control with git in order to use many of GitHub's collaboration features.
+That's the first thing to understand. GitHub is a web-based platform for hosting code and collaborating with other people. On the other hand, **Git is a command-line version control system (VCS)** that you can download and install. It runs on your local computer as well as under the hood on GitHub. You need to understand something about version control with Git in order to use many of GitHub's collaboration features.
 
 #### A little history and nomenclature
 
-Git has been around [since the mid-2000s](https://en.wikipedia.org/wiki/Git). It was originally written by Linus Torvalds specifically for use in development of the Linux kernel. Git is [FOSS](https://foundations.projectpythia.org/foundations/github/what-is-github#free-and-open-source-software-foss) and comes pre-installed on many Linux and Mac OS systems.
+Git has been around [since the mid-2000s](https://en.wikipedia.org/wiki/Git). It was originally written by Linus Torvalds specifically for use in development of the Linux kernel. Git is [FOSS](https://foundations.projectpythia.org/foundations/github/what-is-github#free-and-open-source-software-foss) and comes pre-installed on many Linux and macOS systems.
 
[…]

[…]
