+++ foundations/github/github-setup-advanced.md
@@ -7,7 +7,7 @@
 
 ## Overview:
 
-1. Configure your GitHub account for secure logins via ssh and/or https
+1. Configure your GitHub account for secure logins via SSH and/or HTTPS
 1. Set up notifications on repositories you own or follow
 
 ## Prerequisites
@@ -27,19 +27,19 @@
 
 When you signed up for your free account on [GitHub](https://github.com), you established a _user ID_ and its corresponding _password_. Many of the repositories that GitHub serves are readable from anywhere, not even requiring a GitHub account.
 
-However, especially when you use the git command-line interface to access a GitHub-hosted repo, there are cases when you need to provide an additional set of login credentials. Some of these cases are:
+However, especially when you use the Git command-line interface to access a GitHub-hosted repo, there are cases when you need to provide an additional set of login credentials. Some of these cases are:
 
 1. When you want to clone a _private_, as opposed to _public_ GitHub repository (**read-access**)
 2. When you wish to _push_ to a repo (**write-access**)
 
 For these use-cases, you won't be able to simply type your GitHub user ID and password from the command line. Instead, you need to set up _access tokens_ that live in two places: in your GitHub account, and in your local computer's file system.
 
-GitHub supports two means of key-based access: via _https_, and via _ssh_.
+GitHub supports two means of key-based access: via _HTTPS_, and via _SSH_.
 
-For example, one can clone [Project Pythia's Sandbox repository](https://github.com/ProjectPythia/github-sandbox) using a URL for the https protocol:
+For example, one can clone [Project Pythia's Sandbox repository](https://github.com/ProjectPythia/github-sandbox) using a URL for the HTTPS protocol:
 
 ```{image} ../../images/GitHub_Setup_Advanced_https_URL.png
-:alt: GitHub Clone https
+:alt: GitHub clone HTTPS
 :width: 600px
 ```
 
@@ -50,7 +50,7 @@ The URL in this case is **https://github.com/ProjectPythia/github-sandbox.git**
 Similarly, if you click on the **SSH** tab:
 
 ```{image} ../../images/GitHub_Setup_Advanced_ssh_URL.png
-:alt: GitHub Clone ssh
+:alt: GitHub clone SSH
 :width: 600px
 ```
 
@@ -58,7 +58,7 @@ Similarly, if you click on the **SSH** tab:
 
 Here, the URL is **git@github.com:ProjectPythia/github-sandbox.git**
 
-## Generate a secure personal access token for https
+## Generate a secure personal access token for HTTPS
 
 First, you will create a secure token in your GitHub account settings, and then use the token on your local computer.
 
@@ -66,7 +66,7 @@ Follow the steps with helpful screenshots at [GitHub's PAT Creation](https://doc
 
 ```{admonition} Tip:
 :class: tip
-If using the *https* protocol to *push* to a remote repo, you must have generated and downloaded a personal access token. You *may* also need it when cloning, *if* the remote repo is *not* open to all.
+If using the *HTTPS* protocol to *push* to a remote repo, you must have generated and downloaded a personal access token. You *may* also need it when cloning, *if* the remote repo is *not* open to all.
 ```
 
 ## Generate an SSH public/private keypair
@@ -77,21 +77,21 @@ Follow the steps with helpful screenshots at [GitHub's Connecting to GitHub with
 
 ```{admonition} Tip:
 :class: tip
-If using the *ssh* protocol to clone *or* push, you *must* have generated and created an ssh key-pair.
+If using the *SSH* protocol to clone *or* push, you *must* have generated and created an SSH key-pair.
 ```
 
 ---
 
 ```{admonition} HTTPS vs SSH: Either is fine!
[…]
