### 2026-02-21 b7989ca
update Git install instructions to point to Carpentries instructions
