+++ _preview/607/build/_shared/q-HTTITPIC.js
@@ -1,0 +1,1 @@
+import{b as p}from"/_preview/607/build/_shared/chunk-OZE3FFNP.js";function s(n){return new RegExp("^("+n.join("|")+")$")}function r(n,e){var o=n.sol(),i=n.next();if(c=null,o){if(i=="/")return(e.tokenize=k)(n,e);if(i=="\\")return n.eol()||/\s/.test(n.peek())?(n.skipToEnd(),/^\\\s*$/.test(n.current())?(e.tokenize=m)(n):e.tokenize=r,"comment"):(e.tokenize=r,"builtin")}if(/\s/.test(i))return n.peek()=="/"?(n.skipToEnd(),"comment"):"null";if(i=='"')return(e.tokenize=a)(n,e);if(i=="`")return n.eatWhile(/[A-Za-z\d_:\/.]/),"macroName";if(i=="."&&/\d/.test(n.peek())||/\d/.test(i)){var t=null;return n.backUp(1),n.match(/^\d{4}\.\d{2}(m|\.\d{2}([DT](\d{2}(:\d{2}(:\d{2}(\.\d{1,9})?)?)?)?)?)/)||n.match(/^\d+D(\d{2}(:\d{2}(:\d{2}(\.\d{1,9})?)?)?)/)||n.match(/^\d{2}:\d{2}(:\d{2}(\.\d{1,9})?)?/)||n.match(/^\d+[ptuv]{1}/)?t="temporal":(n.match(/^0[NwW]{1}/)||n.match(/^0x[\da-fA-F]*/)||n.match(/^[01]+[b]{1}/)||n.match(/^\d+[chijn]{1}/)||n.match(/-?\d*(\.\d*)?(e[+\-]?\d+)?(e|f)?/))&&(t="number"),t&&(!(i=n.peek())||x.test(i))?t:(n.next(),"error")}return/[A-Za-z]|\./.test(i)?(n.eatWhile(/[A-Za-z._\d]/),f.test(n.current())?"keyword":"variable"):/[|/&^!+:\\\-*%$=~#;@><\.,?_\']/.test(i)||/[{}\(\[\]\)]/.test(i)?null:"error"}function k(n,e){return n.skipToEnd(),/^\/\s*$/.test(n.current())?(e.tokenize=v)(n,e):e.tokenize=r,"comment"}function v(n,e){var o=n.sol()&&n.peek()=="\\";return n.skipToEnd(),o&&/^\\\s*$/.test(n.current())&&(e.tokenize=r),"comment"}function m(n){return n.skipToEnd(),"comment"}function a(n,e){for(var o=!1,i,t=!1;i=n.next();){if(i=='"'&&!o){t=!0;break}o=!o&&i=="\\"}return t&&(e.tokenize=r),"string"}function l(n,e,o){n.context={prev:n.context,indent:n.indent,col:o,type:e}}function u(n){n.indent=n.context.indent,n.context=n.context.prev}var c,f,x,h,g=p(()=>{f=s(["abs","acos","aj","aj0","all","and","any","asc","asin","asof","atan","attr","avg","avgs","bin","by","ceiling","cols","cor","cos","count","cov","cross","csv","cut","delete","deltas","desc","dev","differ","distinct","div","do","each","ej","enlist","eval","except","exec","exit","exp","fby","fills","first","fkeys","flip","floor","from","get","getenv","group","gtime","hclose","hcount","hdel","hopen","hsym","iasc","idesc","if","ij","in","insert","inter","inv","key","keys","last","like","list","lj","load","log","lower","lsq","ltime","ltrim","mavg","max","maxs","mcount","md5","mdev","med","meta","min","mins","mmax","mmin","mmu","mod","msum","neg","next","not","null","or","over","parse",
[…]
