### 2026-02-21 9d5bd05
Use --strict for myst build
