+++ .github/workflows/nightly-build.yaml
@@ -11,6 +11,7 @@ jobs:
     uses: ProjectPythia/cookbook-actions/.github/workflows/build-book.yaml@main
     with:
       base_url: ''
+      build_command: 'myst build --execute --html --strict'
 
   deploy:
     needs: build
