### 2026-02-21 beff1ff
Use --strict for link checker only
