+++ .github/workflows/nightly-build.yaml
@@ -11,7 +11,6 @@ jobs:
     uses: ProjectPythia/cookbook-actions/.github/workflows/build-book.yaml@main
     with:
       base_url: ''
-      build_command: 'myst build --execute --html --strict'
 
   deploy:
     needs: build
@@ -20,3 +19,5 @@ jobs:
   link-check:
     if: ${{ github.repository_owner == 'ProjectPythia' }}
     uses: ProjectPythia/cookbook-actions/.github/workflows/link-checker.yaml@main
+    with:
+      build_command: 'myst build --check-links --strict'
