+++ _preview/609/build/_shared/php-template-3LFCUAPB.js
@@ -1,0 +1,1 @@
+import{c as i}from"/_preview/609/build/_shared/chunk-OZE3FFNP.js";var a=i((t,e)=>{function l(n){return{name:"PHP template",subLanguage:"xml",contains:[{begin:/<\?(php|=)?/,end:/\?>/,subLanguage:"php",contains:[{begin:"/\\*",end:"\\*/",skip:!0},{begin:'b"',end:'"',skip:!0},{begin:"b'",end:"'",skip:!0},n.inherit(n.APOS_STRING_MODE,{illegal:null,className:null,contains:null,skip:!0}),n.inherit(n.QUOTE_STRING_MODE,{illegal:null,className:null,contains:null,skip:!0})]}]}}e.exports=l});export default a();
