+++ _preview/610/build/_shared/timeline-definition-IT6M3QCI-PRBIVFFM.js
@@ -1,0 +1,61 @@
+import{M as ot,O as ct,P as ht,W as dt,c as st,d as at,e as lt}from"/_preview/610/build/_shared/chunk-JU57NILB.js";import{b as s,c as it,d as S,j as V,t as Q}from"/_preview/610/build/_shared/chunk-7DW7L33J.js";import"/_preview/610/build/_shared/chunk-KQM5FBHR.js";import"/_preview/610/build/_shared/chunk-OZE3FFNP.js";var X=function(){var n=s(function(m,i,a,c){for(a=a||{},c=m.length;c--;a[m[c]]=i);return a},"o"),t=[6,8,10,11,12,14,16,17,20,21],e=[1,9],l=[1,10],r=[1,11],d=[1,12],h=[1,13],g=[1,16],f=[1,17],p={trace:s(function(){},"trace"),yy:{},symbols_:{error:2,start:3,timeline:4,document:5,EOF:6,line:7,SPACE:8,statement:9,NEWLINE:10,title:11,acc_title:12,acc_title_value:13,acc_descr:14,acc_descr_value:15,acc_descr_multiline_value:16,section:17,period_statement:18,event_statement:19,period:20,event:21,$accept:0,$end:1},terminals_:{2:"error",4:"timeline",6:"EOF",8:"SPACE",10:"NEWLINE",11:"title",12:"acc_title",13:"acc_title_value",14:"acc_descr",15:"acc_descr_value",16:"acc_descr_multiline_value",17:"section",20:"period",21:"event"},productions_:[0,[3,3],[5,0],[5,2],[7,2],[7,1],[7,1],[7,1],[9,1],[9,2],[9,2],[9,1],[9,1],[9,1],[9,1],[18,1],[19,1]],performAction:s(function(i,a,c,u,y,o,E){var b=o.length-1;switch(y){case 1:return o[b-1];case 2:this.$=[];break;case 3:o[b-1].push(o[b]),this.$=o[b-1];break;case 4:case 5:this.$=o[b];break;case 6:case 7:this.$=[];break;case 8:u.getCommonDb().setDiagramTitle(o[b].substr(6)),this.$=o[b].substr(6);break;case 9:this.$=o[b].trim(),u.getCommonDb().setAccTitle(this.$);break;case 10:case 11:this.$=o[b].trim(),u.getCommonDb().setAccDescription(this.$);break;case 12:u.addSection(o[b].substr(8)),this.$=o[b].substr(8);break;case 15:u.addTask(o[b],0,""),this.$=o[b];break;case 16:u.addEvent(o[b].substr(2)),this.$=o[b];break}},"anonymous"),table:[{3:1,4:[1,2]},{1:[3]},n(t,[2,2],{5:3}),{6:[1,4],7:5,8:[1,6],9:7,10:[1,8],11:e,12:l,14:r,16:d,17:h,18:14,19:15,20:g,21:f},n(t,[2,7],{1:[2,1]}),n(t,[2,3]),{9:18,11:e,12:l,14:r,16:d,17:h,18:14,19:15,20:g,21:f},n(t,[2,5]),n(t,[2,6]),n(t,[2,8]),{13:[1,19]},{15:[1,20]},n(t,[2,11]),n(t,[2,12]),n(t,[2,13]),n(t,[2,14]),n(t,[2,15]),n(t,[2,16]),n(t,[2,4]),n(t,[2,9]),n(t,[2,10])],defaultActions:{},parseError:s(function(i,a){if(a.recoverable)this.trace(i);else{var c=new Error(i);throw c.hash=a,c}},"parseError"),parse:s(function(i){var a=this,c=[0],u=[],y=[null],o=
[…]
