+++ _preview/609/build/_shared/leaf-XTEKWXNH.js
@@ -1,1 +1,0 @@
-import{c as a}from"/_preview/609/build/_shared/chunk-OZE3FFNP.js";var s=a((t,e)=>{function n(i){return{name:"Leaf",contains:[{className:"function",begin:"#+[A-Za-z_0-9]*\\(",end:/ \{/,returnBegin:!0,excludeEnd:!0,contains:[{className:"keyword",begin:"#+"},{className:"title",begin:"[A-Za-z_][A-Za-z_0-9]*"},{className:"params",begin:"\\(",end:"\\)",endsParent:!0,contains:[{className:"string",begin:'"',end:'"'},{className:"variable",begin:"[A-Za-z_][A-Za-z_0-9]*"}]}]}]}}e.exports=n});export default s();
