+++ _preview/607/build/github-workflows-eabaac12134bc2b7f24f00f173188776.md
@@ -1,469 +1,0 @@
-# GitHub Workflows
-
-```{image} ../../images/Git-Logo-2Color.png
-:alt: Git Logo
-:width: 400px
-```
-
-A workflow is a series of activities or tasks that must be completed sequentially or parallel to achieve the desired outcome. Here we outline two different GitHub workflows that take you through the steps leading up to opening a Pull Request.
-
-## Overview:
-
-1. GitHub workflows overview
-1. Git Feature Branch Workflow
-1. Forking workflow
-
-## Prerequisites
-
-| Concepts                                         | Importance  | Notes |
-| ------------------------------------------------ | ----------- | ----- |
-| [What is GitHub](what-is-github.md)              | Necessary   |       |
-| [GitHub Repositories](github-repos.md)           | Necessary   |       |
-| [Cloning and Forking](github-cloning-forking.md) | Necessary   |       |
-| [Basic Version Control with _git_](basic-git.md) | Necessary   |       |
-| [Issues and Discussions](github-issues.md)       | Recommended |       |
-| [Branches](git-branches.md)                      | Necessary   |       |
-| [Pull Requests](github-pull-request.md)          | Necessary   |       |
-| [Reviewing Pull Requests](review-pr.md)          | Recommended |       |
-
-- **Time to learn**: 60 minutes
-
----
-
-## GitHub workflows
-
-GitHub, together with Git, are powerful tools for managing and
-collaborating on all kinds of digital assets, such as software,
-documentation, and even manuscripts for research papers. Like other
-complex software environments, often these tools can be employed
-in many different ways to accomplish the same goal. In order to
-effectively and consistently use Git and GitHub, over the years a
-variety of best practices have evolved for supporting different
-modes of collaboration. Collectively these different models, or
-recipes, are referred to as _workflows_.
-
-A typical sequence of workflow steps consists of the following:
-
-1. A contributor clones a personal remote repository, creating a local copy
-1. The contributor creates a new branch in their local repository
-1. The contributor makes changes to the branch and commits them to
-   their local repository
-1. The contributor _pushes_ the branch to a remote repository
-1. The contributor submits a PR via GitHub
-
-The sequence of steps
-outlined above provides a general framework for submitting a PR.
-But the precise set of steps is highly dependent on the choice of
-workflow for a given project. In this chapter we describe Pull
-Requests for two commonly used workflows: The **Git Feature Branch
-Workflow** and the **Forking Workflow**. The former is simpler and often
-used by teams when everyone on the team is an authorized contributor
-to the destination repository. I.e. all of the contributors have
-write access to the remote repository hosted by GitHub. The latter
-is typically what is needed to contribute to external projects for
-which the contributor is not authorized (i.e. does not have write
-access) to make changes to the destination repository. We briefly
-describe both workflows below, and include the steps necessary to
-make a PR on each.
-
-## Git Feature Branch Workflow
-
-The **Git Feature Branch Workflow** is one of the simplest and oldest
-collaborative workflows that is used for small team projects. The
-key idea behind this workflow, which is also common to the **Forking
-Workflow**, is that all development (all changes) should take place
-on a dedicated Git _feature_ branch, not the _main_ (historically
-referred to as _master_) branch. The motivation behind this is that
-one or more developers can iterate over a feature branch without
-disturbing the contents of the main branch. Consider using the **Git
-Feature Branch Workflow** for GitHub’s most widely used purpose,
-software development. Software modifications are liable to introduce
-bugs. Isolating them to a dedicated branch until they can be fixed
-ensures that a known, or official, version of the software is always
-available and in working order.
-
-```{note}
-Avoiding making edits directly on the `main` branch is considered best practice for most workflows and projects!
-```
-
[…]
