+++ myst.yml
@@ -8,10 +8,6 @@ project:
   bibliography:
     - references.bib
   authors:
-    - brian-rose
-    - clyne
-    - ktyle
-    - dcamron
     - Project Pythia Community
   doi: https://doi.org/10.5281/zenodo.7884571
   copyright: '2024'
