### 2026-02-24 7d61436
Full author list in how-to-cite.md with CITATION.cff export
