+++ _preview/612/build/_shared/chunk-5736WJYA.js
@@ -1,0 +1,1 @@
+import{a as g}from"/_preview/612/build/_shared/chunk-I232HDZP.js";import{c as t}from"/_preview/612/build/_shared/chunk-OZE3FFNP.js";var l=t((s,r)=>{var n=g();r.exports=a;a.displayName="racket";a.aliases=["rkt"];function a(e){e.register(n),e.languages.racket=e.languages.extend("scheme",{"lambda-parameter":{pattern:/([(\[]lambda\s+[(\[])[^()\[\]'\s]+/,lookbehind:!0}}),e.languages.insertBefore("racket","string",{lang:{pattern:/^#lang.+/m,greedy:!0,alias:"keyword"}}),e.languages.rkt=e.languages.racket}});export{l as a};
