### 2026-02-24 0cf8491
Improved authorship credits (#612)
* Messing around with author metadata

* Full authors on one page only

* Full author list in how-to-cite.md with CITATION.cff export

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
