+++ preamble/how-to-cite.md
@@ -1,10 +1,37 @@
+---
+authors:
+    # Author meta-data is defined upstream at https://github.com/ProjectPythia/pythia-config
+    - brian-rose
+    - jukent
+    - ktyle
+    - clyne
+    - andersy005
+    - dcamron
+    - dopplershift
+    - mgrover1
+    - kafitzgerald
+    - r-ford
+    - bl-freeman
+    - kmpaul
+    - erogluorhan
+    - lkailynncar
+    - anissa111
+    - jnmorley
+    - name: The Project Pythia Community
+exports:
+    - format: cff
+      title: Pythia Foundations
+---
+
 # How to Cite This Book
 
 The material in Pythia Foundations is licensed for free and open consumption and reuse. All code is served under [Apache 2.0](https://www.apache.org/licenses/LICENSE-2.0), while all non-code content is licensed under [Creative Commons BY 4.0 (CC BY 4.0)](https://creativecommons.org/licenses/by/4.0/). Effectively, this means you are free to share and adapt this material so long as you give appropriate credit to the Project Pythia community.
 
+A brief meta-paper describing Pythia Foundations is currently [under review](https://jose.theoj.org/papers/6ad4fe1df16342a33f750b43772c1b15) at the [Journal of Open Source Education](https://jose.theoj.org). Once a DOI has been issued for this paper, we will post it here as the simplest and recommended way to cite this work.
+
 The source code for the book is [released on GitHub](https://github.com/ProjectPythia/pythia-foundations) and archived on Zenodo. This DOI will always resolve to the latest version of the book source:
 <https://doi.org/10.5281/zenodo.7884571>
 
-If material in Pythia Foundations is useful in published work, you can cite a specific version of the book as:
+If you wish to cite a specific version of the book, you can use one of the [version-specific DOIs from Zenodo](https://zenodo.org/records/15742241), for example @10.5281/zenodo.15634505.
 
-> Rose, B. E. J., Kent, J., Tyle, K., Clyne, J., Banihirwe, A., Camron, D., May, R., Grover, M., Ford, R. R., Paul, K., Morley, J., Eroglu, O., Kailyn, L., & Zacharias, A. (2023). Pythia Foundations (Version v2023.05.01) [https://doi.org/10.5281/zenodo.7884572](https://doi.org/10.5281/zenodo.7884572)
+The full list of Foundations authors and their affiliations is also here on this page! The list grows over time as new individuals make substantial contributions to this ever-evolving work. [Join us!](../appendix/how-to-contribute)
