+++ myst.yml
@@ -7,8 +7,9 @@ project:
   github: https://github.com/projectpythia/pythia-foundations
   bibliography:
     - references.bib
-  author:
-    - Project Pythia Community
+  authors:
+    # The full author list is maintained in preamble/how-to-cite.md
+    - The Project Pythia Community
   doi: https://doi.org/10.5281/zenodo.7884571
   copyright: '2024'
   jupyter:
@@ -33,6 +34,7 @@ project:
     SSH: Secure Shell Protocol
     TCP: Transmission Control Protocol
     VCS: Version Control System
+
   toc:
     - file: landing-page.md
     - title: Preamble
