+++ _preview/612/build/_shared/chunk-5BDQQ53J.js
@@ -1,1 +1,0 @@
-import{a as n}from"/_preview/612/build/_shared/chunk-C2LAERNE.js";import{a as g}from"/_preview/612/build/_shared/chunk-SWM73M5L.js";import{c as a}from"/_preview/612/build/_shared/chunk-OZE3FFNP.js";var l=a((u,r)=>{var i=g(),s=n();r.exports=t;t.displayName="t4Vb";t.aliases=[];function t(e){e.register(i),e.register(s),e.languages["t4-vb"]=e.languages["t4-templating"].createT4("vbnet")}});export{l as a};
