+++ core/overview.md
@@ -11,9 +11,9 @@ As you might know by now that Python is a programming language. To make your job
 Most geoscience data analysis involves working with numerical arrays.
 The default library for dealing with numerical arrays in Python is [NumPy](http://www.numpy.org/).
 It has some built in functions for calculating very simple statistics
-(e.g. maximum, mean, standard deviation),
+(e.g., maximum, mean, standard deviation),
 but for more complex analysis
-(e.g. interpolation, integration, linear algebra)
+(e.g., interpolation, integration, linear algebra)
 the [SciPy](https://scipy.org) library is the default.
 If you’re dealing with particularly large arrays,
 [Dask](https://dask.org/) works with the existing Python ecosystem
@@ -31,8 +31,8 @@ is useful for dealing with non-standard calendars.
 When it comes to data visualization,
 the default library is [Matplotlib](https://matplotlib.org/).
 As you can see at the [Matplotlib gallery](https://matplotlib.org/stable/gallery/index.html),
-this library is great for any simple (e.g. bar charts, contour plots, line graphs),
-static (e.g. .png, .eps, .pdf) plots.
+this library is great for any simple (e.g., bar charts, contour plots, line graphs),
+static (e.g., .png, .eps, .pdf) plots.
 The [Cartopy](https://scitools.org.uk/cartopy/docs/latest/) library
 provides additional plotting functionality for common geographic map projections.
 
@@ -49,7 +49,10 @@ These high-levels libraries aren’t as flexible
 but they can do common tasks with far less effort.
 
 The most popular high-level data science library is undoubtedly [Pandas](http://pandas.pydata.org/).
-The key advance offered by Pandas is the concept of labeled arrays.
+The key advance offered by Pandas is the concept of the DataFrame, which is a labeled array:
+
+![Pandas DataFrame schematic](https://github.com/pandas-dev/pandas/blob/c8e8651a1b9d45eb45a3f6cf7fad4c0152cc84bb/doc/source/_static/schemas/01_table_dataframe.svg)
+
 Rather than referring to the individual elements of a data array using a numeric index
 (as is required with NumPy),
 the actual row and column headings can be used.
@@ -57,22 +60,27 @@ That means information from the cardiac ward on 3 July 2005
 could be obtained from a medical dataset by asking for `data['cardiac'].loc['2005-07-03']`,
 rather than having to remember the numeric index corresponding to that ward and date.
 This labeled array feature,
-combined with a bunch of other features that streamline common statistical and plotting tasks
+combined with a bunch of other features that streamline common statistical (e.g., computing averages) and plotting tasks
 traditionally performed with SciPy, datetime and Matplotlib,
 greatly simplifies the code development process (read: less lines of code).
 
 One of the limitations of Pandas
 is that it’s only able to handle one- or two-dimensional (i.e. tabular) data arrays.
 The [Xarray](http://xarray.pydata.org/) library was therefore created
-to extend the labelled array concept to x-dimensional arrays.
+to extend the labelled array concept to N-dimensional arrays:
+
+![Xarray Dataset schematic](https://github.com/pydata/xarray/blob/5f670a74392e3b625dba283c75c6dc2a43be808b/doc/_static/dataset-diagram.png)
+
 Not all of the Pandas functionality is available
 (which is a trade-off associated with being able to handle multi-dimensional arrays),
-but the ability to refer to array elements by their actual latitude (e.g. 20 South),
-longitude (e.g. 50 East), height (e.g. 500 hPa) and time (e.g. 2015-04-27), for example,
+but the ability to refer to array elements by their actual latitude (e.g., 20 South),
[…]
