### 2026-02-25 88ac4b2
Update core/overview.md
Co-authored-by: Katelyn FitzGerald <7872563+kafitzgerald@users.noreply.github.com>
