+++ _preview/614/build/_shared/dist-W5L4JBWE.js
@@ -1,1 +1,0 @@
-import{Aa as $,Da as Y,Ea as S,Na as R,Ra as o,Sa as p,fa as a,ha as O,ia as _,na as r,qa as e,xa as s,ya as X,za as l}from"/_preview/614/build/_shared/chunk-6FEZWRIS.js";import{b as i}from"/_preview/614/build/_shared/chunk-OZE3FFNP.js";var q,T,t,Z=i(()=>{p();_();q=a({null:O.null,instanceof:O.operatorKeyword,this:O.self,"new super assert open to with void":O.keyword,"class interface extends implements enum var":O.definitionKeyword,"module package import":O.moduleKeyword,"switch while for if else case default do break continue return try catch finally throw":O.controlKeyword,["requires exports opens uses provides public private protected static transitive abstract final strictfp synchronized native transient volatile throws"]:O.modifier,IntegerLiteral:O.integer,FloatingPointLiteral:O.float,"StringLiteral TextBlock":O.string,CharacterLiteral:O.character,LineComment:O.lineComment,BlockComment:O.blockComment,BooleanLiteral:O.bool,PrimitiveType:O.standard(O.typeName),TypeName:O.typeName,Identifier:O.variableName,"MethodName/Identifier":O.function(O.variableName),Definition:O.definition(O.variableName),ArithOp:O.arithmeticOperator,LogicOp:O.logicOperator,BitOp:O.bitwiseOperator,CompareOp:O.compareOperator,AssignOp:O.definitionOperator,UpdateOp:O.updateOperator,Asterisk:O.punctuation,Label:O.labelName,"( )":O.paren,"[ ]":O.squareBracket,"{ }":O.brace,".":O.derefOperator,", ;":O.separator}),T={__proto__:null,true:34,false:34,null:42,void:46,byte:48,short:48,int:48,long:48,char:48,float:48,double:48,boolean:48,
[…]
