### 2026-02-26 0dff938
Update Overview section (#613)
* Fix whitespace

* Update core/overview.md

Co-authored-by: Katelyn FitzGerald <7872563+kafitzgerald@users.noreply.github.com>

---------

Co-authored-by: Katelyn FitzGerald <7872563+kafitzgerald@users.noreply.github.com>
