+++ _preview/613/build/_shared/architecture-U656AL7Q-J6BQICAQ.js
@@ -1,1 +1,0 @@
-import{a as e,b as r}from"/_preview/613/build/_shared/chunk-B7PDIXGE.js";import"/_preview/613/build/_shared/chunk-ILONG3JM.js";import"/_preview/613/build/_shared/chunk-JMV5HLZN.js";import"/_preview/613/build/_shared/chunk-FZ2S7OYD.js";import"/_preview/613/build/_shared/chunk-JEM6JXYA.js";import"/_preview/613/build/_shared/chunk-34XIY2DH.js";import"/_preview/613/build/_shared/chunk-OZE3FFNP.js";export{e as ArchitectureModule,r as createArchitectureServices};
