+++ _preview/613/build/_shared/chunk-AUNDXEXT.js
@@ -1,1 +1,0 @@
-import{c as n}from"/_preview/613/build/_shared/chunk-OZE3FFNP.js";var o=n((a,t)=>{t.exports=r;r.displayName="prolog";r.aliases=[];function r(e){e.languages.prolog={comment:{pattern:/\/\*[\s\S]*?\*\/|%.*/,greedy:!0},string:{pattern:/(["'])(?:\1\1|\\(?:\r\n|[\s\S])|(?!\1)[^\\\r\n])*\1(?!\1)/,greedy:!0},builtin:/\b(?:fx|fy|xf[xy]?|yfx?)\b/,function:/\b[a-z]\w*(?:(?=\()|\/\d+)/,number:/\b\d+(?:\.\d*)?/,operator:/[:\\=><\-?*@\/;+^|!$.]+|\b(?:is|mod|not|xor)\b/,punctuation:/[(){}\[\],]/}}});export{o as a};

[…]
