+++ build/root-DSHAHCKW.js
@@ -1,0 +1,11 @@
+import{Aa as je,Ja as ke,L as u,a as Qe,d as Y,ea as be,f as ce,k as ge,ma as we,na as ue,q as he,ta as fe,ua as xe,xa as ye,za as ve}from"/build/_shared/chunk-XSZVU6CB.js";import"/build/_shared/chunk-C4DFGG5C.js";import"/build/_shared/chunk-J7TUH54J.js";import"/build/_shared/chunk-FZ2S7OYD.js";import"/build/_shared/chunk-JEM6JXYA.js";import"/build/_shared/chunk-34XIY2DH.js";import"/build/_shared/chunk-KQM5FBHR.js";import"/build/_shared/chunk-OCWQY3HK.js";import"/build/_shared/chunk-7HNKBP4B.js";import"/build/_shared/chunk-CUKUDK3R.js";import{a as me}from"/build/_shared/chunk-3EBOCCHJ.js";import"/build/_shared/chunk-O4VQNZ62.js";import{r as A,s as T,t as N}from"/build/_shared/chunk-4OEDG4JQ.js";import{a as pe}from"/build/_shared/chunk-GUCIBHGO.js";import{f as se,l as le}from"/build/_shared/chunk-AQ2CODAG.js";import{a as H,d as S}from"/build/_shared/chunk-JJXTQVMA.js";import{f as x}from"/build/_shared/chunk-OZE3FFNP.js";var ze="/build/_assets/app-BWFVNXVE.css";var Te="/build/_assets/thebe-core-VKVHG5VY.css";var xt=x(Qe());var tt="ENTRIES",Me="KEYS",Re="VALUES",y="",q=class{constructor(e,t){let r=e._tree,i=Array.from(r.keys());this.set=e,this._type=t,this._path=i.length>0?[{node:r,keys:i}]:[]}next(){let e=this.dive();return this.backtrack(),e}dive(){if(this._path.length===0)return{done:!0,value:void 0};let{node:e,keys:t}=P(this._path);if(P(t)===y)return{done:!1,value:this.result()};let r=e.get(P(t));return this._path.push({node:r,keys:Array.from(r.keys())}),this.dive()}backtrack(){if(this._path.length===0)return;let e=P(this._path).keys;e.pop(),!(e.length>0)&&(this._path.pop(),this.backtrack())}key(){return this.set._prefix+this._path.map(({keys:e})=>P(e)).filter(e=>e!==y).join("")}value(){return P(this._path).node.get(y)}result(){switch(this._type){case Re:return this.value();case Me:return this.key();default:return[this.key(),this.value()]}}[Symbol.iterator](){return this}},P=o=>o[o.length-1],rt=(o,e,t)=>{let r=new Map;if(e===void 0)return r;let i=e.length+1,a=i+t,n=new Uint8Array(a*i).fill(t+1);for(let d=0;d<i;++d)n[d]=d;for(let d=1;d<a;++d)n[d*i]=d;return Le(o,e,t,r,n,1,i,""),r},Le=(o,e,t,r,i,a,n,d)=>{let s=a*n;e:for(let l of o.keys())if(l===y){let p=i[s-1];p<=t&&r.set(d,[o.get(l),p])}else{let p=a;for(let g=0;g<l.length;++g,++p){let c=l[g],m=n*p,h=m-n,w=i[m],b=Math.max(0,p-t-1),j=Math.min(n-1,p+t);for(let k=b;k<j;++k){let $=c!==e[k],X=i[h+k]+ +$,W=i[h+k+1]+1,D=i[m+k]+1,V=i[m+k+1]=Math.min(X,W,D);V<w&&(w=V)}if(w>t)continue e}Le(o.get(l
[…]
