+++ build/what-is-github-49a133271da98e22053d185d3f201c93.md
@@ -1,0 +1,135 @@
+# What is GitHub?
+
+```{image} ../../images/GitHub-logo.png
+:alt: GitHub Logo
+:width: 400px
+```
+
+## Overview:
+
+1.  What is GitHub?
+1.  No experience necessary!
+1.  Free and open-source software (FOSS)
+1.  Version control systems (VCS)
+1.  GitHub = FOSS + VCS + Web
+1.  Register for a free GitHub account
+
+## Prerequisites
+
+| Concepts | Importance | Notes |
+| -------- | ---------- | ----- |
+| None     |            |       |
+
+- **Time to learn**: 15 minutes
+
+---
+
+## What is GitHub?
+
+[GitHub](https://github.com) is a web-based platform for the dissemination of free and open-source software.
+
+If you are reading this lesson, you are already using GitHub, as that is where Project Pythia hosts its content!
+
+GitHub provides the following:
+
+1. _Version control_ for free and open-source software and other digital assets
+1. Project _discussion forums_
+1. _DevOps_ to facilitate building and testing software
+1. _Bug_ reporting, patching, and tracking
+1. _Documentation_ hosting
+1. An environment that fosters _collaboration_
+
+Although GitHub can host any digital asset, the most common use case for GitHub is for individuals or organizations to house _repositories_ of _free_ and _open-source software_:
+
+## No experience necessary!
+
+You do not need to be an experienced software developer or be proficient in version control to make use of GitHub! Perhaps, though, you have used a particular package (e.g., Xarray or Matplotlib) and have had questions about its usage, noticed a bug, or had an idea for a new feature for the package! You can participate in a project's development via GitHub the same way you might have interacted with its developers via email in the past.
+
+## Free and open-source software (FOSS)
+
+Much of what we term the _scientific Python software ecosystem_ consists of _free and open-source software_. Often abbreviated as **FOSS**, this means:
+
+1.  The software is free of charge, and
+1.  The various files which contain the _software code_ are publicly available.
+
+```{admonition} Did you know?
+:class: info
+The [Python language](https://python.org) itself is an example of *FOSS*!
+```
+
+FOSS is nothing new. For example, the [Linux kernel source code](https://kernel.org) has been available to download for many years.
+
+```{admonition} Free $\neq$ open source!
+:class: tip
+Just because a software package may be free does not mean that its source code is open! For example, although Nvidia makes its video drivers available for free download, the source code for those drivers is proprietary.
+```
+
+Arguably, the greatest advantage of open-source software is that it enables _collaborative sharing_, and thus community feedback.
+
+Types of community input may include the following:
+
+1. _Issues_: usage questions, bug reports, feature requests
+1. _Pull requests_: a user can ask that that their changes/additions be incorporated into the project
+1. _Discussions_: a community forum on the open source project
+
+## Version control systems (VCS)
+
+We will discuss version control in more detail later in this series, but the need to track and manage changes to a project, especially one that involves software, has long been known. Over the years, FOSS developers have used VCS such as _cvs_, _svn_, and most recently, _git_. All of these systems are _command-line tools_.
+
+## FOSS and VCS on the Internet
+
+A successful FOSS project needs to be accessible via the web. As mentioned before, the Linux kernel and the Python language have long been available using first-generation remote access protocols such as FTP and HTTP, and SSH. Later, VCS tools such as cvs and svn established their own TCP protocols for remote access. With the advent of _git_, web-based services that supported HTTP(S) and SSH sprung up. Each of these VCS leverages the concept of a particular FOSS project as a <i>code repository</i>.
+
+```{admonition} Did you know?
[…]
