### 2026-03-02 d731296
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/PyCQA/isort: 8.0.0 → 8.0.1](https://github.com/PyCQA/isort/compare/8.0.0...8.0.1)
