+++ _preview/616/build/_shared/chunk-RBHDENDK.js
@@ -1,0 +1,1 @@
+import{c as r}from"/_preview/616/build/_shared/chunk-OZE3FFNP.js";var a=r((n,t)=>{t.exports=e;e.displayName="goModule";e.aliases=[];function e(o){o.languages["go-mod"]=o.languages["go-module"]={comment:{pattern:/\/\/.*/,greedy:!0},version:{pattern:/(^|[\s()[\],])v\d+\.\d+\.\d+(?:[+-][-+.\w]*)?(?![^\s()[\],])/,lookbehind:!0,alias:"number"},"go-version":{pattern:/((?:^|\s)go\s+)\d+(?:\.\d+){1,2}/,lookbehind:!0,alias:"number"},keyword:{pattern:/^([ \t]*)(?:exclude|go|module|replace|require|retract)\b/m,lookbehind:!0},operator:/=>/,punctuation:/[()[\],]/}}});export{a};
