### 2026-03-04 3d3415f
add additional resource
