### 2026-03-04 aa81572
minor changes to capitalization, etc. for consistency
