+++ foundations/github/github-cloning-forking.md
@@ -7,27 +7,18 @@
 
 ## Overview:
 
-1. Cloning and forking a git repository
+1. Cloning and forking a Git repository
 1. Cloning a repository
 1. Forking a repository
 
 ## Prerequisites
 
-<<<<<<< Updated upstream
-| Concepts                                   | Importance  | Notes                        |
-| ------------------------------------------ | ----------- | ---------------------------- |
-| [What is GitHub?](what-is-github.md)       | Necessary   | GitHub user account required |
-| [GitHub Repositories](github-repos.md)     | Necessary   |                              |
-| [Issues and Discussions](github-issues.md) | Recommended |                              |
-| Command-line shell                         | Helpful     |                              |
-=======
 | Concepts                                                      | Importance  | Notes                        |
 | ------------------------------------------------------------- | ----------- | ---------------------------- |
 | [What is GitHub?](what-is-github.md)                          | Necessary   | GitHub user account required |
 | [GitHub Repositories](github-repos.md)                        | Necessary   |                              |
 | [Issues and Discussions](github-issues.md)                    | Recommended |                              |
 | [The Unix Shell](https://swcarpentry.github.io/shell-novice/) | Helpful     | Or equivalent                |
->>>>>>> Stashed changes
 
 - **Time to learn**: 30 minutes
 
@@ -51,12 +42,12 @@ destination of the clone operation is your local computer, you will
 own the cloned contents. In either case, whether you clone or fork,
 any changes you make to the newly created repository will not impact
 the original without taking explicit action (e.g. performing a
-_push_ or submitting a _Pull Request_, the topics of later sections
+_push_ or submitting a _pull request_, the topics of later sections
 in this guide).
 
 Cloning and forking are often used together (more on this later).
-The illustration below demonstrates the operation of a Fork of a
-remote repository (UPSTREAM), followed by a clone of the newly
+The illustration below demonstrates the process of forking a
+remote repository (UPSTREAM), followed by cloning the newly
 created ORIGIN.
 
 ![clone-and-fork](../../images/github-clone-fork.png)
@@ -67,7 +58,7 @@ Cloning is ideal for the following scenarios:
 
 1. You wish to download, build, and install the latest version of a software package.
 1. You would like to experiment with a repository on your local computer, but do not desire to maintain a separate copy of it (termed a _fork_, to be covered later in this lesson) on your GitHub account.
-1. You have previously _forked_ a repository to your own GitHub account, and now wish to make changes to it for possible incorporation into the original repo, via a _Pull Request_.
+1. You have previously _forked_ a repository to your own GitHub account, and now wish to make changes to it for possible incorporation into the original repo, via a _pull request_.
 
 Let's consider the 2nd scenario. Say you wish to copy a GitHub repository to a computer you have access to (which could be your own computer, or one you have access to at work or school).
 
@@ -120,7 +111,7 @@ Resolving deltas: 100% (3/3), done.
 ```
 
 ```{note}
-Don't have git on your system already? Check out [these installation guides from the Carpentries](https://carpentries.github.io/workshop-template/install_instructions/#git).
+Don't have Git on your system already? Check out [these installation guides from the Carpentries](https://carpentries.github.io/workshop-template/install_instructions/#git).
 ```
 
 Now, you can `cd` into the `github-sandbox` directory which has been created and populated with the exact contents of the **origin**'s repository at the time you cloned it. If you have a Python installation, you could then type
@@ -135,7 +126,7 @@ to run the sample Python script. You should see the following output:
 Hello, Python learners!
 ```
 
[…]
