### 2026-03-04 9ce332d
add terms to glossary
