+++ foundations/github/github-pull-request.md
@@ -138,7 +138,7 @@ To someone with write permissions on the repository, the ability to merge will l
 
 However, this PR will NOT be merged, as the GitHub-Sandbox repository is intended to be static.
 
-## GitHub Workflows
+## Git Workflows
 
 The above demonstration is an example of the Git **Forking Workflow**, because we forked the [GitHub Sandbox repository](https://github.com/ProjectPythia/github-sandbox) before making our feature branches. This is most common when you do NOT have write-access to the upstream repository.
 
@@ -148,11 +148,7 @@ The steps leading up to creating your PR depend on your workflow. The main diffe
 the contributor now, for the Feature Workflow, navigates to the upstream, remote
 repository, not a personal remote fork, and initiates the PR there.
 
-<<<<<<< Updated upstream
-We will cover [GitHub Workflows](github-workflows.md) in greater detail in the next chapter.
-=======
 We will cover [Git workflows](git-workflows.md) in greater detail in the next chapter.
->>>>>>> Stashed changes
 
 ---
 
@@ -162,7 +158,7 @@ We will cover [Git workflows](git-workflows.md) in greater detail in the next ch
   that you have made to one repository are integrated (merged) into
   another repository.
 - The steps that lead up to
-  the PR depend your GitHub Workflow.
+  the PR depend your Git workflow.
 
 ### What's Next?
 
