### 2026-03-04 499dc2e
change name of resources section for consistency
