+++ environment.yml
@@ -13,6 +13,7 @@ dependencies:
   - pyproj
   - pythia-datasets
   - python
+  - pytz
   - scipy
   - ffmpeg
   - xarray
