### 2026-03-04 c707079
jupyterbook link fails jupyterbooks link checking :P
