### 2026-03-05 #308 @brian-rose
Hi @acocac, the work is still ongoing but wrapping up soon! We've had a lot of challenges to address with the evolving computational and publishing infrastructure, alongside the reviewer suggestions. This has all been necessary and worthwhile but time-consuming.

I expect that we will be finished addressing all the reviewer comments and ready for eyes on the new-and-improved Pythia Foundations in about one more week.
