### 2026-03-06 0459694
Update appendix/glossary.md
Co-authored-by: Brian Rose <brose@albany.edu>
