### 2026-03-06 6776c68
Update appendix/glossary.md
Co-authored-by: Brian Rose <brose@albany.edu>
